Major moves
moved all hard coded data to rawdata static class.  put all penrose stuff (like tile data) in the penrose class.  few other minor tweaks and fixes
