+++ penrose_simulator/Assets/core/Penrose.cs
@@ -1,28 +1,31 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
+using Random = UnityEngine.Random;
 
 [RequireComponent(typeof(MeshFilter))]
 [RequireComponent(typeof(MeshRenderer))]
 public class Penrose : MonoBehaviour {
 
-  private const int Total = 900;
-
+  public const int Total = 900;
+  public const float FullScale = 1.0f / 140.0f;
 
   public Color bgColor = Color.gray;
 
   [Header("Display Size")]
   public float scale = 0.003f;
+
   public float gapScale = 0.9f;
 
   [HideInInspector]
   public Color[] buffer = new Color[Total]; // input buffer
 
   [HideInInspector]
-  public Tiles geometry;
+  public TileData[] tiles;
 
   private readonly Vector3[] vertices = new Vector3[Total * 2 * 3];
   private readonly int[] triangles = new int[Total * 2 * 3];
   private readonly Color[] colors = new Color[Total * 2 * 3];
-  
+
   private Mesh mesh;
   private MeshFilter meshFilter;
   private MeshRenderer meshRenderer;
@@ -33,10 +36,7 @@ public class Penrose : MonoBehaviour {
   private void Awake() {
     meshFilter   = GetComponent<MeshFilter>();
     meshRenderer = GetComponent<MeshRenderer>();
-    material =
-      new Material(Shader.Find("Unlit/Penrose")) {hideFlags = HideFlags.HideAndDontSave, name = "PenMaterial"};
-
-    geometry = new Tiles();
+    material     = new Material(Shader.Find("Unlit/Penrose")) {hideFlags = HideFlags.HideAndDontSave, name = "PenMaterial"};
   }
 
   private void GenerateMesh() {
@@ -44,10 +44,10 @@ public class Penrose : MonoBehaviour {
     var j = 0;
 
     // grab the geometry
-    for(int n = 0; n < Geometry.Length; n += 6) {
-      var a = new Vector3(Geometry[j++] * scale, Geometry[j++] * scale, 0f);
-      var b = new Vector3(Geometry[j++] * scale, Geometry[j++] * scale, 0f);
-      var c = new Vector3(Geometry[j++] * scale, Geometry[j++] * scale, 0f);
+    for(int n = 0; n < RawData.Mesh.Length; n += 6) {
+      var a = new Vector3(RawData.Mesh[j++] * scale, RawData.Mesh[j++] * scale, 0f);
+      var b = new Vector3(RawData.Mesh[j++] * scale, RawData.Mesh[j++] * scale, 0f);
+      var c = new Vector3(RawData.Mesh[j++] * scale, RawData.Mesh[j++] * scale, 0f);
 
       var ab = b - a;
       var ac = c - a;
@@ -78,31 +78,40 @@ public class Penrose : MonoBehaviour {
       i += 3;
     }
 
-    mesh = new Mesh { vertices = vertices, triangles = triangles, colors = colors, name = "PenMesh", hideFlags = HideFlags.HideAndDontSave };
+    mesh = new Mesh {vertices = vertices, triangles = triangles, colors = colors, name = "PenMesh", hideFlags = HideFlags.HideAndDontSave};
 
     meshFilter.mesh       = mesh;
     meshRenderer.material = material;
   }
 
+  private void GenerateTiles() {
+    var j = 0;
+    tiles = new TileData[Total];
+    for(var i = 0; i < Total; i++) {
+      var t = new TileData {
+        neighbors = new int[4], type = RawData.Tiles[j++], center = {x = RawData.Tiles[j++] * FullScale, y = RawData.Tiles[j++] * FullScale}
+      };
+
+      for(var a = 0; a < 4; a++) t.neighbors[a] = RawData.Tiles[j++];
+
+      tiles[i] = t;
+    }
+  }
+
   void Start() {
     GenerateMesh();
+    GenerateTiles();
     bgBrightness = bgColor.grayscale;
-  }
-  
-  void Update() {
-    UpdateVertexColors();
-    mesh.RecalculateNormals();
   }
 
   void UpdateVertexColors() {
     // color all the mesh vertices
     var x = 0;
-    
+
     for(var i = 0; i < buffer.Length; i++) {
-      
       var color = buffer[i];
       color = FadeColorToBgColor(color);
-      
+
       // set the vertex color
       for(int j = 0; j < 6; j++) colors[x++] = color;
     }
@@ -110,8 +119,12 @@ public class Penrose : MonoBehaviour {
     mesh.colors = colors;
   }
 
-  private Color FadeColorToBgColor(Color color) {
+  void LateUpdate() {
+    UpdateVertexColors();
+    mesh.RecalculateNormals();
+  }
 
+  private Color FadeColorToBgColor(Color color) {
     // Lerp between color and background
     color = Color.Lerp(bgColor, color, color.grayscale);
 
@@ -128,1808 +141,22 @@ public class Penrose : MonoBehaviour {
     return color;
   }
 
-  private static readonly int[] Geometry = {
-82,932,0,1046,-82,932,
-82,932,0,819,-82,932,
-0,1046,0,1186,133,1142,
-0,1046,133,1002,133,1142,
-82,932,0,1046,133,1002,
-82,932,215,889,133,1002,
-0,733,133,776,133,636,
-0,733,0,593,133,636,
-133,776,215,889,82,932,
-133,776,0,819,82,932,
-0,733,133,776,0,819,
-0,733,-133,776,0,819,
-266,1186,133,1142,215,1029,
-266,1186,348,1072,215,1029,
-133,1142,0,1186,82,1299,
-133,1142,215,1256,82,1299,
-266,1186,133,1142,215,1256,
-266,1186,348,1299,215,1256,
-215,1029,133,1142,133,1002,
-215,1029,215,889,133,1002,
-215,1029,215,889,348,932,
-215,1029,348,1072,348,932,
-133,323,266,366,266,226,
-133,323,133,183,266,226,
-266,366,348,479,215,523,
-266,366,133,409,215,523,
-133,323,266,366,133,409,
-133,323,0,366,133,409,
-82,113,0,226,-82,113,
-82,113,0,0,-82,113,
-0,226,0,366,133,323,
-0,226,133,183,133,323,
-82,113,0,226,133,183,
-82,113,215,69,133,183,
-348,619,430,733,297,776,
-348,619,215,663,297,776,
-430,733,564,776,564,636,
-430,733,430,593,564,636,
-348,619,430,733,430,593,
-348,619,348,479,430,593,
-133,636,215,523,82,479,
-133,636,0,593,82,479,
-215,523,348,479,348,619,
-215,523,215,663,348,619,
-133,636,215,523,215,663,
-133,636,133,776,215,663,
-82,479,215,523,133,409,
-82,479,0,366,133,409,
-82,479,0,366,-82,479,
-82,479,0,593,-82,479,
-430,733,564,776,481,889,
-430,733,348,846,481,889,
-297,776,430,733,348,846,
-297,776,215,889,348,846,
-133,776,215,889,297,776,
-133,776,215,663,297,776,
-348,932,481,889,481,1029,
-348,932,348,1072,481,1029,
-348,932,481,889,348,846,
-348,932,215,889,348,846,
-779,1299,697,1186,830,1142,
-779,1299,912,1256,830,1142,
-697,1186,564,1142,564,1282,
-697,1186,697,1326,564,1282,
-266,1186,348,1299,430,1186,
-266,1186,348,1072,430,1186,
-430,1186,348,1299,481,1256,
-430,1186,564,1142,481,1256,
-430,1186,564,1142,481,1029,
-430,1186,348,1072,481,1029,
-646,1029,646,889,779,932,
-646,1029,779,1072,779,932,
-646,889,564,776,481,889,
-646,889,564,1002,481,889,
-646,1029,646,889,564,1002,
-646,1029,564,1142,564,1002,
-697,1186,564,1142,646,1029,
-697,1186,779,1072,646,1029,
-830,1142,697,1186,779,1072,
-830,1142,912,1029,779,1072,
-481,1029,481,889,564,1002,
-481,1029,564,1142,564,1002,
-861,366,994,323,912,209,
-861,366,779,253,912,209,
-994,323,1128,366,1045,479,
-994,323,912,436,1045,479,
-861,366,994,323,912,436,
-861,366,779,479,912,436,
-697,226,697,366,564,323,
-697,226,564,183,564,323,
-697,366,779,479,861,366,
-697,366,779,253,861,366,
-697,226,697,366,779,253,
-697,226,779,113,779,253,
-1210,479,1343,523,1261,636,
-1210,479,1128,593,1261,636,
-1343,523,1476,479,1394,366,
-1343,523,1261,409,1394,366,
-1210,479,1343,523,1261,409,
-1210,479,1128,366,1261,409,
-1045,619,1045,479,912,523,
-1045,619,912,663,912,523,
-1045,479,1128,366,1210,479,
-1045,479,1128,593,1210,479,
-1045,619,1045,479,1128,593,
-1045,619,1128,733,1128,593,
-912,523,1045,479,912,436,
-912,523,779,479,912,436,
-912,523,779,479,779,619,
-912,523,912,663,779,619,
-266,366,348,479,430,366,
-266,366,348,253,430,366,
-266,226,266,366,348,253,
-266,226,348,113,348,253,
-82,113,215,69,133,-43,
-82,113,0,0,133,-43,
-215,69,348,113,266,226,
-215,69,133,183,266,226,
-481,523,564,636,646,523,
-481,523,564,409,646,523,
-481,523,564,636,430,593,
-481,523,348,479,430,593,
-564,323,430,366,430,226,
-564,323,564,183,430,226,
-430,366,348,479,481,523,
-430,366,564,409,481,523,
-564,323,430,366,564,409,
-564,323,697,366,564,409,
-430,226,430,366,348,253,
-430,226,348,113,348,253,
-430,226,348,113,481,69,
-430,226,564,183,481,69,
-564,636,564,776,697,733,
-564,636,697,593,697,733,
-646,523,564,636,697,593,
-646,523,779,479,697,593,
-697,366,779,479,646,523,
-697,366,564,409,646,523,
-779,619,697,733,830,776,
-779,619,912,663,830,776,
-779,619,697,733,697,593,
-779,619,779,479,697,593,
-1774,889,1907,932,1825,1046,
-1774,889,1692,1002,1825,1046,
-1907,932,2040,889,1958,776,
-1907,932,1825,819,1958,776,
-1774,889,1907,932,1825,819,
-1774,889,1692,776,1825,819,
-1609,1029,1609,889,1476,932,
-1609,1029,1476,1072,1476,932,
-1609,889,1692,776,1774,889,
-1609,889,1692,1002,1774,889,
-1609,1029,1609,889,1692,1002,
-1609,1029,1692,1142,1692,1002,
-2122,776,2256,733,2173,619,
-2122,776,2040,663,2173,619,
-2256,733,2389,776,2306,889,
-2256,733,2173,846,2306,889,
-2122,776,2256,733,2173,846,
-2122,776,2040,889,2173,846,
-1958,636,1958,776,1825,733,
-1958,636,1825,593,1825,733,
-1958,776,2040,889,2122,776,
-1958,776,2040,663,2122,776,
-1958,636,1958,776,2040,663,
-1958,636,2040,523,2040,663,
-1825,733,1958,776,1825,819,
-1825,733,1692,776,1825,819,
-1825,733,1692,776,1692,636,
-1825,733,1825,593,1692,636,
-1178,1029,1178,889,1045,932,
-1178,1029,1045,1072,1045,932,
-1178,889,1261,776,1343,889,
-1178,889,1261,1002,1343,889,
-1178,1029,1178,889,1261,1002,
-1178,1029,1261,1142,1261,1002,
-994,1142,1128,1186,1045,1299,
-994,1142,912,1256,1045,1299,
-1128,1186,1261,1142,1178,1029,
-1128,1186,1045,1072,1178,1029,
-994,1142,1128,1186,1045,1072,
-994,1142,912,1029,1045,1072,
-1394,733,1476,619,1558,733,
-1394,733,1476,846,1558,733,
-1476,619,1476,479,1343,523,
-1476,619,1343,663,1343,523,
-1394,733,1476,619,1343,663,
-1394,733,1261,776,1343,663,
-1476,932,1343,889,1343,1029,
-1476,932,1476,1072,1343,1029,
-1343,889,1261,776,1394,733,
-1343,889,1476,846,1394,733,
-1476,932,1343,889,1476,846,
-1476,932,1609,889,1476,846,
-1343,1029,1343,889,1261,1002,
-1343,1029,1261,1142,1261,1002,
-1343,1029,1261,1142,1394,1186,
-1343,1029,1476,1072,1394,1186,
-1476,619,1476,479,1609,523,
-1476,619,1609,663,1609,523,
-1558,733,1476,619,1609,663,
-1558,733,1692,776,1609,663,
-1609,889,1692,776,1558,733,
-1609,889,1476,846,1558,733,
-1692,636,1609,523,1742,479,
-1692,636,1825,593,1742,479,
-1692,636,1609,523,1609,663,
-1692,636,1692,776,1609,663,
-1178,889,1261,776,1128,733,
-1178,889,1045,846,1128,733,
-1045,932,1178,889,1045,846,
-1045,932,912,889,1045,846,
-994,1142,912,1029,830,1142,
-994,1142,912,1256,830,1142,
-912,1029,912,889,1045,932,
-912,1029,1045,1072,1045,932,
-1261,636,1343,523,1343,663,
-1261,636,1261,776,1343,663,
-1045,619,1128,733,994,776,
-1045,619,912,663,994,776,
-1128,733,1261,776,1261,636,
-1128,733,1128,593,1261,636,
-994,776,1128,733,1045,846,
-994,776,912,889,1045,846,
-994,776,912,889,830,776,
-994,776,912,663,830,776,
-646,889,564,776,697,733,
-646,889,779,846,697,733,
-779,932,646,889,779,846,
-779,932,912,889,779,846,
-912,1029,912,889,779,932,
-912,1029,779,1072,779,932,
-830,776,697,733,779,846,
-830,776,912,889,779,846,
-1907,932,2040,889,2040,1029,
-1907,932,1907,1072,2040,1029,
-1825,1046,1907,932,1907,1072,
-1825,1046,1825,1186,1907,1072,
-1609,1029,1692,1142,1558,1186,
-1609,1029,1476,1072,1558,1186,
-1692,1142,1825,1186,1825,1046,
-1692,1142,1692,1002,1825,1046,
-2173,932,2306,889,2306,1029,
-2173,932,2173,1072,2306,1029,
-2173,932,2306,889,2173,846,
-2173,932,2040,889,2173,846,
-2122,1142,2040,1029,1958,1142,
-2122,1142,2040,1256,1958,1142,
-2040,1029,2040,889,2173,932,
-2040,1029,2173,1072,2173,932,
-2122,1142,2040,1029,2173,1072,
-2122,1142,2256,1186,2173,1072,
-1958,1142,2040,1029,1907,1072,
-1958,1142,1825,1186,1907,1072,
-1958,1142,1825,1186,1907,1299,
-1958,1142,2040,1256,1907,1299,
-1128,1186,1261,1142,1261,1282,
-1128,1186,1128,1326,1261,1282,
-1558,1186,1476,1299,1394,1186,
-1558,1186,1476,1072,1394,1186,
-1558,1186,1476,1299,1609,1256,
-1558,1186,1692,1142,1609,1256,
-1394,1186,1476,1299,1343,1256,
-1394,1186,1261,1142,1343,1256,
-1692,1142,1825,1186,1742,1299,
-1692,1142,1609,1256,1742,1299,
-2389,1282,2389,1142,2256,1186,
-2389,1282,2256,1326,2256,1186,
-2306,889,2389,776,2471,889,
-2306,889,2389,1002,2471,889,
-2306,1029,2306,889,2389,1002,
-2306,1029,2389,1142,2389,1002,
-2122,1142,2256,1186,2173,1299,
-2122,1142,2040,1256,2173,1299,
-2256,1186,2389,1142,2306,1029,
-2256,1186,2173,1072,2306,1029,
-2604,932,2471,889,2471,1029,
-2604,932,2604,1072,2471,1029,
-2471,889,2389,776,2522,733,
-2471,889,2604,846,2522,733,
-2604,932,2471,889,2604,846,
-2604,932,2737,889,2604,846,
-2737,1029,2737,889,2604,932,
-2737,1029,2604,1072,2604,932,
-2655,1142,2737,1029,2604,1072,
-2655,1142,2522,1186,2604,1072,
-2389,1282,2389,1142,2522,1186,
-2389,1282,2522,1326,2522,1186,
-2471,1029,2471,889,2389,1002,
-2471,1029,2389,1142,2389,1002,
-2655,1142,2522,1186,2604,1299,
-2655,1142,2737,1256,2604,1299,
-2522,1186,2389,1142,2471,1029,
-2522,1186,2604,1072,2471,1029,
-2389,-689,2471,-803,2338,-846,
-2389,-689,2256,-733,2338,-846,
-2471,-803,2604,-846,2604,-706,
-2471,-803,2471,-663,2604,-706,
-2389,-689,2471,-803,2471,-663,
-2389,-689,2389,-549,2471,-663,
-2173,-706,2256,-593,2122,-549,
-2173,-706,2040,-663,2122,-549,
-2256,-593,2389,-549,2389,-689,
-2256,-593,2256,-733,2389,-689,
-2173,-706,2256,-593,2256,-733,
-2173,-706,2173,-846,2256,-733,
-2686,-593,2604,-706,2522,-593,
-2686,-593,2604,-479,2522,-593,
-2604,-706,2604,-846,2737,-803,
-2604,-706,2737,-663,2737,-803,
-2686,-593,2604,-706,2737,-663,
-2686,-593,2820,-549,2737,-663,
-2522,-593,2604,-706,2471,-663,
-2522,-593,2389,-549,2471,-663,
-2522,-593,2389,-549,2471,-436,
-2522,-593,2604,-479,2471,-436,
-1825,-453,1907,-339,1774,-296,
-1825,-453,1692,-409,1774,-296,
-1907,-339,2040,-296,2040,-436,
-1907,-339,1907,-479,2040,-436,
-1825,-453,1907,-339,1907,-479,
-1825,-453,1825,-593,1907,-479,
-1609,-436,1692,-549,1558,-593,
-1609,-436,1476,-479,1558,-593,
-1692,-549,1825,-593,1825,-453,
-1692,-549,1692,-409,1825,-453,
-1609,-436,1692,-549,1692,-409,
-1609,-436,1609,-296,1692,-409,
-2173,-339,2306,-296,2306,-436,
-2173,-339,2173,-479,2306,-436,
-2306,-296,2389,-183,2256,-139,
-2306,-296,2173,-253,2256,-139,
-2173,-339,2306,-296,2173,-253,
-2173,-339,2040,-296,2173,-253,
-2122,-549,2040,-436,1958,-549,
-2122,-549,2040,-663,1958,-549,
-2040,-436,2040,-296,2173,-339,
-2040,-436,2173,-479,2173,-339,
-2122,-549,2040,-436,2173,-479,
-2122,-549,2256,-593,2173,-479,
-1958,-549,2040,-436,1907,-479,
-1958,-549,1825,-593,1907,-479,
-1958,-549,1825,-593,1907,-706,
-1958,-549,2040,-663,1907,-706,
-2306,-296,2389,-183,2471,-296,
-2306,-296,2389,-409,2471,-296,
-2306,-436,2306,-296,2389,-409,
-2306,-436,2389,-549,2389,-409,
-2256,-593,2389,-549,2306,-436,
-2256,-593,2173,-479,2306,-436,
-2471,-436,2471,-296,2604,-339,
-2471,-436,2604,-479,2604,-339,
-2471,-436,2471,-296,2389,-409,
-2471,-436,2389,-549,2389,-409,
-2686,-593,2820,-549,2737,-436,
-2686,-593,2604,-479,2737,-436,
-2737,-436,2737,-296,2604,-339,
-2737,-436,2604,-479,2604,-339,
-2655,-183,2522,-139,2604,-26,
-2655,-183,2737,-69,2604,-26,
-2522,-139,2389,-183,2471,-296,
-2522,-139,2604,-253,2471,-296,
-2655,-183,2522,-139,2604,-253,
-2655,-183,2737,-296,2604,-253,
-2820,-183,2737,-296,2655,-183,
-2820,-183,2737,-69,2655,-183,
-2604,-339,2471,-296,2604,-253,
-2604,-339,2737,-296,2604,-253,
-994,323,1128,366,1128,226,
-994,323,994,183,1128,226,
-912,209,994,323,994,183,
-912,209,912,69,994,183,
-697,226,779,113,646,69,
-697,226,564,183,646,69,
-779,113,912,69,912,209,
-779,113,779,253,912,209,
-1261,323,1394,366,1394,226,
-1261,323,1261,183,1394,226,
-1261,323,1394,366,1261,409,
-1261,323,1128,366,1261,409,
-1210,113,1128,226,1045,113,
-1210,113,1128,1,1045,113,
-1128,226,1128,366,1261,323,
-1128,226,1261,183,1261,323,
-1210,113,1128,226,1261,183,
-1210,113,1343,69,1261,183,
-1045,113,1128,226,994,183,
-1045,113,912,69,994,183,
-1045,113,912,69,994,-43,
-1045,113,1128,1,994,-43,
-348,-26,430,-139,297,-183,
-348,-26,215,-69,297,-183,
-430,-139,564,-183,564,-43,
-430,-139,430,0,564,-43,
-348,-26,430,-139,430,0,
-348,-26,348,113,430,0,
-215,69,348,113,348,-26,
-215,69,215,-69,348,-26,
-133,-43,215,69,215,-69,
-133,-43,133,-183,215,-69,
-697,-139,830,-183,830,-43,
-697,-139,697,0,830,-43,
-830,-183,912,-296,779,-339,
-830,-183,697,-226,779,-339,
-697,-139,830,-183,697,-226,
-697,-139,564,-183,697,-226,
-646,69,564,-43,481,69,
-646,69,564,183,481,69,
-564,-43,564,-183,697,-139,
-564,-43,697,0,697,-139,
-646,69,564,-43,697,0,
-646,69,779,113,697,0,
-481,69,564,-43,430,0,
-481,69,348,113,430,0,
-830,-183,912,-296,994,-183,
-830,-183,912,-69,994,-183,
-830,-43,830,-183,912,-69,
-830,-43,912,69,912,-69,
-779,113,912,69,830,-43,
-779,113,697,0,830,-43,
-994,-43,994,-183,1128,-139,
-994,-43,1128,1,1128,-139,
-994,-43,994,-183,912,-69,
-994,-43,912,69,912,-69,
-1958,323,2091,366,2091,226,
-1958,323,1958,183,2091,226,
-2091,366,2173,479,2040,523,
-2091,366,1958,409,2040,523,
-1958,323,2091,366,1958,409,
-1958,323,1825,366,1958,409,
-1907,113,1825,226,1742,113,
-1907,113,1825,1,1742,113,
-1825,226,1825,366,1958,323,
-1825,226,1958,183,1958,323,
-1907,113,1825,226,1958,183,
-1907,113,2040,69,1958,183,
-2256,733,2389,776,2389,636,
-2256,733,2256,593,2389,636,
-2173,619,2256,733,2256,593,
-2173,619,2173,479,2256,593,
-1958,636,2040,523,1907,479,
-1958,636,1825,593,1907,479,
-2040,523,2173,479,2173,619,
-2040,523,2040,663,2173,619,
-1907,479,2040,523,1958,409,
-1907,479,1825,366,1958,409,
-1907,479,1825,366,1742,479,
-1907,479,1825,593,1742,479,
-1558,-139,1476,-26,1394,-139,
-1558,-139,1476,-253,1394,-139,
-1476,-26,1476,113,1609,69,
-1476,-26,1609,-69,1609,69,
-1558,-139,1476,-26,1609,-69,
-1558,-139,1692,-183,1609,-69,
-1476,-339,1609,-296,1609,-436,
-1476,-339,1476,-479,1609,-436,
-1609,-296,1692,-183,1558,-139,
-1609,-296,1476,-253,1558,-139,
-1476,-339,1609,-296,1476,-253,
-1476,-339,1343,-296,1476,-253,
-1558,226,1558,366,1692,323,
-1558,226,1692,183,1692,323,
-1558,366,1476,479,1394,366,
-1558,366,1476,253,1394,366,
-1558,226,1558,366,1476,253,
-1558,226,1476,113,1476,253,
-1742,113,1609,69,1692,-43,
-1742,113,1825,1,1692,-43,
-1609,69,1476,113,1558,226,
-1609,69,1692,183,1558,226,
-1742,113,1609,69,1692,183,
-1742,113,1825,226,1692,183,
-1692,-43,1609,69,1609,-69,
-1692,-43,1692,-183,1609,-69,
-1692,-43,1692,-183,1825,-139,
-1692,-43,1825,1,1825,-139,
-1558,366,1476,479,1609,523,
-1558,366,1692,409,1609,523,
-1692,323,1558,366,1692,409,
-1692,323,1825,366,1692,409,
-1825,226,1825,366,1692,323,
-1825,226,1692,183,1692,323,
-1742,479,1609,523,1692,409,
-1742,479,1825,366,1692,409,
-1476,-26,1476,113,1343,69,
-1476,-26,1343,-69,1343,69,
-1394,-139,1476,-26,1343,-69,
-1394,-139,1261,-183,1343,-69,
-1476,-339,1343,-296,1343,-436,
-1476,-339,1476,-479,1343,-436,
-1343,-296,1261,-183,1394,-139,
-1343,-296,1476,-253,1394,-139,
-1394,226,1394,366,1476,253,
-1394,226,1476,113,1476,253,
-1210,113,1343,69,1261,-43,
-1210,113,1128,1,1261,-43,
-1343,69,1476,113,1394,226,
-1343,69,1261,183,1394,226,
-1261,-43,1343,69,1343,-69,
-1261,-43,1261,-183,1343,-69,
-1261,-43,1261,-183,1128,-139,
-1261,-43,1128,1,1128,-139,
-1178,-296,1045,-339,1128,-453,
-1178,-296,1261,-409,1128,-453,
-1045,-339,912,-296,994,-183,
-1045,-339,1128,-226,994,-183,
-1178,-296,1045,-339,1128,-226,
-1178,-296,1261,-183,1128,-226,
-1343,-296,1261,-183,1178,-296,
-1343,-296,1261,-409,1178,-296,
-1343,-436,1343,-296,1261,-409,
-1343,-436,1261,-549,1261,-409,
-1128,-139,994,-183,1128,-226,
-1128,-139,1261,-183,1128,-226,
-2091,366,2173,479,2256,366,
-2091,366,2173,253,2256,366,
-2091,226,2091,366,2173,253,
-2091,226,2173,113,2173,253,
-1907,113,2040,69,1958,-43,
-1907,113,1825,1,1958,-43,
-2040,69,2173,113,2091,226,
-2040,69,1958,183,2091,226,
-2306,523,2389,636,2471,523,
-2306,523,2389,409,2471,523,
-2306,523,2389,636,2256,593,
-2306,523,2173,479,2256,593,
-2389,323,2256,366,2256,226,
-2389,323,2389,183,2256,226,
-2256,366,2173,479,2306,523,
-2256,366,2389,409,2306,523,
-2389,323,2256,366,2389,409,
-2389,323,2522,366,2389,409,
-2256,226,2256,366,2173,253,
-2256,226,2173,113,2173,253,
-2256,226,2173,113,2306,69,
-2256,226,2389,183,2306,69,
-1907,-339,2040,-296,1958,-183,
-1907,-339,1825,-226,1958,-183,
-1774,-296,1907,-339,1825,-226,
-1774,-296,1692,-183,1825,-226,
-1609,-296,1692,-183,1774,-296,
-1609,-296,1692,-409,1774,-296,
-2122,-183,2256,-139,2173,-26,
-2122,-183,2040,-69,2173,-26,
-2122,-183,2256,-139,2173,-253,
-2122,-183,2040,-296,2173,-253,
-1958,-43,1958,-183,1825,-139,
-1958,-43,1825,1,1825,-139,
-1958,-183,2040,-296,2122,-183,
-1958,-183,2040,-69,2122,-183,
-1958,-43,1958,-183,2040,-69,
-1958,-43,2040,69,2040,-69,
-1825,-139,1958,-183,1825,-226,
-1825,-139,1692,-183,1825,-226,
-2256,-139,2389,-183,2389,-43,
-2256,-139,2256,1,2389,-43,
-2173,-26,2256,-139,2256,1,
-2173,-26,2173,113,2256,1,
-2040,69,2173,113,2173,-26,
-2040,69,2040,-69,2173,-26,
-2306,69,2389,-43,2471,69,
-2306,69,2389,183,2471,69,
-2306,69,2389,-43,2256,1,
-2306,69,2173,113,2256,1,
-2686,366,2604,479,2522,366,
-2686,366,2604,253,2522,366,
-2686,226,2686,366,2604,253,
-2686,226,2604,113,2604,253,
-2737,69,2604,113,2686,226,
-2737,69,2820,183,2686,226,
-2389,636,2389,776,2522,733,
-2389,636,2522,593,2522,733,
-2471,523,2389,636,2522,593,
-2471,523,2604,479,2522,593,
-2389,323,2522,366,2522,226,
-2389,323,2389,183,2522,226,
-2522,366,2604,479,2471,523,
-2522,366,2389,409,2471,523,
-2522,226,2522,366,2604,253,
-2522,226,2604,113,2604,253,
-2522,226,2604,113,2471,69,
-2522,226,2389,183,2471,69,
-2522,-139,2389,-183,2389,-43,
-2522,-139,2522,1,2389,-43,
-2604,-26,2522,-139,2522,1,
-2604,-26,2604,113,2522,1,
-2737,69,2604,113,2604,-26,
-2737,69,2737,-69,2604,-26,
-2471,69,2389,-43,2522,1,
-2471,69,2604,113,2522,1,
-2655,776,2522,733,2604,619,
-2655,776,2737,663,2604,619,
-2655,776,2522,733,2604,846,
-2655,776,2737,889,2604,846,
-2820,776,2737,889,2655,776,
-2820,776,2737,663,2655,776,
-2686,366,2604,479,2737,523,
-2686,366,2820,409,2737,523,
-2604,619,2522,733,2522,593,
-2604,619,2604,479,2522,593,
-2737,523,2604,479,2604,619,
-2737,523,2737,663,2604,619,
-2471,-803,2604,-846,2522,-959,
-2471,-803,2389,-916,2522,-959,
-2338,-846,2471,-803,2389,-916,
-2338,-846,2256,-959,2389,-916,
-2173,-706,2173,-846,2040,-803,
-2173,-706,2040,-663,2040,-803,
-2173,-846,2256,-959,2338,-846,
-2173,-846,2256,-733,2338,-846,
-2686,-959,2820,-1002,2737,-1116,
-2686,-959,2604,-1072,2737,-1116,
-2686,-959,2820,-1002,2737,-889,
-2686,-959,2604,-846,2737,-889,
-2522,-1099,2522,-959,2389,-1002,
-2522,-1099,2389,-1142,2389,-1002,
-2522,-959,2604,-846,2686,-959,
-2522,-959,2604,-1072,2686,-959,
-2522,-1099,2522,-959,2604,-1072,
-2522,-1099,2604,-1212,2604,-1072,
-2389,-1002,2522,-959,2389,-916,
-2389,-1002,2256,-959,2389,-916,
-2389,-1002,2256,-959,2256,-1099,
-2389,-1002,2389,-1142,2256,-1099,
-1742,-706,1742,-846,1609,-803,
-1742,-706,1609,-663,1609,-803,
-1742,-846,1825,-959,1907,-846,
-1742,-846,1825,-733,1907,-846,
-1742,-706,1742,-846,1825,-733,
-1742,-706,1825,-593,1825,-733,
-1692,-549,1825,-593,1742,-706,
-1692,-549,1609,-663,1742,-706,
-1558,-593,1692,-549,1609,-663,
-1558,-593,1476,-706,1609,-663,
-1958,-1002,2040,-1116,2122,-1002,
-1958,-1002,2040,-889,2122,-1002,
-2040,-1116,2040,-1256,1907,-1212,
-2040,-1116,1907,-1072,1907,-1212,
-1958,-1002,2040,-1116,1907,-1072,
-1958,-1002,1825,-959,1907,-1072,
-2040,-803,1907,-846,1907,-706,
-2040,-803,2040,-663,1907,-706,
-1907,-846,1825,-959,1958,-1002,
-1907,-846,2040,-889,1958,-1002,
-2040,-803,1907,-846,2040,-889,
-2040,-803,2173,-846,2040,-889,
-1907,-706,1907,-846,1825,-733,
-1907,-706,1825,-593,1825,-733,
-2040,-1116,2040,-1256,2173,-1212,
-2040,-1116,2173,-1072,2173,-1212,
-2122,-1002,2040,-1116,2173,-1072,
-2122,-1002,2256,-959,2173,-1072,
-2173,-846,2256,-959,2122,-1002,
-2173,-846,2040,-889,2122,-1002,
-2256,-1099,2173,-1212,2306,-1256,
-2256,-1099,2389,-1142,2306,-1256,
-2256,-1099,2173,-1212,2173,-1072,
-2256,-1099,2256,-959,2173,-1072,
-2655,-1369,2737,-1256,2604,-1212,
-2655,-1369,2522,-1326,2604,-1212,
-2522,-1099,2604,-1212,2471,-1256,
-2522,-1099,2389,-1142,2471,-1256,
-2604,-1212,2737,-1256,2737,-1116,
-2604,-1212,2604,-1072,2737,-1116,
-2471,-1256,2604,-1212,2522,-1326,
-2471,-1256,2389,-1369,2522,-1326,
-2471,-1256,2389,-1369,2306,-1256,
-2471,-1256,2389,-1142,2306,-1256,
-2122,-1369,2040,-1256,2173,-1212,
-2122,-1369,2256,-1326,2173,-1212,
-2306,-1256,2173,-1212,2256,-1326,
-2306,-1256,2389,-1369,2256,-1326,
-614,-706,614,-846,481,-803,
-614,-706,481,-663,481,-803,
-614,-846,697,-959,779,-846,
-614,-846,697,-733,779,-846,
-614,-706,614,-846,697,-733,
-614,-706,697,-593,697,-733,
-430,-593,564,-549,481,-436,
-430,-593,348,-479,481,-436,
-564,-549,697,-593,614,-706,
-564,-549,481,-663,614,-706,
-430,-593,564,-549,481,-663,
-430,-593,348,-706,481,-663,
-830,-1002,912,-1116,994,-1002,
-830,-1002,912,-889,994,-1002,
-912,-1116,912,-1256,779,-1212,
-912,-1116,779,-1072,779,-1212,
-830,-1002,912,-1116,779,-1072,
-830,-1002,697,-959,779,-1072,
-912,-803,779,-846,779,-706,
-912,-803,912,-663,779,-706,
-779,-846,697,-959,830,-1002,
-779,-846,912,-889,830,-1002,
-912,-803,779,-846,912,-889,
-912,-803,1045,-846,912,-889,
-779,-706,779,-846,697,-733,
-779,-706,697,-593,697,-733,
-779,-706,697,-593,830,-549,
-779,-706,912,-663,830,-549,
-430,-139,564,-183,481,-296,
-430,-139,348,-253,481,-296,
-297,-183,430,-139,348,-253,
-297,-183,215,-296,348,-253,
-133,-43,133,-183,1,-139,
-133,-43,0,0,1,-139,
-133,-183,215,-296,297,-183,
-133,-183,215,-69,297,-183,
-646,-296,779,-339,697,-453,
-646,-296,564,-409,697,-453,
-646,-296,779,-339,697,-226,
-646,-296,564,-183,697,-226,
-481,-436,481,-296,348,-339,
-481,-436,348,-479,348,-339,
-481,-296,564,-183,646,-296,
-481,-296,564,-409,646,-296,
-481,-436,481,-296,564,-409,
-481,-436,564,-549,564,-409,
-348,-339,481,-296,348,-253,
-348,-339,215,-296,348,-253,
-348,-339,215,-296,215,-436,
-348,-339,348,-479,215,-436,
-779,-339,912,-296,912,-436,
-779,-339,779,-479,912,-436,
-697,-453,779,-339,779,-479,
-697,-453,697,-593,779,-479,
-564,-549,697,-593,697,-453,
-564,-549,564,-409,697,-453,
-830,-549,912,-436,994,-549,
-830,-549,912,-663,994,-549,
-830,-549,912,-436,779,-479,
-830,-549,697,-593,779,-479,
-1476,-1212,1343,-1256,1343,-1116,
-1476,-1212,1476,-1072,1343,-1116,
-1343,-1256,1261,-1369,1394,-1412,
-1343,-1256,1476,-1299,1394,-1412,
-1476,-1212,1343,-1256,1476,-1299,
-1476,-1212,1609,-1256,1476,-1299,
-1343,-803,1210,-846,1210,-706,
-1343,-803,1343,-663,1210,-706,
-1210,-846,1128,-959,1261,-1002,
-1210,-846,1343,-889,1261,-1002,
-1343,-803,1210,-846,1343,-889,
-1343,-803,1476,-846,1343,-889,
-1394,-593,1476,-706,1558,-593,
-1394,-593,1476,-479,1558,-593,
-1476,-706,1476,-846,1343,-803,
-1476,-706,1343,-663,1343,-803,
-1394,-593,1476,-706,1343,-663,
-1394,-593,1261,-549,1343,-663,
-1128,-1099,1045,-1212,1178,-1256,
-1128,-1099,1261,-1142,1178,-1256,
-1045,-1212,912,-1256,912,-1116,
-1045,-1212,1045,-1072,912,-1116,
-1128,-1099,1045,-1212,1045,-1072,
-1128,-1099,1128,-959,1045,-1072,
-1343,-1116,1261,-1002,1394,-959,
-1343,-1116,1476,-1072,1394,-959,
-1261,-1002,1128,-959,1128,-1099,
-1261,-1002,1261,-1142,1128,-1099,
-1343,-1116,1261,-1002,1261,-1142,
-1343,-1116,1343,-1256,1261,-1142,
-1394,-959,1261,-1002,1343,-889,
-1394,-959,1476,-846,1343,-889,
-1394,-959,1476,-846,1558,-959,
-1394,-959,1476,-1072,1558,-959,
-1045,-1212,912,-1256,994,-1369,
-1045,-1212,1128,-1326,994,-1369,
-1178,-1256,1045,-1212,1128,-1326,
-1178,-1256,1261,-1369,1128,-1326,
-1343,-1256,1261,-1369,1178,-1256,
-1343,-1256,1261,-1142,1178,-1256,
-1210,-846,1128,-959,1045,-846,
-1210,-846,1128,-733,1045,-846,
-1210,-706,1210,-846,1128,-733,
-1210,-706,1128,-593,1128,-733,
-1394,-593,1261,-549,1343,-436,
-1394,-593,1476,-479,1343,-436,
-1261,-549,1128,-593,1210,-706,
-1261,-549,1343,-663,1210,-706,
-994,-1002,912,-1116,1045,-1072,
-994,-1002,1128,-959,1045,-1072,
-912,-803,1045,-846,1045,-706,
-912,-803,912,-663,1045,-706,
-1045,-846,1128,-959,994,-1002,
-1045,-846,912,-889,994,-1002,
-1045,-706,1045,-846,1128,-733,
-1045,-706,1128,-593,1128,-733,
-1045,-706,1128,-593,994,-549,
-1045,-706,912,-663,994,-549,
-1045,-339,912,-296,912,-436,
-1045,-339,1045,-479,912,-436,
-1128,-453,1045,-339,1045,-479,
-1128,-453,1128,-593,1045,-479,
-1261,-549,1128,-593,1128,-453,
-1261,-549,1261,-409,1128,-453,
-994,-549,912,-436,1045,-479,
-994,-549,1128,-593,1045,-479,
-1476,-1212,1609,-1256,1609,-1116,
-1476,-1212,1476,-1072,1609,-1116,
-1609,-1256,1692,-1369,1558,-1412,
-1609,-1256,1476,-1299,1558,-1412,
-1742,-846,1825,-959,1692,-1002,
-1742,-846,1609,-889,1692,-1002,
-1609,-803,1742,-846,1609,-889,
-1609,-803,1476,-846,1609,-889,
-1476,-706,1476,-846,1609,-803,
-1476,-706,1609,-663,1609,-803,
-1825,-1099,1907,-1212,1774,-1256,
-1825,-1099,1692,-1142,1774,-1256,
-1825,-1099,1907,-1212,1907,-1072,
-1825,-1099,1825,-959,1907,-1072,
-1609,-1116,1692,-1002,1558,-959,
-1609,-1116,1476,-1072,1558,-959,
-1692,-1002,1825,-959,1825,-1099,
-1692,-1002,1692,-1142,1825,-1099,
-1609,-1116,1692,-1002,1692,-1142,
-1609,-1116,1609,-1256,1692,-1142,
-1558,-959,1692,-1002,1609,-889,
-1558,-959,1476,-846,1609,-889,
-1907,-1212,2040,-1256,1958,-1369,
-1907,-1212,1825,-1326,1958,-1369,
-1774,-1256,1907,-1212,1825,-1326,
-1774,-1256,1692,-1369,1825,-1326,
-1609,-1256,1692,-1369,1774,-1256,
-1609,-1256,1692,-1142,1774,-1256,
-614,-846,697,-959,564,-1002,
-614,-846,481,-889,564,-1002,
-481,-803,614,-846,481,-889,
-481,-803,348,-846,481,-889,
-430,-593,348,-706,266,-593,
-430,-593,348,-479,266,-593,
-348,-706,348,-846,481,-803,
-348,-706,481,-663,481,-803,
-697,-1099,779,-1212,646,-1256,
-697,-1099,564,-1142,646,-1256,
-697,-1099,779,-1212,779,-1072,
-697,-1099,697,-959,779,-1072,
-481,-1116,564,-1002,430,-959,
-481,-1116,348,-1072,430,-959,
-564,-1002,697,-959,697,-1099,
-564,-1002,564,-1142,697,-1099,
-481,-1116,564,-1002,564,-1142,
-481,-1116,481,-1256,564,-1142,
-430,-959,564,-1002,481,-889,
-430,-959,348,-846,481,-889,
-430,-959,348,-846,266,-959,
-430,-959,348,-1072,266,-959,
-82,-339,5,-453,-82,-339,
-82,-339,2,-226,-82,-339,
-5,-453,7,-593,133,-549,
-5,-453,133,-409,133,-549,
-82,-339,5,-453,133,-409,
-82,-339,215,-296,133,-409,
-133,-183,215,-296,82,-339,
-133,-183,2,-226,82,-339,
-1,-139,133,-183,2,-226,
-1,-139,-133,-183,2,-226,
-82,-706,82,-846,215,-803,
-82,-706,215,-663,215,-803,
-82,-846,1,-959,-82,-846,
-82,-846,8,-733,-82,-846,
-82,-706,82,-846,8,-733,
-82,-706,7,-593,8,-733,
-266,-593,133,-549,215,-436,
-266,-593,348,-479,215,-436,
-133,-549,7,-593,82,-706,
-133,-549,215,-663,82,-706,
-266,-593,133,-549,215,-663,
-266,-593,348,-706,215,-663,
-215,-436,133,-549,133,-409,
-215,-436,215,-296,133,-409,
-82,-846,1,-959,133,-1002,
-82,-846,215,-889,133,-1002,
-215,-803,82,-846,215,-889,
-215,-803,348,-846,215,-889,
-348,-706,348,-846,215,-803,
-348,-706,215,-663,215,-803,
-266,-959,133,-1002,215,-1116,
-266,-959,348,-1072,215,-1116,
-266,-959,133,-1002,215,-889,
-266,-959,348,-846,215,-889,
-830,-1369,912,-1256,779,-1212,
-830,-1369,697,-1326,779,-1212,
-430,-1412,564,-1369,481,-1256,
-430,-1412,348,-1299,481,-1256,
-297,-1369,430,-1412,348,-1299,
-297,-1369,215,-1256,348,-1299,
-646,-1256,779,-1212,697,-1326,
-646,-1256,564,-1369,697,-1326,
-481,-1116,481,-1256,348,-1212,
-481,-1116,348,-1072,348,-1212,
-481,-1256,564,-1369,646,-1256,
-481,-1256,564,-1142,646,-1256,
-348,-1212,481,-1256,348,-1299,
-348,-1212,215,-1256,348,-1299,
-348,-1212,215,-1256,215,-1116,
-348,-1212,348,-1072,215,-1116,
-82,-1212,1,-1099,-82,-1212,
-82,-1212,1,-1326,-82,-1212,
-1,-1099,1,-959,133,-1002,
-1,-1099,133,-1142,133,-1002,
-82,-1212,1,-1099,133,-1142,
-82,-1212,215,-1256,133,-1142,
-133,-1369,215,-1256,82,-1212,
-133,-1369,1,-1326,82,-1212,
-215,-1116,133,-1002,133,-1142,
-215,-1116,215,-1256,133,-1142,
--481,-803,-614,-846,-614,-706,
--481,-803,-481,-663,-614,-706,
--614,-846,-697,-959,-564,-1002,
--614,-846,-481,-889,-564,-1002,
--481,-803,-614,-846,-481,-889,
--481,-803,-348,-846,-481,-889,
--430,-593,-348,-706,-266,-593,
--430,-593,-348,-479,-266,-593,
--348,-706,-348,-846,-481,-803,
--348,-706,-481,-663,-481,-803,
--430,-593,-348,-706,-481,-663,
--430,-593,-564,-549,-481,-663,
--697,-1099,-779,-1212,-646,-1256,
--697,-1099,-564,-1142,-646,-1256,
--779,-1212,-912,-1256,-912,-1116,
--779,-1212,-779,-1072,-912,-1116,
--697,-1099,-779,-1212,-779,-1072,
--697,-1099,-697,-959,-779,-1072,
--481,-1116,-564,-1002,-430,-959,
--481,-1116,-348,-1072,-430,-959,
--564,-1002,-697,-959,-697,-1099,
--564,-1002,-564,-1142,-697,-1099,
--481,-1116,-564,-1002,-564,-1142,
--481,-1116,-481,-1256,-564,-1142,
--430,-959,-564,-1002,-481,-889,
--430,-959,-348,-846,-481,-889,
--430,-959,-348,-846,-266,-959,
--430,-959,-348,-1072,-266,-959,
-5,-453,7,-593,-133,-549,
-5,-453,-133,-409,-133,-549,
--82,-339,5,-453,-133,-409,
--82,-339,-215,-296,-133,-409,
-1,-139,-133,-183,-133,-43,
-1,-139,0,0,-133,-43,
--133,-183,-215,-296,-82,-339,
--133,-183,2,-226,-82,-339,
--82,-706,-82,-846,-215,-803,
--82,-706,-215,-663,-215,-803,
--82,-706,-82,-846,8,-733,
--82,-706,7,-593,8,-733,
--266,-593,-133,-549,-215,-436,
--266,-593,-348,-479,-215,-436,
--133,-549,7,-593,-82,-706,
--133,-549,-215,-663,-82,-706,
--266,-593,-133,-549,-215,-663,
--266,-593,-348,-706,-215,-663,
--215,-436,-133,-549,-133,-409,
--215,-436,-215,-296,-133,-409,
--215,-436,-215,-296,-348,-339,
--215,-436,-348,-479,-348,-339,
--82,-846,1,-959,-133,-1002,
--82,-846,-215,-889,-133,-1002,
--215,-803,-82,-846,-215,-889,
--215,-803,-348,-846,-215,-889,
--348,-706,-348,-846,-215,-803,
--348,-706,-215,-663,-215,-803,
--266,-959,-133,-1002,-215,-1116,
--266,-959,-348,-1072,-215,-1116,
--266,-959,-133,-1002,-215,-889,
--266,-959,-348,-846,-215,-889,
--830,-1369,-912,-1256,-779,-1212,
--830,-1369,-697,-1326,-779,-1212,
--430,-1412,-564,-1369,-481,-1256,
--430,-1412,-348,-1299,-481,-1256,
--297,-1369,-430,-1412,-348,-1299,
--297,-1369,-215,-1256,-348,-1299,
--646,-1256,-779,-1212,-697,-1326,
--646,-1256,-564,-1369,-697,-1326,
--481,-1116,-481,-1256,-348,-1212,
--481,-1116,-348,-1072,-348,-1212,
--481,-1256,-564,-1369,-646,-1256,
--481,-1256,-564,-1142,-646,-1256,
--348,-1212,-481,-1256,-348,-1299,
--348,-1212,-215,-1256,-348,-1299,
--348,-1212,-215,-1256,-215,-1116,
--348,-1212,-348,-1072,-215,-1116,
-1,-1099,1,-959,-133,-1002,
-1,-1099,-133,-1142,-133,-1002,
--82,-1212,1,-1099,-133,-1142,
--82,-1212,-215,-1256,-133,-1142,
--133,-1369,-215,-1256,-82,-1212,
--133,-1369,1,-1326,-82,-1212,
--215,-1116,-133,-1002,-133,-1142,
--215,-1116,-215,-1256,-133,-1142,
--2338,-846,-2471,-803,-2389,-689,
--2338,-846,-2256,-733,-2389,-689,
--2471,-803,-2604,-846,-2522,-959,
--2471,-803,-2389,-916,-2522,-959,
--2338,-846,-2471,-803,-2389,-916,
--2338,-846,-2256,-959,-2389,-916,
--2173,-706,-2173,-846,-2040,-803,
--2173,-706,-2040,-663,-2040,-803,
--2173,-846,-2256,-959,-2338,-846,
--2173,-846,-2256,-733,-2338,-846,
--2173,-706,-2173,-846,-2256,-733,
--2173,-706,-2256,-593,-2256,-733,
--2686,-959,-2820,-1002,-2737,-1116,
--2686,-959,-2604,-1072,-2737,-1116,
--2686,-959,-2820,-1002,-2737,-889,
--2686,-959,-2604,-846,-2737,-889,
--2522,-1099,-2522,-959,-2389,-1002,
--2522,-1099,-2389,-1142,-2389,-1002,
--2522,-959,-2604,-846,-2686,-959,
--2522,-959,-2604,-1072,-2686,-959,
--2522,-1099,-2522,-959,-2604,-1072,
--2522,-1099,-2604,-1212,-2604,-1072,
--2389,-1002,-2522,-959,-2389,-916,
--2389,-1002,-2256,-959,-2389,-916,
--2389,-1002,-2256,-959,-2256,-1099,
--2389,-1002,-2389,-1142,-2256,-1099,
--1742,-706,-1742,-846,-1609,-803,
--1742,-706,-1609,-663,-1609,-803,
--1742,-846,-1825,-959,-1907,-846,
--1742,-846,-1825,-733,-1907,-846,
--1742,-706,-1742,-846,-1825,-733,
--1742,-706,-1825,-593,-1825,-733,
--1558,-593,-1692,-549,-1609,-436,
--1558,-593,-1476,-479,-1609,-436,
--1692,-549,-1825,-593,-1742,-706,
--1692,-549,-1609,-663,-1742,-706,
--1558,-593,-1692,-549,-1609,-663,
--1558,-593,-1476,-706,-1609,-663,
--1958,-1002,-2040,-1116,-2122,-1002,
--1958,-1002,-2040,-889,-2122,-1002,
--2040,-1116,-2040,-1256,-1907,-1212,
--2040,-1116,-1907,-1072,-1907,-1212,
--1958,-1002,-2040,-1116,-1907,-1072,
--1958,-1002,-1825,-959,-1907,-1072,
--2040,-803,-1907,-846,-1907,-706,
--2040,-803,-2040,-663,-1907,-706,
--1907,-846,-1825,-959,-1958,-1002,
--1907,-846,-2040,-889,-1958,-1002,
--2040,-803,-1907,-846,-2040,-889,
--2040,-803,-2173,-846,-2040,-889,
--1907,-706,-1907,-846,-1825,-733,
--1907,-706,-1825,-593,-1825,-733,
--1907,-706,-1825,-593,-1958,-549,
--1907,-706,-2040,-663,-1958,-549,
--2040,-1116,-2040,-1256,-2173,-1212,
--2040,-1116,-2173,-1072,-2173,-1212,
--2122,-1002,-2040,-1116,-2173,-1072,
--2122,-1002,-2256,-959,-2173,-1072,
--2173,-846,-2256,-959,-2122,-1002,
--2173,-846,-2040,-889,-2122,-1002,
--2256,-1099,-2173,-1212,-2306,-1256,
--2256,-1099,-2389,-1142,-2306,-1256,
--2256,-1099,-2173,-1212,-2173,-1072,
--2256,-1099,-2256,-959,-2173,-1072,
--2655,-1369,-2737,-1256,-2604,-1212,
--2655,-1369,-2522,-1326,-2604,-1212,
--2522,-1099,-2604,-1212,-2471,-1256,
--2522,-1099,-2389,-1142,-2471,-1256,
--2604,-1212,-2737,-1256,-2737,-1116,
--2604,-1212,-2604,-1072,-2737,-1116,
--2471,-1256,-2604,-1212,-2522,-1326,
--2471,-1256,-2389,-1369,-2522,-1326,
--2471,-1256,-2389,-1369,-2306,-1256,
--2471,-1256,-2389,-1142,-2306,-1256,
--2122,-1369,-2040,-1256,-2173,-1212,
--2122,-1369,-2256,-1326,-2173,-1212,
--2306,-1256,-2173,-1212,-2256,-1326,
--2306,-1256,-2389,-1369,-2256,-1326,
--614,-846,-697,-959,-779,-846,
--614,-846,-697,-733,-779,-846,
--614,-706,-614,-846,-697,-733,
--614,-706,-697,-593,-697,-733,
--430,-593,-564,-549,-481,-436,
--430,-593,-348,-479,-481,-436,
--564,-549,-697,-593,-614,-706,
--564,-549,-481,-663,-614,-706,
--830,-1002,-912,-1116,-994,-1002,
--830,-1002,-912,-889,-994,-1002,
--830,-1002,-912,-1116,-779,-1072,
--830,-1002,-697,-959,-779,-1072,
--912,-803,-779,-846,-779,-706,
--912,-803,-912,-663,-779,-706,
--779,-846,-697,-959,-830,-1002,
--779,-846,-912,-889,-830,-1002,
--912,-803,-779,-846,-912,-889,
--912,-803,-1045,-846,-912,-889,
--779,-706,-779,-846,-697,-733,
--779,-706,-697,-593,-697,-733,
--779,-706,-697,-593,-830,-549,
--779,-706,-912,-663,-830,-549,
--297,-183,-430,-139,-348,-26,
--297,-183,-215,-69,-348,-26,
--430,-139,-564,-183,-481,-296,
--430,-139,-348,-253,-481,-296,
--297,-183,-430,-139,-348,-253,
--297,-183,-215,-296,-348,-253,
--133,-183,-215,-296,-297,-183,
--133,-183,-215,-69,-297,-183,
--133,-43,-133,-183,-215,-69,
--133,-43,-215,69,-215,-69,
--646,-296,-779,-339,-697,-453,
--646,-296,-564,-409,-697,-453,
--779,-339,-912,-296,-830,-183,
--779,-339,-697,-226,-830,-183,
--646,-296,-779,-339,-697,-226,
--646,-296,-564,-183,-697,-226,
--481,-436,-481,-296,-348,-339,
--481,-436,-348,-479,-348,-339,
--481,-296,-564,-183,-646,-296,
--481,-296,-564,-409,-646,-296,
--481,-436,-481,-296,-564,-409,
--481,-436,-564,-549,-564,-409,
--348,-339,-481,-296,-348,-253,
--348,-339,-215,-296,-348,-253,
--779,-339,-912,-296,-912,-436,
--779,-339,-779,-479,-912,-436,
--697,-453,-779,-339,-779,-479,
--697,-453,-697,-593,-779,-479,
--564,-549,-697,-593,-697,-453,
--564,-549,-564,-409,-697,-453,
--830,-549,-912,-436,-994,-549,
--830,-549,-912,-663,-994,-549,
--830,-549,-912,-436,-779,-479,
--830,-549,-697,-593,-779,-479,
--1476,-1212,-1343,-1256,-1343,-1116,
--1476,-1212,-1476,-1072,-1343,-1116,
--1343,-1256,-1261,-1369,-1394,-1412,
--1343,-1256,-1476,-1299,-1394,-1412,
--1476,-1212,-1343,-1256,-1476,-1299,
--1476,-1212,-1609,-1256,-1476,-1299,
--1343,-803,-1210,-846,-1210,-706,
--1343,-803,-1343,-663,-1210,-706,
--1210,-846,-1128,-959,-1261,-1002,
--1210,-846,-1343,-889,-1261,-1002,
--1343,-803,-1210,-846,-1343,-889,
--1343,-803,-1476,-846,-1343,-889,
--1394,-593,-1476,-706,-1558,-593,
--1394,-593,-1476,-479,-1558,-593,
--1476,-706,-1476,-846,-1343,-803,
--1476,-706,-1343,-663,-1343,-803,
--1394,-593,-1476,-706,-1343,-663,
--1394,-593,-1261,-549,-1343,-663,
--1128,-1099,-1045,-1212,-1178,-1256,
--1128,-1099,-1261,-1142,-1178,-1256,
--1045,-1212,-912,-1256,-912,-1116,
--1045,-1212,-1045,-1072,-912,-1116,
--1128,-1099,-1045,-1212,-1045,-1072,
--1128,-1099,-1128,-959,-1045,-1072,
--1343,-1116,-1261,-1002,-1394,-959,
--1343,-1116,-1476,-1072,-1394,-959,
--1261,-1002,-1128,-959,-1128,-1099,
--1261,-1002,-1261,-1142,-1128,-1099,
--1343,-1116,-1261,-1002,-1261,-1142,
--1343,-1116,-1343,-1256,-1261,-1142,
--1394,-959,-1261,-1002,-1343,-889,
--1394,-959,-1476,-846,-1343,-889,
--1394,-959,-1476,-846,-1558,-959,
--1394,-959,-1476,-1072,-1558,-959,
--1045,-1212,-912,-1256,-994,-1369,
--1045,-1212,-1128,-1326,-994,-1369,
--1178,-1256,-1045,-1212,-1128,-1326,
--1178,-1256,-1261,-1369,-1128,-1326,
--1343,-1256,-1261,-1369,-1178,-1256,
--1343,-1256,-1261,-1142,-1178,-1256,
--1210,-846,-1128,-959,-1045,-846,
--1210,-846,-1128,-733,-1045,-846,
--1210,-706,-1210,-846,-1128,-733,
--1210,-706,-1128,-593,-1128,-733,
--1394,-593,-1261,-549,-1343,-436,
--1394,-593,-1476,-479,-1343,-436,
--1261,-549,-1128,-593,-1210,-706,
--1261,-549,-1343,-663,-1210,-706,
--994,-1002,-912,-1116,-1045,-1072,
--994,-1002,-1128,-959,-1045,-1072,
--912,-803,-1045,-846,-1045,-706,
--912,-803,-912,-663,-1045,-706,
--1045,-846,-1128,-959,-994,-1002,
--1045,-846,-912,-889,-994,-1002,
--1045,-706,-1045,-846,-1128,-733,
--1045,-706,-1128,-593,-1128,-733,
--1045,-706,-1128,-593,-994,-549,
--1045,-706,-912,-663,-994,-549,
--1128,-453,-1045,-339,-1178,-296,
--1128,-453,-1261,-409,-1178,-296,
--1045,-339,-912,-296,-912,-436,
--1045,-339,-1045,-479,-912,-436,
--1128,-453,-1045,-339,-1045,-479,
--1128,-453,-1128,-593,-1045,-479,
--1261,-549,-1128,-593,-1128,-453,
--1261,-549,-1261,-409,-1128,-453,
--1343,-436,-1261,-549,-1261,-409,
--1343,-436,-1343,-296,-1261,-409,
--994,-549,-912,-436,-1045,-479,
--994,-549,-1128,-593,-1045,-479,
--1476,-1212,-1609,-1256,-1609,-1116,
--1476,-1212,-1476,-1072,-1609,-1116,
--1609,-1256,-1692,-1369,-1558,-1412,
--1609,-1256,-1476,-1299,-1558,-1412,
--1742,-846,-1825,-959,-1692,-1002,
--1742,-846,-1609,-889,-1692,-1002,
--1609,-803,-1742,-846,-1609,-889,
--1609,-803,-1476,-846,-1609,-889,
--1476,-706,-1476,-846,-1609,-803,
--1476,-706,-1609,-663,-1609,-803,
--1825,-1099,-1907,-1212,-1774,-1256,
--1825,-1099,-1692,-1142,-1774,-1256,
--1825,-1099,-1907,-1212,-1907,-1072,
--1825,-1099,-1825,-959,-1907,-1072,
--1609,-1116,-1692,-1002,-1558,-959,
--1609,-1116,-1476,-1072,-1558,-959,
--1692,-1002,-1825,-959,-1825,-1099,
--1692,-1002,-1692,-1142,-1825,-1099,
--1609,-1116,-1692,-1002,-1692,-1142,
--1609,-1116,-1609,-1256,-1692,-1142,
--1558,-959,-1692,-1002,-1609,-889,
--1558,-959,-1476,-846,-1609,-889,
--1907,-1212,-2040,-1256,-1958,-1369,
--1907,-1212,-1825,-1326,-1958,-1369,
--1774,-1256,-1907,-1212,-1825,-1326,
--1774,-1256,-1692,-1369,-1825,-1326,
--1609,-1256,-1692,-1369,-1774,-1256,
--1609,-1256,-1692,-1142,-1774,-1256,
--2471,-803,-2604,-846,-2604,-706,
--2471,-803,-2471,-663,-2604,-706,
--2389,-689,-2471,-803,-2471,-663,
--2389,-689,-2389,-549,-2471,-663,
--2173,-706,-2256,-593,-2122,-549,
--2173,-706,-2040,-663,-2122,-549,
--2256,-593,-2389,-549,-2389,-689,
--2256,-593,-2256,-733,-2389,-689,
--2686,-593,-2604,-706,-2522,-593,
--2686,-593,-2604,-479,-2522,-593,
--2604,-706,-2604,-846,-2737,-803,
--2604,-706,-2737,-663,-2737,-803,
--2686,-593,-2604,-706,-2737,-663,
--2686,-593,-2820,-549,-2737,-663,
--2522,-593,-2604,-706,-2471,-663,
--2522,-593,-2389,-549,-2471,-663,
--2522,-593,-2389,-549,-2471,-436,
--2522,-593,-2604,-479,-2471,-436,
--1825,-453,-1907,-339,-1774,-296,
--1825,-453,-1692,-409,-1774,-296,
--1907,-339,-2040,-296,-2040,-436,
--1907,-339,-1907,-479,-2040,-436,
--1825,-453,-1907,-339,-1907,-479,
--1825,-453,-1825,-593,-1907,-479,
--1692,-549,-1825,-593,-1825,-453,
--1692,-549,-1692,-409,-1825,-453,
--1609,-436,-1692,-549,-1692,-409,
--1609,-436,-1609,-296,-1692,-409,
--2173,-339,-2306,-296,-2306,-436,
--2173,-339,-2173,-479,-2306,-436,
--2306,-296,-2389,-183,-2256,-140,
--2306,-296,-2173,-253,-2256,-140,
--2173,-339,-2306,-296,-2173,-253,
--2173,-339,-2040,-296,-2173,-253,
--2122,-549,-2040,-436,-1958,-549,
--2122,-549,-2040,-663,-1958,-549,
--2040,-436,-2040,-296,-2173,-339,
--2040,-436,-2173,-479,-2173,-339,
--2122,-549,-2040,-436,-2173,-479,
--2122,-549,-2256,-593,-2173,-479,
--1958,-549,-2040,-436,-1907,-479,
--1958,-549,-1825,-593,-1907,-479,
--2306,-296,-2389,-183,-2471,-296,
--2306,-296,-2389,-409,-2471,-296,
--2306,-436,-2306,-296,-2389,-409,
--2306,-436,-2389,-549,-2389,-409,
--2256,-593,-2389,-549,-2306,-436,
--2256,-593,-2173,-479,-2306,-436,
--2471,-436,-2471,-296,-2604,-339,
--2471,-436,-2604,-479,-2604,-339,
--2471,-436,-2471,-296,-2389,-409,
--2471,-436,-2389,-549,-2389,-409,
--2686,-593,-2820,-549,-2737,-436,
--2686,-593,-2604,-479,-2737,-436,
--2737,-436,-2737,-296,-2604,-339,
--2737,-436,-2604,-479,-2604,-339,
--2655,-183,-2522,-140,-2604,-26,
--2655,-183,-2737,-70,-2604,-26,
--2522,-140,-2389,-183,-2471,-296,
--2522,-140,-2604,-253,-2471,-296,
--2655,-183,-2522,-140,-2604,-253,
--2655,-183,-2737,-296,-2604,-253,
--2820,-183,-2737,-296,-2655,-183,
--2820,-183,-2737,-70,-2655,-183,
--2604,-339,-2471,-296,-2604,-253,
--2604,-339,-2737,-296,-2604,-253,
--912,209,-994,323,-861,366,
--912,209,-779,253,-861,366,
--994,323,-1128,366,-1128,226,
--994,323,-994,183,-1128,226,
--912,209,-994,323,-994,183,
--912,209,-912,69,-994,183,
--697,226,-779,113,-646,69,
--697,226,-564,183,-646,69,
--779,113,-912,69,-912,209,
--779,113,-779,253,-912,209,
--697,226,-779,113,-779,253,
--697,226,-697,366,-779,253,
--1261,323,-1394,366,-1394,226,
--1261,323,-1261,183,-1394,226,
--1394,366,-1476,479,-1343,523,
--1394,366,-1261,409,-1343,523,
--1261,323,-1394,366,-1261,409,
--1261,323,-1128,366,-1261,409,
--1210,113,-1128,226,-1045,113,
--1210,113,-1128,-1,-1045,113,
--1128,226,-1128,366,-1261,323,
--1128,226,-1261,183,-1261,323,
--1210,113,-1128,226,-1261,183,
--1210,113,-1343,69,-1261,183,
--1045,113,-1128,226,-994,183,
--1045,113,-912,69,-994,183,
--1045,113,-912,69,-994,-43,
--1045,113,-1128,-1,-994,-43,
--430,-139,-564,-183,-564,-43,
--430,-139,-430,-3,-564,-43,
--348,-26,-430,-139,-430,-3,
--348,-26,-348,113,-430,-3,
--133,-43,-215,69,-82,113,
--133,-43,0,0,-82,113,
--215,69,-348,113,-348,-26,
--215,69,-215,-69,-348,-26,
--697,-139,-830,-183,-830,-43,
--697,-139,-697,-7,-830,-43,
--697,-139,-830,-183,-697,-226,
--697,-139,-564,-183,-697,-226,
--646,69,-564,-43,-481,69,
--646,69,-564,183,-481,69,
--564,-43,-564,-183,-697,-139,
--564,-43,-697,-7,-697,-139,
--646,69,-564,-43,-697,-7,
--646,69,-779,113,-697,-7,
--481,69,-564,-43,-430,-3,
--481,69,-348,113,-430,-3,
--481,69,-348,113,-430,226,
--481,69,-564,183,-430,226,
--830,-183,-912,-296,-994,-183,
--830,-183,-912,-69,-994,-183,
--830,-43,-830,-183,-912,-69,
--830,-43,-912,69,-912,-69,
--779,113,-912,69,-830,-43,
--779,113,-697,-7,-830,-43,
--994,-43,-994,-183,-1128,-139,
--994,-43,-1128,-1,-1128,-139,
--994,-43,-994,-183,-912,-69,
--994,-43,-912,69,-912,-69,
--1958,323,-2091,366,-2091,226,
--1958,323,-1958,183,-2091,226,
--2091,366,-2173,479,-2040,523,
--2091,366,-1958,409,-2040,523,
--1958,323,-2091,366,-1958,409,
--1958,323,-1825,366,-1958,409,
--1907,113,-1825,226,-1742,113,
--1907,113,-1825,-3,-1742,113,
--1825,226,-1825,366,-1958,323,
--1825,226,-1958,183,-1958,323,
--1907,113,-1825,226,-1958,183,
--1907,113,-2040,69,-1958,183,
--2173,619,-2256,733,-2122,776,
--2173,619,-2040,663,-2122,776,
--2256,733,-2389,776,-2389,636,
--2256,733,-2256,593,-2389,636,
--2173,619,-2256,733,-2256,593,
--2173,619,-2173,479,-2256,593,
--1958,636,-2040,523,-1907,479,
--1958,636,-1825,593,-1907,479,
--2040,523,-2173,479,-2173,619,
--2040,523,-2040,663,-2173,619,
--1958,636,-2040,523,-2040,663,
--1958,636,-1958,776,-2040,663,
--1907,479,-2040,523,-1958,409,
--1907,479,-1825,366,-1958,409,
--1907,479,-1825,366,-1742,479,
--1907,479,-1825,593,-1742,479,
--1558,-139,-1476,-26,-1394,-139,
--1558,-139,-1476,-253,-1394,-139,
--1476,-26,-1476,113,-1609,69,
--1476,-26,-1609,-70,-1609,69,
--1558,-139,-1476,-26,-1609,-70,
--1558,-139,-1692,-183,-1609,-70,
--1476,-339,-1609,-296,-1609,-436,
--1476,-339,-1476,-479,-1609,-436,
--1609,-296,-1692,-183,-1558,-139,
--1609,-296,-1476,-253,-1558,-139,
--1476,-339,-1609,-296,-1476,-253,
--1476,-339,-1343,-296,-1476,-253,
--1558,226,-1558,366,-1692,323,
--1558,226,-1692,183,-1692,323,
--1558,366,-1476,479,-1394,366,
--1558,366,-1476,253,-1394,366,
--1558,226,-1558,366,-1476,253,
--1558,226,-1476,113,-1476,253,
--1742,113,-1609,69,-1692,-43,
--1742,113,-1825,-3,-1692,-43,
--1609,69,-1476,113,-1558,226,
--1609,69,-1692,183,-1558,226,
--1742,113,-1609,69,-1692,183,
--1742,113,-1825,226,-1692,183,
--1692,-43,-1609,69,-1609,-70,
--1692,-43,-1692,-183,-1609,-70,
--1692,-43,-1692,-183,-1825,-140,
--1692,-43,-1825,-3,-1825,-140,
--1558,366,-1476,479,-1609,523,
--1558,366,-1692,409,-1609,523,
--1692,323,-1558,366,-1692,409,
--1692,323,-1825,366,-1692,409,
--1825,226,-1825,366,-1692,323,
--1825,226,-1692,183,-1692,323,
--1742,479,-1609,523,-1692,636,
--1742,479,-1825,593,-1692,636,
--1742,479,-1609,523,-1692,409,
--1742,479,-1825,366,-1692,409,
--1476,-26,-1476,113,-1343,69,
--1476,-26,-1343,-70,-1343,69,
--1394,-139,-1476,-26,-1343,-70,
--1394,-139,-1261,-183,-1343,-70,
--1476,-339,-1343,-296,-1343,-436,
--1476,-339,-1476,-479,-1343,-436,
--1343,-296,-1261,-183,-1394,-139,
--1343,-296,-1476,-253,-1394,-139,
--1394,226,-1394,366,-1476,253,
--1394,226,-1476,113,-1476,253,
--1210,113,-1343,69,-1261,-43,
--1210,113,-1128,-1,-1261,-43,
--1343,69,-1476,113,-1394,226,
--1343,69,-1261,183,-1394,226,
--1261,-43,-1343,69,-1343,-70,
--1261,-43,-1261,-183,-1343,-70,
--1261,-43,-1261,-183,-1128,-139,
--1261,-43,-1128,-1,-1128,-139,
--1045,-339,-912,-296,-994,-183,
--1045,-339,-1128,-226,-994,-183,
--1178,-296,-1045,-339,-1128,-226,
--1178,-296,-1261,-183,-1128,-226,
--1343,-296,-1261,-183,-1178,-296,
--1343,-296,-1261,-409,-1178,-296,
--1128,-139,-994,-183,-1128,-226,
--1128,-139,-1261,-183,-1128,-226,
--2091,366,-2173,479,-2256,366,
--2091,366,-2173,253,-2256,366,
--2091,226,-2091,366,-2173,253,
--2091,226,-2173,113,-2173,253,
--1907,113,-2040,69,-1958,-43,
--1907,113,-1825,-3,-1958,-43,
--2040,69,-2173,113,-2091,226,
--2040,69,-1958,183,-2091,226,
--2306,523,-2389,636,-2471,523,
--2306,523,-2389,409,-2471,523,
--2306,523,-2389,636,-2256,593,
--2306,523,-2173,479,-2256,593,
--2389,323,-2256,366,-2256,226,
--2389,323,-2389,183,-2256,226,
--2256,366,-2173,479,-2306,523,
--2256,366,-2389,409,-2306,523,
--2389,323,-2256,366,-2389,409,
--2389,323,-2522,366,-2389,409,
--2256,226,-2256,366,-2173,253,
--2256,226,-2173,113,-2173,253,
--2256,226,-2173,113,-2306,69,
--2256,226,-2389,183,-2306,69,
--1907,-339,-2040,-296,-1958,-183,
--1907,-339,-1825,-226,-1958,-183,
--1774,-296,-1907,-339,-1825,-226,
--1774,-296,-1692,-183,-1825,-226,
--1609,-296,-1692,-183,-1774,-296,
--1609,-296,-1692,-409,-1774,-296,
--2122,-183,-2256,-140,-2173,-26,
--2122,-183,-2040,-70,-2173,-26,
--2122,-183,-2256,-140,-2173,-253,
--2122,-183,-2040,-296,-2173,-253,
--1958,-43,-1958,-183,-1825,-140,
--1958,-43,-1825,-3,-1825,-140,
--1958,-183,-2040,-296,-2122,-183,
--1958,-183,-2040,-70,-2122,-183,
--1958,-43,-1958,-183,-2040,-70,
--1958,-43,-2040,69,-2040,-70,
--1825,-140,-1958,-183,-1825,-226,
--1825,-140,-1692,-183,-1825,-226,
--2256,-140,-2389,-183,-2389,-43,
--2256,-140,-2256,-4,-2389,-43,
--2173,-26,-2256,-140,-2256,-4,
--2173,-26,-2173,113,-2256,-4,
--2040,69,-2173,113,-2173,-26,
--2040,69,-2040,-70,-2173,-26,
--2306,69,-2389,-43,-2471,69,
--2306,69,-2389,183,-2471,69,
--2306,69,-2389,-43,-2256,-4,
--2306,69,-2173,113,-2256,-4,
--2686,366,-2604,479,-2522,366,
--2686,366,-2604,253,-2522,366,
--2686,226,-2686,366,-2604,253,
--2686,226,-2604,113,-2604,253,
--2737,69,-2604,113,-2686,226,
--2737,69,-2820,183,-2686,226,
--2389,636,-2389,776,-2522,733,
--2389,636,-2522,593,-2522,733,
--2471,523,-2389,636,-2522,593,
--2471,523,-2604,479,-2522,593,
--2389,323,-2522,366,-2522,226,
--2389,323,-2389,183,-2522,226,
--2522,366,-2604,479,-2471,523,
--2522,366,-2389,409,-2471,523,
--2522,226,-2522,366,-2604,253,
--2522,226,-2604,113,-2604,253,
--2522,226,-2604,113,-2471,69,
--2522,226,-2389,183,-2471,69,
--2522,-140,-2389,-183,-2389,-43,
--2522,-140,-2522,-5,-2389,-43,
--2604,-26,-2522,-140,-2522,-5,
--2604,-26,-2604,113,-2522,-5,
--2737,69,-2604,113,-2604,-26,
--2737,69,-2737,-70,-2604,-26,
--2471,69,-2389,-43,-2522,-5,
--2471,69,-2604,113,-2522,-5,
--2655,776,-2522,733,-2604,619,
--2655,776,-2737,663,-2604,619,
--2522,733,-2389,776,-2471,889,
--2522,733,-2604,846,-2471,889,
--2655,776,-2522,733,-2604,846,
--2655,776,-2737,889,-2604,846,
--2820,776,-2737,889,-2655,776,
--2820,776,-2737,663,-2655,776,
--2686,366,-2604,479,-2737,523,
--2686,366,-2820,409,-2737,523,
--2604,619,-2522,733,-2522,593,
--2604,619,-2604,479,-2522,593,
--2737,523,-2604,479,-2604,619,
--2737,523,-2737,663,-2604,619,
--994,323,-1128,366,-1045,479,
--994,323,-912,436,-1045,479,
--861,366,-994,323,-912,436,
--861,366,-779,479,-912,436,
--697,226,-697,366,-564,323,
--697,226,-564,183,-564,323,
--697,366,-779,479,-861,366,
--697,366,-779,253,-861,366,
--1210,479,-1343,523,-1261,636,
--1210,479,-1128,593,-1261,636,
--1210,479,-1343,523,-1261,409,
--1210,479,-1128,366,-1261,409,
--1045,619,-1045,479,-912,523,
--1045,619,-912,663,-912,523,
--1045,479,-1128,366,-1210,479,
--1045,479,-1128,593,-1210,479,
--1045,619,-1045,479,-1128,593,
--1045,619,-1128,733,-1128,593,
--912,523,-1045,479,-912,436,
--912,523,-779,479,-912,436,
--912,523,-779,479,-779,619,
--912,523,-912,663,-779,619,
--266,226,-266,366,-133,323,
--266,226,-133,183,-133,323,
--266,366,-348,479,-430,366,
--266,366,-348,253,-430,366,
--266,226,-266,366,-348,253,
--266,226,-348,113,-348,253,
--215,69,-348,113,-266,226,
--215,69,-133,183,-266,226,
--82,113,-215,69,-133,183,
--82,113,0,226,-133,183,
--481,523,-564,636,-646,523,
--481,523,-564,409,-646,523,
--564,636,-564,776,-430,733,
--564,636,-430,593,-430,733,
--481,523,-564,636,-430,593,
--481,523,-348,479,-430,593,
--564,323,-430,366,-430,226,
--564,323,-564,183,-430,226,
--430,366,-348,479,-481,523,
--430,366,-564,409,-481,523,
--564,323,-430,366,-564,409,
--564,323,-697,366,-564,409,
--430,226,-430,366,-348,253,
--430,226,-348,113,-348,253,
--564,636,-564,776,-697,733,
--564,636,-697,593,-697,733,
--646,523,-564,636,-697,593,
--646,523,-779,479,-697,593,
--697,366,-779,479,-646,523,
--697,366,-564,409,-646,523,
--779,619,-697,733,-830,776,
--779,619,-912,663,-830,776,
--779,619,-697,733,-697,593,
--779,619,-779,479,-697,593,
--1774,889,-1907,932,-1825,1046,
--1774,889,-1692,1002,-1825,1046,
--1907,932,-2040,889,-1958,776,
--1907,932,-1825,819,-1958,776,
--1774,889,-1907,932,-1825,819,
--1774,889,-1692,776,-1825,819,
--1609,1029,-1609,889,-1476,932,
--1609,1029,-1476,1072,-1476,932,
--1609,889,-1692,776,-1774,889,
--1609,889,-1692,1002,-1774,889,
--1609,1029,-1609,889,-1692,1002,
--1609,1029,-1692,1142,-1692,1002,
--2256,733,-2389,776,-2306,889,
--2256,733,-2173,846,-2306,889,
--2122,776,-2256,733,-2173,846,
--2122,776,-2040,889,-2173,846,
--1958,636,-1958,776,-1825,733,
--1958,636,-1825,593,-1825,733,
--1958,776,-2040,889,-2122,776,
--1958,776,-2040,663,-2122,776,
--1825,733,-1958,776,-1825,819,
--1825,733,-1692,776,-1825,819,
--1825,733,-1692,776,-1692,636,
--1825,733,-1825,593,-1692,636,
--1178,1029,-1178,889,-1045,932,
--1178,1029,-1045,1072,-1045,932,
--1178,889,-1261,776,-1343,889,
--1178,889,-1261,1002,-1343,889,
--1178,1029,-1178,889,-1261,1002,
--1178,1029,-1261,1142,-1261,1002,
--994,1142,-1128,1186,-1045,1299,
--994,1142,-912,1256,-1045,1299,
--1128,1186,-1261,1142,-1178,1029,
--1128,1186,-1045,1072,-1178,1029,
--994,1142,-1128,1186,-1045,1072,
--994,1142,-912,1029,-1045,1072,
--1394,733,-1476,619,-1558,733,
--1394,733,-1476,846,-1558,733,
--1476,619,-1476,479,-1343,523,
--1476,619,-1343,663,-1343,523,
--1394,733,-1476,619,-1343,663,
--1394,733,-1261,776,-1343,663,
--1476,932,-1343,889,-1343,1029,
--1476,932,-1476,1072,-1343,1029,
--1343,889,-1261,776,-1394,733,
--1343,889,-1476,846,-1394,733,
--1476,932,-1343,889,-1476,846,
--1476,932,-1609,889,-1476,846,
--1343,1029,-1343,889,-1261,1002,
--1343,1029,-1261,1142,-1261,1002,
--1343,1029,-1261,1142,-1394,1186,
--1343,1029,-1476,1072,-1394,1186,
--1476,619,-1476,479,-1609,523,
--1476,619,-1609,663,-1609,523,
--1558,733,-1476,619,-1609,663,
--1558,733,-1692,776,-1609,663,
--1609,889,-1692,776,-1558,733,
--1609,889,-1476,846,-1558,733,
--1692,636,-1609,523,-1609,663,
--1692,636,-1692,776,-1609,663,
--1178,889,-1261,776,-1128,733,
--1178,889,-1045,846,-1128,733,
--1045,932,-1178,889,-1045,846,
--1045,932,-912,889,-1045,846,
--994,1142,-912,1029,-830,1142,
--994,1142,-912,1256,-830,1142,
--912,1029,-912,889,-1045,932,
--912,1029,-1045,1072,-1045,932,
--1261,636,-1343,523,-1343,663,
--1261,636,-1261,776,-1343,663,
--1045,619,-1128,733,-994,776,
--1045,619,-912,663,-994,776,
--1128,733,-1261,776,-1261,636,
--1128,733,-1128,593,-1261,636,
--994,776,-1128,733,-1045,846,
--994,776,-912,889,-1045,846,
--994,776,-912,889,-830,776,
--994,776,-912,663,-830,776,
--779,932,-646,889,-646,1029,
--779,932,-779,1072,-646,1029,
--646,889,-564,776,-697,733,
--646,889,-779,846,-697,733,
--779,932,-646,889,-779,846,
--779,932,-912,889,-779,846,
--912,1029,-912,889,-779,932,
--912,1029,-779,1072,-779,932,
--830,1142,-912,1029,-779,1072,
--830,1142,-697,1186,-779,1072,
--830,776,-697,733,-779,846,
--830,776,-912,889,-779,846,
--1907,932,-2040,889,-2040,1029,
--1907,932,-1907,1072,-2040,1029,
--1825,1046,-1907,932,-1907,1072,
--1825,1046,-1825,1186,-1907,1072,
--1609,1029,-1692,1142,-1558,1186,
--1609,1029,-1476,1072,-1558,1186,
--1692,1142,-1825,1186,-1825,1046,
--1692,1142,-1692,1002,-1825,1046,
--2173,932,-2306,889,-2306,1029,
--2173,932,-2173,1072,-2306,1029,
--2173,932,-2306,889,-2173,846,
--2173,932,-2040,889,-2173,846,
--2122,1142,-2040,1029,-1958,1142,
--2122,1142,-2040,1256,-1958,1142,
--2040,1029,-2040,889,-2173,932,
--2040,1029,-2173,1072,-2173,932,
--2122,1142,-2040,1029,-2173,1072,
--2122,1142,-2256,1186,-2173,1072,
--1958,1142,-2040,1029,-1907,1072,
--1958,1142,-1825,1186,-1907,1072,
--1958,1142,-1825,1186,-1907,1299,
--1958,1142,-2040,1256,-1907,1299,
--1128,1186,-1261,1142,-1261,1282,
--1128,1186,-1128,1326,-1261,1282,
--1558,1186,-1476,1299,-1394,1186,
--1558,1186,-1476,1072,-1394,1186,
--1558,1186,-1476,1299,-1609,1256,
--1558,1186,-1692,1142,-1609,1256,
--1394,1186,-1476,1299,-1343,1256,
--1394,1186,-1261,1142,-1343,1256,
--1692,1142,-1825,1186,-1742,1299,
--1692,1142,-1609,1256,-1742,1299,
--2389,1282,-2389,1142,-2256,1186,
--2389,1282,-2256,1326,-2256,1186,
--2306,889,-2389,776,-2471,889,
--2306,889,-2389,1002,-2471,889,
--2306,1029,-2306,889,-2389,1002,
--2306,1029,-2389,1142,-2389,1002,
--2122,1142,-2256,1186,-2173,1299,
--2122,1142,-2040,1256,-2173,1299,
--2256,1186,-2389,1142,-2306,1029,
--2256,1186,-2173,1072,-2306,1029,
--2604,932,-2471,889,-2471,1029,
--2604,932,-2604,1072,-2471,1029,
--2604,932,-2471,889,-2604,846,
--2604,932,-2737,889,-2604,846,
--2737,1029,-2737,889,-2604,932,
--2737,1029,-2604,1072,-2604,932,
--2655,1142,-2737,1029,-2604,1072,
--2655,1142,-2522,1186,-2604,1072,
--2389,1282,-2389,1142,-2522,1186,
--2389,1282,-2522,1326,-2522,1186,
--2471,1029,-2471,889,-2389,1002,
--2471,1029,-2389,1142,-2389,1002,
--2655,1142,-2522,1186,-2604,1299,
--2655,1142,-2737,1256,-2604,1299,
--2522,1186,-2389,1142,-2471,1029,
--2522,1186,-2604,1072,-2471,1029,
-0,1046,0,1186,-133,1142,
-0,1046,-133,1002,-133,1142,
--82,932,0,1046,-133,1002,
--82,932,-215,889,-133,1002,
-0,733,-133,776,-133,636,
-0,733,0,593,-133,636,
--133,776,-215,889,-82,932,
--133,776,0,819,-82,932,
--266,1186,-133,1142,-215,1029,
--266,1186,-348,1072,-215,1029,
--133,1142,0,1186,-82,1299,
--133,1142,-215,1256,-82,1299,
--266,1186,-133,1142,-215,1256,
--266,1186,-348,1299,-215,1256,
--215,1029,-133,1142,-133,1002,
--215,1029,-215,889,-133,1002,
--215,1029,-215,889,-348,932,
--215,1029,-348,1072,-348,932,
--266,366,-348,479,-215,523,
--266,366,-133,409,-215,523,
--133,323,-266,366,-133,409,
--133,323,0,366,-133,409,
-0,226,0,366,-133,323,
-0,226,-133,183,-133,323,
--348,619,-430,733,-297,776,
--348,619,-215,663,-297,776,
--348,619,-430,733,-430,593,
--348,619,-348,479,-430,593,
--133,636,-215,523,-82,479,
--133,636,0,593,-82,479,
--215,523,-348,479,-348,619,
--215,523,-215,663,-348,619,
--133,636,-215,523,-215,663,
--133,636,-133,776,-215,663,
--82,479,-215,523,-133,409,
--82,479,0,366,-133,409,
--430,733,-564,776,-481,889,
--430,733,-348,846,-481,889,
--297,776,-430,733,-348,846,
--297,776,-215,889,-348,846,
--133,776,-215,889,-297,776,
--133,776,-215,663,-297,776,
--348,932,-481,889,-481,1029,
--348,932,-348,1072,-481,1029,
--348,932,-481,889,-348,846,
--348,932,-215,889,-348,846,
--779,1299,-697,1186,-830,1142,
--779,1299,-912,1256,-830,1142,
--697,1186,-564,1142,-564,1282,
--697,1186,-697,1326,-564,1282,
--266,1186,-348,1299,-430,1186,
--266,1186,-348,1072,-430,1186,
--430,1186,-348,1299,-481,1256,
--430,1186,-564,1142,-481,1256,
--430,1186,-564,1142,-481,1029,
--430,1186,-348,1072,-481,1029,
--646,889,-564,776,-481,889,
--646,889,-564,1002,-481,889,
--646,1029,-646,889,-564,1002,
--646,1029,-564,1142,-564,1002,
--697,1186,-564,1142,-646,1029,
--697,1186,-779,1072,-646,1029,
--481,1029,-481,889,-564,1002,
--481,1029,-564,1142,-564,1002,
+  [Serializable]
+  public class TileData {
 
-    };
+    public Vector2 center;
+    public int[] neighbors;
+    public int type;
+
+    public int GetRandomNeighbor() {
+      var neighbor                = neighbors[Random.Range(0, 4)];
+      if(neighbor == -1) neighbor = GetRandomNeighbor();
+      return neighbor;
+    }
+
+    public override string ToString() =>
+      $"{type}, ({center.x},{center.y}), ({neighbors[0]}, {neighbors[1]}, {neighbors[3]}, {neighbors[0]})";
+
+  }
 
 }
