+++ penrose_simulator/Assets/core/RawData.cs
@@ -1,0 +1,2711 @@
+﻿
+public static class RawData {
+
+public static readonly int[] Tiles = {
+0,0,932,2,4,869,871,
+0,66,1094,2,7,9,868,
+1,107,967,0,1,4,9,
+0,66,684,5,20,22,870,
+0,107,854,0,2,5,27,
+1,0,776,3,4,870,871,
+0,240,1107,8,9,10,32,
+0,107,1221,1,8,-1,-1,
+1,240,1221,6,7,32,-1,
+1,174,1016,1,2,6,10,
+0,282,981,6,9,28,29,
+0,199,274,13,15,56,58,
+0,240,444,13,21,23,55,
+1,133,366,11,12,15,23,
+0,0,113,16,57,661,781,
+0,66,274,11,13,16,879,
+1,107,148,14,15,57,58,
+0,323,698,19,21,26,27,
+0,497,684,19,25,60,66,
+1,389,606,17,18,21,60,
+0,107,558,3,22,23,24,
+0,282,571,12,17,19,22,
+1,174,649,3,20,21,27,
+1,107,444,12,13,20,24,
+0,0,479,20,23,882,885,
+0,456,811,18,26,29,36,
+1,323,811,17,25,27,29,
+0,215,776,4,17,22,26,
+0,415,981,10,29,34,40,
+1,348,889,10,25,26,28,
+0,804,1221,39,106,-1,-1,
+0,630,1234,38,-1,-1,-1,
+0,348,1186,6,8,33,34,
+1,456,1221,32,34,-1,-1,
+0,456,1107,28,32,33,40,
+0,712,981,37,38,114,115,
+0,564,889,25,37,40,113,
+1,605,1016,35,36,38,40,
+0,671,1107,31,35,37,39,
+1,804,1107,30,38,106,115,
+1,522,1016,28,34,36,37,
+0,887,288,43,45,185,187,
+0,1020,401,43,51,53,184,
+1,887,401,41,42,45,53,
+0,630,274,46,61,63,186,
+0,779,366,41,43,46,68,
+1,738,239,44,45,186,187,
+0,1235,558,49,51,108,110,
+0,1368,444,49,92,189,231,
+1,1235,444,47,48,51,189,
+0,979,571,52,53,54,109,
+0,1128,479,42,47,49,52,
+1,1086,606,50,51,109,110,
+1,912,479,42,43,50,54,
+0,846,571,50,53,69,70,
+0,348,366,12,56,62,64,
+1,307,239,11,55,58,64,
+0,107,35,14,16,199,352,
+0,240,148,11,16,56,198,
+0,564,523,60,62,67,68,
+1,456,558,18,19,59,62,
+0,497,274,44,63,64,65,
+0,456,444,55,59,60,63,
+1,564,366,44,61,62,68,
+1,389,239,55,56,61,65,
+0,456,148,61,64,203,206,
+0,630,684,18,67,70,113,
+1,671,558,59,66,68,70,
+0,671,444,45,59,63,67,
+0,804,698,54,70,112,116,
+1,738,606,54,66,67,69,
+0,1799,967,73,75,118,120,
+0,1932,854,73,81,83,117,
+1,1799,854,71,72,75,83,
+0,1543,981,76,94,96,119,
+0,1692,889,71,73,76,101,
+1,1650,1016,74,75,119,120,
+0,2148,698,79,81,219,221,
+0,2281,811,79,122,134,218,
+1,2148,811,77,78,81,122,
+0,1891,684,82,83,84,220,
+0,2040,776,72,77,79,82,
+1,1999,649,80,81,220,221,
+1,1825,776,72,73,80,84,
+0,1758,684,80,83,102,103,
+0,1112,981,87,89,105,107,
+0,1261,889,87,95,97,104,
+1,1220,1016,85,86,89,97,
+0,1020,1221,90,106,-1,-1,
+0,1153,1107,85,87,90,128,
+1,1020,1107,88,89,106,107,
+0,1476,733,93,95,100,101,
+0,1410,571,48,93,99,108,
+1,1368,698,91,92,95,108,
+0,1410,981,74,96,97,98,
+0,1368,811,86,91,93,96,
+1,1476,889,74,94,95,101,
+1,1302,1016,86,87,94,98,
+0,1368,1107,94,97,129,131,
+0,1543,571,92,100,103,238,
+1,1584,698,91,99,101,103,
+0,1584,811,75,91,96,100,
+0,1717,558,84,103,223,241,
+1,1650,649,84,99,100,102,
+0,1153,811,86,105,110,111,
+1,1045,889,85,104,107,111,
+0,912,1142,30,39,88,90,
+0,979,981,85,90,105,115,
+1,1302,649,47,92,93,110,
+0,1020,698,50,52,111,112,
+0,1194,684,47,52,104,108,
+1,1020,811,104,105,109,112,
+0,912,776,69,109,111,116,
+0,671,811,36,66,114,116,
+1,779,889,35,113,115,116,
+0,846,981,35,39,107,114,
+1,804,811,69,112,113,114,
+0,1974,981,72,118,124,126,
+1,1866,1059,71,117,120,126,
+0,1584,1107,74,76,129,130,
+0,1758,1094,71,76,118,132,
+0,2240,981,122,124,135,137,
+1,2173,889,78,79,121,124,
+0,2040,1142,125,126,127,136,
+0,2107,981,117,121,122,125,
+1,2148,1107,123,124,136,137,
+1,1932,1107,117,118,123,127,
+0,1932,1221,123,126,-1,-1,
+0,1194,1234,89,-1,-1,-1,
+0,1476,1186,98,119,130,131,
+1,1584,1221,119,129,132,-1,
+1,1368,1221,98,129,-1,-1,
+0,1717,1221,120,130,-1,-1,
+0,2322,1234,137,143,-1,-1,
+0,2389,889,78,135,139,144,
+1,2348,1016,121,134,137,144,
+0,2148,1221,123,125,-1,-1,
+0,2281,1107,121,125,133,135,
+0,2538,981,140,141,144,146,
+0,2496,811,134,140,285,296,
+1,2604,889,138,139,141,296,
+0,2671,981,138,140,142,-1,
+1,2630,1107,141,145,146,-1,
+0,2455,1234,133,146,-1,-1,
+1,2430,1016,134,135,138,146,
+0,2630,1221,142,-1,-1,-1,
+0,2496,1107,138,142,143,144,
+0,2363,-768,149,151,302,304,
+0,2538,-754,149,154,156,301,
+1,2430,-676,147,148,151,156,
+0,2148,-628,152,167,169,303,
+0,2322,-641,147,149,152,174,
+1,2214,-719,150,151,303,304,
+0,2604,-593,155,156,157,177,
+0,2671,-754,148,155,-1,-1,
+1,2712,-628,153,154,177,-1,
+1,2496,-628,148,149,153,157,
+0,2496,-514,153,156,175,176,
+0,1799,-374,160,162,269,270,
+0,1974,-388,160,168,170,268,
+1,1866,-466,158,159,162,170,
+0,1584,-514,163,227,316,372,
+0,1758,-501,158,160,163,315,
+1,1650,-423,161,162,227,270,
+0,2240,-388,166,168,173,174,
+0,2281,-218,166,172,272,277,
+1,2173,-296,164,165,168,272,
+0,2040,-549,150,169,170,171,
+0,2107,-388,159,164,166,169,
+1,2148,-514,150,167,168,174,
+1,1932,-514,159,160,167,171,
+0,1932,-628,167,170,320,323,
+0,2389,-296,165,173,176,180,
+1,2348,-423,164,172,174,176,
+0,2281,-514,151,164,169,173,
+0,2538,-388,157,176,178,183,
+1,2430,-423,157,172,173,175,
+0,2712,-514,153,155,178,-1,
+0,2671,-388,175,177,183,-1,
+0,2630,-104,181,182,292,293,
+0,2496,-218,172,181,183,291,
+1,2630,-218,179,180,182,183,
+0,2737,-183,179,181,-1,-1,
+1,2604,-296,175,178,180,181,
+0,1061,274,42,185,191,193,
+1,953,196,41,184,187,193,
+0,671,148,44,46,203,205,
+0,846,161,41,46,185,209,
+0,1327,274,189,191,246,248,
+1,1261,366,48,49,188,191,
+0,1128,113,192,193,194,247,
+0,1194,274,184,188,189,192,
+1,1235,148,190,191,247,248,
+1,1020,148,184,185,190,194,
+0,1020,35,190,193,210,211,
+0,323,-105,197,198,351,353,
+0,497,-91,197,204,206,350,
+1,389,-13,195,196,198,206,
+0,282,21,58,195,197,199,
+1,174,-56,57,198,352,353,
+0,763,-91,202,204,208,209,
+0,804,-261,202,207,355,361,
+1,697,-183,200,201,204,355,
+0,564,70,65,186,205,206,
+0,630,-91,196,200,202,205,
+1,671,35,186,203,204,209,
+1,456,35,65,196,197,203,
+0,912,-183,201,208,211,252,
+1,871,-56,200,207,209,211,
+0,804,35,187,200,205,208,
+0,1061,-91,194,211,250,256,
+1,953,-56,194,207,208,210,
+0,2024,274,214,216,258,260,
+0,2066,444,214,221,222,257,
+1,1958,366,212,213,216,222,
+0,1825,113,217,233,235,259,
+0,1891,274,212,214,217,240,
+1,1932,148,215,216,259,260,
+0,2322,684,78,219,262,285,
+1,2214,606,77,218,221,262,
+0,1932,558,80,82,222,223,
+0,2107,571,77,82,213,219,
+1,1932,444,213,214,220,223,
+0,1825,479,102,220,222,241,
+0,1476,-140,226,228,243,245,
+0,1543,21,226,234,236,242,
+1,1584,-105,224,225,228,236,
+0,1543,-388,161,163,229,244,
+0,1584,-218,224,226,229,270,
+1,1476,-296,227,228,244,245,
+0,1625,274,232,234,239,240,
+0,1476,366,48,232,238,246,
+1,1517,239,230,231,234,246,
+0,1717,35,215,235,236,237,
+0,1584,148,225,230,232,235,
+1,1717,148,215,233,234,240,
+1,1650,-56,225,226,233,237,
+0,1758,-91,233,236,273,276,
+0,1584,444,99,231,239,241,
+1,1692,366,230,238,240,241,
+0,1758,274,216,230,235,239,
+1,1717,444,102,223,238,239,
+0,1410,21,225,243,248,249,
+1,1368,-105,224,242,245,249,
+0,1410,-388,227,229,255,388,
+0,1368,-218,224,229,243,254,
+1,1435,239,188,231,232,248,
+0,1235,35,190,192,249,250,
+0,1368,148,188,192,242,246,
+1,1302,-56,242,243,247,250,
+0,1194,-91,210,247,249,256,
+0,1153,-374,253,254,396,397,
+0,1020,-261,207,253,256,395,
+1,1153,-261,251,252,254,256,
+0,1261,-296,245,251,253,255,
+1,1302,-423,244,254,388,397,
+1,1128,-183,210,250,252,253,
+0,2173,366,213,258,264,266,
+1,2132,239,212,257,260,266,
+0,1932,35,215,217,273,275,
+0,2066,148,212,217,258,279,
+0,2389,523,262,264,286,288,
+1,2281,558,218,219,261,264,
+0,2322,274,265,266,267,287,
+0,2281,444,257,261,262,265,
+1,2389,366,263,264,287,288,
+1,2214,239,257,258,263,267,
+0,2281,148,263,266,280,281,
+0,1932,-261,159,269,274,276,
+1,1799,-261,158,268,270,276,
+0,1692,-296,158,163,228,269,
+0,2148,-104,272,274,278,279,
+1,2148,-218,165,166,271,274,
+0,1891,-91,237,259,275,276,
+0,2040,-183,268,271,272,275,
+1,1999,-56,259,273,274,279,
+1,1825,-183,237,268,269,273,
+0,2322,-91,165,278,281,291,
+1,2214,-13,271,277,279,281,
+0,2107,21,260,271,275,278,
+0,2389,70,267,281,290,294,
+1,2281,35,267,277,278,280,
+0,2604,366,283,288,289,298,
+1,2645,239,282,284,289,-1,
+0,2712,148,283,293,-1,-1,
+0,2455,684,139,218,286,299,
+1,2496,558,261,285,288,299,
+0,2455,274,263,265,289,290,
+0,2496,444,261,265,282,286,
+1,2563,239,282,283,287,290,
+0,2496,148,280,287,289,294,
+0,2455,-91,180,277,292,294,
+1,2563,-13,179,291,293,294,
+0,2671,21,179,284,292,-1,
+1,2496,35,280,290,291,292,
+0,2630,698,296,297,299,300,
+1,2630,811,139,140,295,297,
+0,2737,776,295,296,-1,-1,
+0,2712,444,282,300,-1,-1,
+1,2563,606,285,286,295,300,
+0,2671,571,295,298,299,-1,
+0,2496,-881,148,302,308,310,
+1,2363,-881,147,301,304,310,
+0,2107,-754,150,152,320,322,
+0,2256,-846,147,152,302,326,
+0,2712,-1037,306,308,331,-1,
+1,2712,-924,305,308,-1,-1,
+0,2455,-1051,309,310,311,330,
+0,2604,-959,301,305,306,309,
+1,2563,-1086,307,308,330,331,
+1,2389,-959,301,302,307,311,
+0,2322,-1051,307,310,327,328,
+0,1676,-754,314,315,402,403,
+0,1825,-846,314,321,323,401,
+1,1784,-719,312,313,315,323,
+0,1717,-628,162,312,314,316,
+1,1584,-628,161,315,372,403,
+0,2040,-1002,319,321,325,326,
+0,1974,-1164,319,324,405,410,
+1,1932,-1037,317,318,321,405,
+0,1974,-754,171,303,322,323,
+0,1932,-924,313,317,319,322,
+1,2040,-846,303,320,321,326,
+1,1866,-719,171,313,314,320,
+0,2107,-1164,318,325,328,334,
+1,2148,-1037,317,324,326,328,
+0,2148,-924,304,317,322,325,
+0,2281,-1177,311,328,333,335,
+1,2214,-1086,311,324,325,327,
+0,2630,-1291,331,332,-1,-1,
+0,2496,-1177,307,309,332,333,
+0,2671,-1164,305,309,329,-1,
+1,2496,-1291,329,330,333,-1,
+0,2389,-1256,327,330,332,335,
+0,2148,-1291,324,335,-1,-1,
+1,2281,-1291,327,333,334,-1,
+0,548,-754,338,340,414,416,
+0,697,-846,338,346,348,413,
+1,656,-719,336,337,340,348,
+0,456,-514,341,356,358,415,
+0,589,-628,336,338,341,363,
+1,456,-628,339,340,415,416,
+0,912,-1002,344,346,390,392,
+0,846,-1164,344,376,418,441,
+1,804,-1037,342,343,346,418,
+0,846,-754,347,348,349,391,
+0,804,-924,337,342,344,347,
+1,912,-846,345,346,391,392,
+1,738,-719,337,338,345,349,
+0,804,-628,345,348,364,365,
+0,456,-218,196,351,357,359,
+1,323,-218,195,350,353,359,
+0,66,-91,57,199,428,470,
+0,215,-183,195,199,351,427,
+0,671,-374,355,357,362,363,
+1,671,-261,201,202,354,357,
+0,415,-388,339,358,359,360,
+0,564,-296,350,354,355,358,
+1,522,-423,339,356,357,363,
+1,348,-296,350,351,356,360,
+0,282,-388,356,359,432,435,
+0,846,-388,201,362,365,395,
+1,738,-466,354,361,363,365,
+0,630,-501,340,354,358,362,
+0,912,-549,349,365,394,398,
+1,804,-514,349,361,362,364,
+0,1410,-1164,368,378,380,399,
+0,1368,-1334,368,385,-1,-1,
+1,1476,-1256,366,367,399,400,
+0,1276,-754,371,373,387,389,
+0,1235,-924,371,379,381,386,
+1,1343,-846,369,370,373,381,
+0,1476,-593,161,316,374,388,
+0,1410,-754,369,371,374,403,
+1,1368,-628,372,373,388,389,
+0,1153,-1177,377,379,384,385,
+0,979,-1164,343,377,383,390,
+1,1086,-1086,375,376,379,390,
+0,1368,-1037,366,380,381,382,
+0,1194,-1051,370,375,377,380,
+1,1302,-1129,366,378,379,385,
+1,1368,-924,370,371,378,382,
+0,1476,-959,378,381,406,409,
+0,1020,-1291,376,384,-1,-1,
+1,1153,-1291,375,383,385,-1,
+0,1261,-1256,367,375,380,384,
+0,1128,-846,370,387,392,393,
+1,1169,-719,369,386,389,393,
+0,1368,-514,244,255,372,374,
+0,1235,-628,369,374,387,397,
+1,1020,-1037,342,376,377,392,
+0,979,-754,345,347,393,394,
+0,1020,-924,342,347,386,390,
+1,1086,-719,386,387,391,394,
+0,1020,-628,364,391,393,398,
+0,979,-388,252,361,396,398,
+1,1086,-466,251,395,397,398,
+0,1194,-501,251,255,389,396,
+1,1020,-514,364,394,395,396,
+0,1543,-1164,366,368,406,408,
+0,1584,-1334,368,412,-1,-1,
+0,1717,-924,313,402,407,409,
+1,1609,-846,312,401,403,409,
+0,1543,-754,312,316,373,402,
+0,1799,-1177,405,407,411,412,
+1,1866,-1086,318,319,404,407,
+0,1584,-1037,382,399,408,409,
+0,1758,-1051,401,404,405,408,
+1,1650,-1129,399,406,407,412,
+1,1584,-924,382,401,402,406,
+0,1932,-1291,318,411,-1,-1,
+1,1799,-1291,404,410,412,-1,
+0,1692,-1256,400,404,408,411,
+0,589,-924,337,414,420,422,
+1,481,-846,336,413,416,422,
+0,348,-593,339,341,432,434,
+0,415,-754,336,341,414,438,
+0,671,-1177,418,420,444,446,
+1,738,-1086,343,344,417,420,
+0,456,-1037,421,422,423,445,
+0,630,-1051,413,417,418,421,
+1,522,-1129,419,420,445,446,
+1,456,-924,413,414,419,423,
+0,348,-959,419,422,439,440,
+0,4,-339,426,427,469,471,
+0,66,-501,426,433,435,468,
+1,107,-374,424,425,427,435,
+0,107,-261,353,424,426,428,
+1,2,-183,352,427,470,471,
+0,148,-754,431,433,437,438,
+0,1,-846,431,436,473,479,
+1,41,-719,429,430,433,473,
+0,240,-514,360,415,434,435,
+0,107,-628,425,429,431,434,
+1,240,-628,415,432,433,438,
+1,174,-423,360,425,426,432,
+0,107,-924,430,437,440,450,
+1,215,-846,429,436,438,440,
+0,282,-754,416,429,434,437,
+0,240,-1037,423,440,448,453,
+1,240,-924,423,436,437,439,
+0,804,-1291,343,444,-1,-1,
+0,456,-1334,443,446,447,-1,
+1,323,-1334,442,447,-1,-1,
+1,671,-1291,417,441,446,-1,
+0,415,-1164,419,421,447,448,
+0,564,-1256,417,421,442,444,
+1,348,-1256,442,443,445,448,
+0,282,-1164,439,445,447,453,
+0,1,-1212,451,452,493,494,
+0,66,-1051,436,451,453,492,
+1,107,-1177,449,450,452,453,
+0,107,-1291,449,451,-1,-1,
+1,174,-1129,439,448,450,451,
+0,-548,-754,456,458,536,538,
+0,-589,-924,456,464,466,535,
+1,-481,-846,454,455,458,466,
+0,-348,-593,459,474,476,537,
+0,-415,-754,454,456,459,481,
+1,-456,-628,457,458,537,538,
+0,-671,-1177,462,464,487,489,
+0,-846,-1164,462,484,540,573,
+1,-738,-1086,460,461,464,540,
+0,-456,-1037,465,466,467,488,
+0,-630,-1051,455,460,462,465,
+1,-522,-1129,463,464,488,489,
+1,-456,-924,455,456,463,467,
+0,-348,-959,463,466,482,483,
+0,-66,-501,425,469,475,477,
+1,-107,-374,424,468,471,477,
+0,-66,-91,352,428,550,661,
+0,-107,-261,424,428,469,549,
+0,-148,-754,473,475,480,481,
+1,-41,-719,430,431,472,475,
+0,-240,-514,457,476,477,478,
+0,-107,-628,468,472,473,476,
+1,-240,-628,457,474,475,481,
+1,-174,-423,468,469,474,478,
+0,-282,-388,474,477,554,557,
+0,-107,-924,430,480,483,492,
+1,-215,-846,472,479,481,483,
+0,-282,-754,458,472,476,480,
+0,-240,-1037,467,483,491,495,
+1,-240,-924,467,479,480,482,
+0,-804,-1291,461,487,-1,-1,
+0,-456,-1334,486,489,490,-1,
+1,-323,-1334,485,490,-1,-1,
+1,-671,-1291,460,484,489,-1,
+0,-415,-1164,463,465,490,491,
+0,-564,-1256,460,465,485,487,
+1,-348,-1256,485,486,488,491,
+0,-282,-1164,482,488,490,495,
+0,-66,-1051,450,479,493,495,
+1,-107,-1177,449,492,494,495,
+0,-107,-1291,449,493,-1,-1,
+1,-174,-1129,482,491,492,493,
+0,-2363,-768,498,500,613,615,
+0,-2496,-881,498,505,507,612,
+1,-2363,-881,496,497,500,507,
+0,-2107,-754,501,518,520,614,
+0,-2256,-846,496,498,501,525,
+1,-2214,-719,499,500,614,615,
+0,-2712,-1037,503,505,530,-1,
+1,-2712,-924,502,505,-1,-1,
+0,-2455,-1051,506,507,508,529,
+0,-2604,-959,497,502,503,506,
+1,-2563,-1086,504,505,529,530,
+1,-2389,-959,497,498,504,508,
+0,-2322,-1051,504,507,526,527,
+0,-1676,-754,511,513,601,602,
+0,-1825,-846,511,519,521,600,
+1,-1784,-719,509,510,513,521,
+0,-1584,-514,514,569,625,692,
+0,-1717,-628,509,511,514,624,
+1,-1584,-628,512,513,569,602,
+0,-2040,-1002,517,519,524,525,
+0,-1974,-1164,517,523,604,609,
+1,-1932,-1037,515,516,519,604,
+0,-1974,-754,499,520,521,522,
+0,-1932,-924,510,515,517,520,
+1,-2040,-846,499,518,519,525,
+1,-1866,-719,510,511,518,522,
+0,-1932,-628,518,521,629,632,
+0,-2107,-1164,516,524,527,533,
+1,-2148,-1037,515,523,525,527,
+0,-2148,-924,500,515,520,524,
+0,-2281,-1177,508,527,532,534,
+1,-2214,-1086,508,523,524,526,
+0,-2630,-1291,530,531,-1,-1,
+0,-2496,-1177,504,506,531,532,
+0,-2671,-1164,502,506,528,-1,
+1,-2496,-1291,528,529,532,-1,
+0,-2389,-1256,526,529,531,534,
+0,-2148,-1291,523,534,-1,-1,
+1,-2281,-1291,526,532,533,-1,
+0,-697,-846,455,536,542,544,
+1,-656,-719,454,535,538,544,
+0,-456,-514,457,459,554,556,
+0,-589,-628,454,459,536,560,
+0,-912,-1002,540,542,587,589,
+1,-804,-1037,461,462,539,542,
+0,-846,-754,543,544,545,588,
+0,-804,-924,535,539,540,543,
+1,-912,-846,541,542,588,589,
+1,-738,-719,535,536,541,545,
+0,-804,-628,541,544,561,562,
+0,-323,-105,548,549,660,662,
+0,-456,-218,548,555,557,659,
+1,-323,-218,546,547,549,557,
+0,-215,-183,471,546,548,550,
+1,-174,-56,470,549,661,662,
+0,-671,-374,553,555,559,560,
+0,-804,-261,553,558,664,670,
+1,-671,-261,551,552,555,664,
+0,-415,-388,478,537,556,557,
+0,-564,-296,547,551,553,556,
+1,-522,-423,537,554,555,560,
+1,-348,-296,478,547,548,554,
+0,-846,-388,552,559,562,593,
+1,-738,-466,551,558,560,562,
+0,-630,-501,538,551,556,559,
+0,-912,-549,545,562,591,597,
+1,-804,-514,545,558,559,561,
+0,-1410,-1164,565,575,577,598,
+0,-1368,-1334,565,582,-1,-1,
+1,-1476,-1256,563,564,598,599,
+0,-1276,-754,568,570,584,586,
+0,-1235,-924,568,576,578,583,
+1,-1343,-846,566,567,570,578,
+0,-1476,-593,512,514,571,585,
+0,-1410,-754,566,568,571,602,
+1,-1368,-628,569,570,585,586,
+0,-1153,-1177,574,576,581,582,
+0,-979,-1164,461,574,580,587,
+1,-1086,-1086,572,573,576,587,
+0,-1368,-1037,563,577,578,579,
+0,-1194,-1051,567,572,574,577,
+1,-1302,-1129,563,575,576,582,
+1,-1368,-924,567,568,575,579,
+0,-1476,-959,575,578,605,608,
+0,-1020,-1291,573,581,-1,-1,
+1,-1153,-1291,572,580,582,-1,
+0,-1261,-1256,564,572,577,581,
+0,-1128,-846,567,584,589,590,
+1,-1169,-719,566,583,586,590,
+0,-1368,-514,569,571,596,710,
+0,-1235,-628,566,571,584,595,
+1,-1020,-1037,539,573,574,589,
+0,-979,-754,541,543,590,591,
+0,-1020,-924,539,543,583,587,
+1,-1086,-719,583,584,588,591,
+0,-1020,-628,561,588,590,597,
+0,-1153,-374,594,595,718,719,
+0,-979,-388,558,594,597,717,
+1,-1086,-466,592,593,595,597,
+0,-1194,-501,586,592,594,596,
+1,-1302,-423,585,595,710,719,
+1,-1020,-514,561,591,593,594,
+0,-1543,-1164,563,565,605,607,
+0,-1584,-1334,565,611,-1,-1,
+0,-1717,-924,510,601,606,608,
+1,-1609,-846,509,600,602,608,
+0,-1543,-754,509,514,570,601,
+0,-1799,-1177,604,606,610,611,
+1,-1866,-1086,516,517,603,606,
+0,-1584,-1037,579,598,607,608,
+0,-1758,-1051,600,603,604,607,
+1,-1650,-1129,598,605,606,611,
+1,-1584,-924,579,600,601,605,
+0,-1932,-1291,516,610,-1,-1,
+1,-1799,-1291,603,609,611,-1,
+0,-1692,-1256,599,603,607,610,
+0,-2538,-754,497,613,617,619,
+1,-2430,-676,496,612,615,619,
+0,-2148,-628,499,501,629,631,
+0,-2322,-641,496,501,613,635,
+0,-2604,-593,618,619,620,638,
+0,-2671,-754,612,618,-1,-1,
+1,-2712,-628,616,617,638,-1,
+1,-2496,-628,612,613,616,620,
+0,-2496,-514,616,619,636,637,
+0,-1799,-374,623,624,733,734,
+0,-1974,-388,623,630,632,732,
+1,-1866,-466,621,622,624,632,
+0,-1758,-501,513,621,623,625,
+1,-1650,-423,512,624,692,734,
+0,-2240,-388,628,630,634,635,
+0,-2281,-218,628,633,736,741,
+1,-2173,-296,626,627,630,736,
+0,-2040,-549,522,614,631,632,
+0,-2107,-388,622,626,628,631,
+1,-2148,-514,614,629,630,635,
+1,-1932,-514,522,622,623,629,
+0,-2389,-296,627,634,637,641,
+1,-2348,-423,626,633,635,637,
+0,-2281,-514,615,626,631,634,
+0,-2538,-388,620,637,639,644,
+1,-2430,-423,620,633,634,636,
+0,-2712,-514,616,618,639,-1,
+0,-2671,-388,636,638,644,-1,
+0,-2630,-105,642,643,756,757,
+0,-2496,-218,633,642,644,755,
+1,-2630,-218,640,641,643,644,
+0,-2737,-183,640,642,-1,-1,
+1,-2604,-296,636,639,641,642,
+0,-887,288,647,649,767,769,
+0,-1061,274,647,655,657,766,
+1,-953,196,645,646,649,657,
+0,-671,148,650,665,667,768,
+0,-846,161,645,647,650,672,
+1,-738,239,648,649,768,769,
+0,-1327,274,653,655,712,714,
+0,-1368,444,653,696,771,813,
+1,-1261,366,651,652,655,771,
+0,-1128,113,656,657,658,713,
+0,-1194,274,646,651,653,656,
+1,-1235,148,654,655,713,714,
+1,-1020,148,646,647,654,658,
+0,-1020,34,654,657,673,674,
+0,-497,-91,547,660,666,668,
+1,-389,-13,546,659,662,668,
+0,-107,35,14,470,550,781,
+0,-282,21,546,550,660,780,
+0,-763,-91,664,666,671,672,
+1,-697,-183,552,553,663,666,
+0,-564,69,648,667,668,669,
+0,-630,-91,659,663,664,667,
+1,-671,34,648,665,666,672,
+1,-456,34,659,660,665,669,
+0,-456,148,665,668,785,788,
+0,-912,-183,552,671,674,717,
+1,-871,-56,663,670,672,674,
+0,-804,34,649,663,667,671,
+0,-1061,-91,658,674,716,720,
+1,-953,-56,658,670,671,673,
+0,-2024,274,677,679,722,724,
+0,-2066,444,677,685,687,721,
+1,-1958,366,675,676,679,687,
+0,-1825,113,680,698,700,723,
+0,-1891,274,675,677,680,705,
+1,-1932,148,678,679,723,724,
+0,-2148,698,683,685,801,803,
+0,-2322,684,683,726,749,800,
+1,-2214,606,681,682,685,726,
+0,-1932,558,686,687,688,802,
+0,-2107,571,676,681,683,686,
+1,-1999,649,684,685,802,803,
+1,-1932,444,676,677,684,688,
+0,-1825,479,684,687,706,707,
+0,-1476,-140,691,693,709,711,
+0,-1543,21,691,699,701,708,
+1,-1584,-105,689,690,693,701,
+0,-1543,-388,512,625,694,710,
+0,-1584,-218,689,691,694,734,
+1,-1476,-296,692,693,710,711,
+0,-1625,274,697,699,704,705,
+0,-1476,366,652,697,703,712,
+1,-1517,239,695,696,699,712,
+0,-1717,34,678,700,701,702,
+0,-1584,148,690,695,697,700,
+1,-1717,148,678,698,699,705,
+1,-1650,-56,690,691,698,702,
+0,-1758,-91,698,701,737,740,
+0,-1584,444,696,704,707,820,
+1,-1692,366,695,703,705,707,
+0,-1758,274,679,695,700,704,
+0,-1717,558,688,707,805,823,
+1,-1717,444,688,703,704,706,
+0,-1410,21,690,709,714,715,
+1,-1368,-105,689,708,711,715,
+0,-1410,-388,585,596,692,694,
+0,-1368,-218,689,694,709,719,
+1,-1435,239,651,696,697,714,
+0,-1235,34,654,656,715,716,
+0,-1368,148,651,656,708,712,
+1,-1302,-56,708,709,713,716,
+0,-1194,-91,673,713,715,720,
+0,-1020,-261,593,670,718,720,
+1,-1153,-261,592,717,719,720,
+0,-1261,-296,592,596,711,718,
+1,-1128,-183,673,716,717,718,
+0,-2173,366,676,722,728,730,
+1,-2132,239,675,721,724,730,
+0,-1932,34,678,680,737,739,
+0,-2066,148,675,680,722,743,
+0,-2389,523,726,728,750,752,
+1,-2281,558,682,683,725,728,
+0,-2322,274,729,730,731,751,
+0,-2281,444,721,725,726,729,
+1,-2389,366,727,728,751,752,
+1,-2214,239,721,722,727,731,
+0,-2281,148,727,730,744,745,
+0,-1932,-261,622,733,738,740,
+1,-1799,-261,621,732,734,740,
+0,-1692,-296,621,625,693,733,
+0,-2148,-105,736,738,742,743,
+1,-2148,-218,627,628,735,738,
+0,-1891,-91,702,723,739,740,
+0,-2040,-183,732,735,736,739,
+1,-1999,-56,723,737,738,743,
+1,-1825,-183,702,732,733,737,
+0,-2322,-91,627,742,745,755,
+1,-2214,-13,735,741,743,745,
+0,-2107,21,724,735,739,742,
+0,-2389,69,731,745,754,758,
+1,-2281,34,731,741,742,744,
+0,-2604,366,747,752,753,763,
+1,-2645,239,746,748,753,-1,
+0,-2712,148,747,757,-1,-1,
+0,-2455,684,682,750,760,764,
+1,-2496,558,725,749,752,764,
+0,-2455,274,727,729,753,754,
+0,-2496,444,725,729,746,750,
+1,-2563,239,746,747,751,754,
+0,-2496,148,744,751,753,758,
+0,-2455,-91,641,741,756,758,
+1,-2563,-13,640,755,757,758,
+0,-2671,21,640,748,756,-1,
+1,-2496,34,744,754,755,756,
+0,-2630,698,761,762,764,765,
+0,-2496,811,749,761,856,861,
+1,-2630,811,759,760,762,861,
+0,-2737,776,759,761,-1,-1,
+0,-2712,444,746,765,-1,-1,
+1,-2563,606,749,750,759,765,
+0,-2671,571,759,763,764,-1,
+0,-1020,401,646,767,773,775,
+1,-887,401,645,766,769,775,
+0,-630,274,648,650,785,787,
+0,-779,366,645,650,767,791,
+0,-1235,558,771,773,828,830,
+1,-1235,444,652,653,770,773,
+0,-979,571,774,775,776,829,
+0,-1128,479,766,770,771,774,
+1,-1086,606,772,773,829,830,
+1,-912,479,766,767,772,776,
+0,-846,571,772,775,792,793,
+0,-199,274,779,780,878,879,
+0,-348,366,779,786,788,877,
+1,-307,239,777,778,780,788,
+0,-240,148,662,777,779,781,
+1,-107,148,14,661,780,879,
+0,-564,523,784,786,790,791,
+0,-497,684,784,789,881,886,
+1,-456,558,782,783,786,881,
+0,-497,274,669,768,787,788,
+0,-456,444,778,782,784,787,
+1,-564,366,768,785,786,791,
+1,-389,239,669,778,779,785,
+0,-630,684,783,790,793,834,
+1,-671,558,782,789,791,793,
+0,-671,444,769,782,787,790,
+0,-804,698,776,793,832,838,
+1,-738,606,776,789,790,792,
+0,-1799,967,796,798,840,842,
+0,-1932,854,796,803,804,839,
+1,-1799,854,794,795,798,804,
+0,-1543,981,799,815,817,841,
+0,-1692,889,794,796,799,822,
+1,-1650,1016,797,798,841,842,
+0,-2281,811,682,801,844,856,
+1,-2148,811,681,800,803,844,
+0,-1891,684,684,686,804,805,
+0,-2040,776,681,686,795,801,
+1,-1825,776,795,796,802,805,
+0,-1758,684,706,802,804,823,
+0,-1112,981,808,810,825,827,
+0,-1261,889,808,816,818,824,
+1,-1220,1016,806,807,810,818,
+0,-1020,1221,811,826,-1,-1,
+0,-1153,1107,806,808,811,850,
+1,-1020,1107,809,810,826,827,
+0,-1476,733,814,816,821,822,
+0,-1410,571,652,814,820,828,
+1,-1368,698,812,813,816,828,
+0,-1410,981,797,817,818,819,
+0,-1368,811,807,812,814,817,
+1,-1476,889,797,815,816,822,
+1,-1302,1016,807,808,815,819,
+0,-1368,1107,815,818,851,853,
+0,-1543,571,703,813,821,823,
+1,-1584,698,812,820,822,823,
+0,-1584,811,798,812,817,821,
+1,-1650,649,706,805,820,821,
+0,-1153,811,807,825,830,831,
+1,-1045,889,806,824,827,831,
+0,-912,1142,809,811,837,891,
+0,-979,981,806,811,825,836,
+1,-1302,649,770,813,814,830,
+0,-1020,698,772,774,831,832,
+0,-1194,684,770,774,824,828,
+1,-1020,811,824,825,829,832,
+0,-912,776,792,829,831,838,
+0,-712,981,835,836,897,898,
+0,-671,811,789,835,838,896,
+1,-779,889,833,834,836,838,
+0,-846,981,827,833,835,837,
+1,-804,1107,826,836,891,898,
+1,-804,811,792,832,834,835,
+0,-1974,981,795,840,846,848,
+1,-1866,1059,794,839,842,848,
+0,-1584,1107,797,799,851,852,
+0,-1758,1094,794,799,840,854,
+0,-2240,981,844,846,857,859,
+1,-2173,889,800,801,843,846,
+0,-2040,1142,847,848,849,858,
+0,-2107,981,839,843,844,847,
+1,-2148,1107,845,846,858,859,
+1,-1932,1107,839,840,845,849,
+0,-1932,1221,845,848,-1,-1,
+0,-1194,1234,810,-1,-1,-1,
+0,-1476,1186,819,841,852,853,
+1,-1584,1221,841,851,854,-1,
+1,-1368,1221,819,851,-1,-1,
+0,-1717,1221,842,852,-1,-1,
+0,-2322,1234,859,864,-1,-1,
+0,-2389,889,760,800,857,865,
+1,-2348,1016,843,856,859,865,
+0,-2148,1221,845,847,-1,-1,
+0,-2281,1107,843,847,855,857,
+0,-2538,981,861,862,865,867,
+1,-2604,889,760,761,860,862,
+0,-2671,981,860,861,863,-1,
+1,-2630,1107,862,866,867,-1,
+0,-2455,1234,855,867,-1,-1,
+1,-2430,1016,856,857,860,867,
+0,-2630,1221,863,-1,-1,-1,
+0,-2496,1107,860,863,864,865,
+0,-66,1094,1,869,873,875,
+1,-107,967,0,868,871,875,
+0,-66,684,3,5,882,884,
+0,-107,854,0,5,869,888,
+0,-240,1107,874,875,876,893,
+0,-107,1221,868,874,-1,-1,
+1,-240,1221,872,873,893,-1,
+1,-174,1016,868,869,872,876,
+0,-282,981,872,875,889,890,
+0,-240,444,778,878,883,885,
+1,-133,366,777,877,879,885,
+0,-66,274,15,777,781,878,
+0,-323,698,881,883,887,888,
+1,-389,606,783,784,880,883,
+0,-107,558,24,870,884,885,
+0,-282,571,877,880,881,884,
+1,-174,649,870,882,883,888,
+1,-107,444,24,877,878,882,
+0,-456,811,783,887,890,896,
+1,-323,811,880,886,888,890,
+0,-215,776,871,880,884,887,
+0,-415,981,876,890,895,899,
+1,-348,889,876,886,887,889,
+0,-804,1221,826,837,-1,-1,
+0,-630,1234,898,-1,-1,-1,
+0,-348,1186,872,874,894,895,
+1,-456,1221,893,895,-1,-1,
+0,-456,1107,889,893,894,899,
+0,-564,889,834,886,897,899,
+1,-605,1016,833,896,898,899,
+0,-671,1107,833,837,892,897,
+1,-522,1016,889,895,896,897,
+  };
+
+public static readonly int[] Mesh = {
+82,932,0,1046,-82,932,
+82,932,0,819,-82,932,
+0,1046,0,1186,133,1142,
+0,1046,133,1002,133,1142,
+82,932,0,1046,133,1002,
+82,932,215,889,133,1002,
+0,733,133,776,133,636,
+0,733,0,593,133,636,
+133,776,215,889,82,932,
+133,776,0,819,82,932,
+0,733,133,776,0,819,
+0,733,-133,776,0,819,
+266,1186,133,1142,215,1029,
+266,1186,348,1072,215,1029,
+133,1142,0,1186,82,1299,
+133,1142,215,1256,82,1299,
+266,1186,133,1142,215,1256,
+266,1186,348,1299,215,1256,
+215,1029,133,1142,133,1002,
+215,1029,215,889,133,1002,
+215,1029,215,889,348,932,
+215,1029,348,1072,348,932,
+133,323,266,366,266,226,
+133,323,133,183,266,226,
+266,366,348,479,215,523,
+266,366,133,409,215,523,
+133,323,266,366,133,409,
+133,323,0,366,133,409,
+82,113,0,226,-82,113,
+82,113,0,0,-82,113,
+0,226,0,366,133,323,
+0,226,133,183,133,323,
+82,113,0,226,133,183,
+82,113,215,69,133,183,
+348,619,430,733,297,776,
+348,619,215,663,297,776,
+430,733,564,776,564,636,
+430,733,430,593,564,636,
+348,619,430,733,430,593,
+348,619,348,479,430,593,
+133,636,215,523,82,479,
+133,636,0,593,82,479,
+215,523,348,479,348,619,
+215,523,215,663,348,619,
+133,636,215,523,215,663,
+133,636,133,776,215,663,
+82,479,215,523,133,409,
+82,479,0,366,133,409,
+82,479,0,366,-82,479,
+82,479,0,593,-82,479,
+430,733,564,776,481,889,
+430,733,348,846,481,889,
+297,776,430,733,348,846,
+297,776,215,889,348,846,
+133,776,215,889,297,776,
+133,776,215,663,297,776,
+348,932,481,889,481,1029,
+348,932,348,1072,481,1029,
+348,932,481,889,348,846,
+348,932,215,889,348,846,
+779,1299,697,1186,830,1142,
+779,1299,912,1256,830,1142,
+697,1186,564,1142,564,1282,
+697,1186,697,1326,564,1282,
+266,1186,348,1299,430,1186,
+266,1186,348,1072,430,1186,
+430,1186,348,1299,481,1256,
+430,1186,564,1142,481,1256,
+430,1186,564,1142,481,1029,
+430,1186,348,1072,481,1029,
+646,1029,646,889,779,932,
+646,1029,779,1072,779,932,
+646,889,564,776,481,889,
+646,889,564,1002,481,889,
+646,1029,646,889,564,1002,
+646,1029,564,1142,564,1002,
+697,1186,564,1142,646,1029,
+697,1186,779,1072,646,1029,
+830,1142,697,1186,779,1072,
+830,1142,912,1029,779,1072,
+481,1029,481,889,564,1002,
+481,1029,564,1142,564,1002,
+861,366,994,323,912,209,
+861,366,779,253,912,209,
+994,323,1128,366,1045,479,
+994,323,912,436,1045,479,
+861,366,994,323,912,436,
+861,366,779,479,912,436,
+697,226,697,366,564,323,
+697,226,564,183,564,323,
+697,366,779,479,861,366,
+697,366,779,253,861,366,
+697,226,697,366,779,253,
+697,226,779,113,779,253,
+1210,479,1343,523,1261,636,
+1210,479,1128,593,1261,636,
+1343,523,1476,479,1394,366,
+1343,523,1261,409,1394,366,
+1210,479,1343,523,1261,409,
+1210,479,1128,366,1261,409,
+1045,619,1045,479,912,523,
+1045,619,912,663,912,523,
+1045,479,1128,366,1210,479,
+1045,479,1128,593,1210,479,
+1045,619,1045,479,1128,593,
+1045,619,1128,733,1128,593,
+912,523,1045,479,912,436,
+912,523,779,479,912,436,
+912,523,779,479,779,619,
+912,523,912,663,779,619,
+266,366,348,479,430,366,
+266,366,348,253,430,366,
+266,226,266,366,348,253,
+266,226,348,113,348,253,
+82,113,215,69,133,-43,
+82,113,0,0,133,-43,
+215,69,348,113,266,226,
+215,69,133,183,266,226,
+481,523,564,636,646,523,
+481,523,564,409,646,523,
+481,523,564,636,430,593,
+481,523,348,479,430,593,
+564,323,430,366,430,226,
+564,323,564,183,430,226,
+430,366,348,479,481,523,
+430,366,564,409,481,523,
+564,323,430,366,564,409,
+564,323,697,366,564,409,
+430,226,430,366,348,253,
+430,226,348,113,348,253,
+430,226,348,113,481,69,
+430,226,564,183,481,69,
+564,636,564,776,697,733,
+564,636,697,593,697,733,
+646,523,564,636,697,593,
+646,523,779,479,697,593,
+697,366,779,479,646,523,
+697,366,564,409,646,523,
+779,619,697,733,830,776,
+779,619,912,663,830,776,
+779,619,697,733,697,593,
+779,619,779,479,697,593,
+1774,889,1907,932,1825,1046,
+1774,889,1692,1002,1825,1046,
+1907,932,2040,889,1958,776,
+1907,932,1825,819,1958,776,
+1774,889,1907,932,1825,819,
+1774,889,1692,776,1825,819,
+1609,1029,1609,889,1476,932,
+1609,1029,1476,1072,1476,932,
+1609,889,1692,776,1774,889,
+1609,889,1692,1002,1774,889,
+1609,1029,1609,889,1692,1002,
+1609,1029,1692,1142,1692,1002,
+2122,776,2256,733,2173,619,
+2122,776,2040,663,2173,619,
+2256,733,2389,776,2306,889,
+2256,733,2173,846,2306,889,
+2122,776,2256,733,2173,846,
+2122,776,2040,889,2173,846,
+1958,636,1958,776,1825,733,
+1958,636,1825,593,1825,733,
+1958,776,2040,889,2122,776,
+1958,776,2040,663,2122,776,
+1958,636,1958,776,2040,663,
+1958,636,2040,523,2040,663,
+1825,733,1958,776,1825,819,
+1825,733,1692,776,1825,819,
+1825,733,1692,776,1692,636,
+1825,733,1825,593,1692,636,
+1178,1029,1178,889,1045,932,
+1178,1029,1045,1072,1045,932,
+1178,889,1261,776,1343,889,
+1178,889,1261,1002,1343,889,
+1178,1029,1178,889,1261,1002,
+1178,1029,1261,1142,1261,1002,
+994,1142,1128,1186,1045,1299,
+994,1142,912,1256,1045,1299,
+1128,1186,1261,1142,1178,1029,
+1128,1186,1045,1072,1178,1029,
+994,1142,1128,1186,1045,1072,
+994,1142,912,1029,1045,1072,
+1394,733,1476,619,1558,733,
+1394,733,1476,846,1558,733,
+1476,619,1476,479,1343,523,
+1476,619,1343,663,1343,523,
+1394,733,1476,619,1343,663,
+1394,733,1261,776,1343,663,
+1476,932,1343,889,1343,1029,
+1476,932,1476,1072,1343,1029,
+1343,889,1261,776,1394,733,
+1343,889,1476,846,1394,733,
+1476,932,1343,889,1476,846,
+1476,932,1609,889,1476,846,
+1343,1029,1343,889,1261,1002,
+1343,1029,1261,1142,1261,1002,
+1343,1029,1261,1142,1394,1186,
+1343,1029,1476,1072,1394,1186,
+1476,619,1476,479,1609,523,
+1476,619,1609,663,1609,523,
+1558,733,1476,619,1609,663,
+1558,733,1692,776,1609,663,
+1609,889,1692,776,1558,733,
+1609,889,1476,846,1558,733,
+1692,636,1609,523,1742,479,
+1692,636,1825,593,1742,479,
+1692,636,1609,523,1609,663,
+1692,636,1692,776,1609,663,
+1178,889,1261,776,1128,733,
+1178,889,1045,846,1128,733,
+1045,932,1178,889,1045,846,
+1045,932,912,889,1045,846,
+994,1142,912,1029,830,1142,
+994,1142,912,1256,830,1142,
+912,1029,912,889,1045,932,
+912,1029,1045,1072,1045,932,
+1261,636,1343,523,1343,663,
+1261,636,1261,776,1343,663,
+1045,619,1128,733,994,776,
+1045,619,912,663,994,776,
+1128,733,1261,776,1261,636,
+1128,733,1128,593,1261,636,
+994,776,1128,733,1045,846,
+994,776,912,889,1045,846,
+994,776,912,889,830,776,
+994,776,912,663,830,776,
+646,889,564,776,697,733,
+646,889,779,846,697,733,
+779,932,646,889,779,846,
+779,932,912,889,779,846,
+912,1029,912,889,779,932,
+912,1029,779,1072,779,932,
+830,776,697,733,779,846,
+830,776,912,889,779,846,
+1907,932,2040,889,2040,1029,
+1907,932,1907,1072,2040,1029,
+1825,1046,1907,932,1907,1072,
+1825,1046,1825,1186,1907,1072,
+1609,1029,1692,1142,1558,1186,
+1609,1029,1476,1072,1558,1186,
+1692,1142,1825,1186,1825,1046,
+1692,1142,1692,1002,1825,1046,
+2173,932,2306,889,2306,1029,
+2173,932,2173,1072,2306,1029,
+2173,932,2306,889,2173,846,
+2173,932,2040,889,2173,846,
+2122,1142,2040,1029,1958,1142,
+2122,1142,2040,1256,1958,1142,
+2040,1029,2040,889,2173,932,
+2040,1029,2173,1072,2173,932,
+2122,1142,2040,1029,2173,1072,
+2122,1142,2256,1186,2173,1072,
+1958,1142,2040,1029,1907,1072,
+1958,1142,1825,1186,1907,1072,
+1958,1142,1825,1186,1907,1299,
+1958,1142,2040,1256,1907,1299,
+1128,1186,1261,1142,1261,1282,
+1128,1186,1128,1326,1261,1282,
+1558,1186,1476,1299,1394,1186,
+1558,1186,1476,1072,1394,1186,
+1558,1186,1476,1299,1609,1256,
+1558,1186,1692,1142,1609,1256,
+1394,1186,1476,1299,1343,1256,
+1394,1186,1261,1142,1343,1256,
+1692,1142,1825,1186,1742,1299,
+1692,1142,1609,1256,1742,1299,
+2389,1282,2389,1142,2256,1186,
+2389,1282,2256,1326,2256,1186,
+2306,889,2389,776,2471,889,
+2306,889,2389,1002,2471,889,
+2306,1029,2306,889,2389,1002,
+2306,1029,2389,1142,2389,1002,
+2122,1142,2256,1186,2173,1299,
+2122,1142,2040,1256,2173,1299,
+2256,1186,2389,1142,2306,1029,
+2256,1186,2173,1072,2306,1029,
+2604,932,2471,889,2471,1029,
+2604,932,2604,1072,2471,1029,
+2471,889,2389,776,2522,733,
+2471,889,2604,846,2522,733,
+2604,932,2471,889,2604,846,
+2604,932,2737,889,2604,846,
+2737,1029,2737,889,2604,932,
+2737,1029,2604,1072,2604,932,
+2655,1142,2737,1029,2604,1072,
+2655,1142,2522,1186,2604,1072,
+2389,1282,2389,1142,2522,1186,
+2389,1282,2522,1326,2522,1186,
+2471,1029,2471,889,2389,1002,
+2471,1029,2389,1142,2389,1002,
+2655,1142,2522,1186,2604,1299,
+2655,1142,2737,1256,2604,1299,
+2522,1186,2389,1142,2471,1029,
+2522,1186,2604,1072,2471,1029,
+2389,-689,2471,-803,2338,-846,
+2389,-689,2256,-733,2338,-846,
+2471,-803,2604,-846,2604,-706,
+2471,-803,2471,-663,2604,-706,
+2389,-689,2471,-803,2471,-663,
+2389,-689,2389,-549,2471,-663,
+2173,-706,2256,-593,2122,-549,
+2173,-706,2040,-663,2122,-549,
+2256,-593,2389,-549,2389,-689,
+2256,-593,2256,-733,2389,-689,
+2173,-706,2256,-593,2256,-733,
+2173,-706,2173,-846,2256,-733,
+2686,-593,2604,-706,2522,-593,
+2686,-593,2604,-479,2522,-593,
+2604,-706,2604,-846,2737,-803,
+2604,-706,2737,-663,2737,-803,
+2686,-593,2604,-706,2737,-663,
+2686,-593,2820,-549,2737,-663,
+2522,-593,2604,-706,2471,-663,
+2522,-593,2389,-549,2471,-663,
+2522,-593,2389,-549,2471,-436,
+2522,-593,2604,-479,2471,-436,
+1825,-453,1907,-339,1774,-296,
+1825,-453,1692,-409,1774,-296,
+1907,-339,2040,-296,2040,-436,
+1907,-339,1907,-479,2040,-436,
+1825,-453,1907,-339,1907,-479,
+1825,-453,1825,-593,1907,-479,
+1609,-436,1692,-549,1558,-593,
+1609,-436,1476,-479,1558,-593,
+1692,-549,1825,-593,1825,-453,
+1692,-549,1692,-409,1825,-453,
+1609,-436,1692,-549,1692,-409,
+1609,-436,1609,-296,1692,-409,
+2173,-339,2306,-296,2306,-436,
+2173,-339,2173,-479,2306,-436,
+2306,-296,2389,-183,2256,-139,
+2306,-296,2173,-253,2256,-139,
+2173,-339,2306,-296,2173,-253,
+2173,-339,2040,-296,2173,-253,
+2122,-549,2040,-436,1958,-549,
+2122,-549,2040,-663,1958,-549,
+2040,-436,2040,-296,2173,-339,
+2040,-436,2173,-479,2173,-339,
+2122,-549,2040,-436,2173,-479,
+2122,-549,2256,-593,2173,-479,
+1958,-549,2040,-436,1907,-479,
+1958,-549,1825,-593,1907,-479,
+1958,-549,1825,-593,1907,-706,
+1958,-549,2040,-663,1907,-706,
+2306,-296,2389,-183,2471,-296,
+2306,-296,2389,-409,2471,-296,
+2306,-436,2306,-296,2389,-409,
+2306,-436,2389,-549,2389,-409,
+2256,-593,2389,-549,2306,-436,
+2256,-593,2173,-479,2306,-436,
+2471,-436,2471,-296,2604,-339,
+2471,-436,2604,-479,2604,-339,
+2471,-436,2471,-296,2389,-409,
+2471,-436,2389,-549,2389,-409,
+2686,-593,2820,-549,2737,-436,
+2686,-593,2604,-479,2737,-436,
+2737,-436,2737,-296,2604,-339,
+2737,-436,2604,-479,2604,-339,
+2655,-183,2522,-139,2604,-26,
+2655,-183,2737,-69,2604,-26,
+2522,-139,2389,-183,2471,-296,
+2522,-139,2604,-253,2471,-296,
+2655,-183,2522,-139,2604,-253,
+2655,-183,2737,-296,2604,-253,
+2820,-183,2737,-296,2655,-183,
+2820,-183,2737,-69,2655,-183,
+2604,-339,2471,-296,2604,-253,
+2604,-339,2737,-296,2604,-253,
+994,323,1128,366,1128,226,
+994,323,994,183,1128,226,
+912,209,994,323,994,183,
+912,209,912,69,994,183,
+697,226,779,113,646,69,
+697,226,564,183,646,69,
+779,113,912,69,912,209,
+779,113,779,253,912,209,
+1261,323,1394,366,1394,226,
+1261,323,1261,183,1394,226,
+1261,323,1394,366,1261,409,
+1261,323,1128,366,1261,409,
+1210,113,1128,226,1045,113,
+1210,113,1128,1,1045,113,
+1128,226,1128,366,1261,323,
+1128,226,1261,183,1261,323,
+1210,113,1128,226,1261,183,
+1210,113,1343,69,1261,183,
+1045,113,1128,226,994,183,
+1045,113,912,69,994,183,
+1045,113,912,69,994,-43,
+1045,113,1128,1,994,-43,
+348,-26,430,-139,297,-183,
+348,-26,215,-69,297,-183,
+430,-139,564,-183,564,-43,
+430,-139,430,0,564,-43,
+348,-26,430,-139,430,0,
+348,-26,348,113,430,0,
+215,69,348,113,348,-26,
+215,69,215,-69,348,-26,
+133,-43,215,69,215,-69,
+133,-43,133,-183,215,-69,
+697,-139,830,-183,830,-43,
+697,-139,697,0,830,-43,
+830,-183,912,-296,779,-339,
+830,-183,697,-226,779,-339,
+697,-139,830,-183,697,-226,
+697,-139,564,-183,697,-226,
+646,69,564,-43,481,69,
+646,69,564,183,481,69,
+564,-43,564,-183,697,-139,
+564,-43,697,0,697,-139,
+646,69,564,-43,697,0,
+646,69,779,113,697,0,
+481,69,564,-43,430,0,
+481,69,348,113,430,0,
+830,-183,912,-296,994,-183,
+830,-183,912,-69,994,-183,
+830,-43,830,-183,912,-69,
+830,-43,912,69,912,-69,
+779,113,912,69,830,-43,
+779,113,697,0,830,-43,
+994,-43,994,-183,1128,-139,
+994,-43,1128,1,1128,-139,
+994,-43,994,-183,912,-69,
+994,-43,912,69,912,-69,
+1958,323,2091,366,2091,226,
+1958,323,1958,183,2091,226,
+2091,366,2173,479,2040,523,
+2091,366,1958,409,2040,523,
+1958,323,2091,366,1958,409,
+1958,323,1825,366,1958,409,
+1907,113,1825,226,1742,113,
+1907,113,1825,1,1742,113,
+1825,226,1825,366,1958,323,
+1825,226,1958,183,1958,323,
+1907,113,1825,226,1958,183,
+1907,113,2040,69,1958,183,
+2256,733,2389,776,2389,636,
+2256,733,2256,593,2389,636,
+2173,619,2256,733,2256,593,
+2173,619,2173,479,2256,593,
+1958,636,2040,523,1907,479,
+1958,636,1825,593,1907,479,
+2040,523,2173,479,2173,619,
+2040,523,2040,663,2173,619,
+1907,479,2040,523,1958,409,
+1907,479,1825,366,1958,409,
+1907,479,1825,366,1742,479,
+1907,479,1825,593,1742,479,
+1558,-139,1476,-26,1394,-139,
+1558,-139,1476,-253,1394,-139,
+1476,-26,1476,113,1609,69,
+1476,-26,1609,-69,1609,69,
+1558,-139,1476,-26,1609,-69,
+1558,-139,1692,-183,1609,-69,
+1476,-339,1609,-296,1609,-436,
+1476,-339,1476,-479,1609,-436,
+1609,-296,1692,-183,1558,-139,
+1609,-296,1476,-253,1558,-139,
+1476,-339,1609,-296,1476,-253,
+1476,-339,1343,-296,1476,-253,
+1558,226,1558,366,1692,323,
+1558,226,1692,183,1692,323,
+1558,366,1476,479,1394,366,
+1558,366,1476,253,1394,366,
+1558,226,1558,366,1476,253,
+1558,226,1476,113,1476,253,
+1742,113,1609,69,1692,-43,
+1742,113,1825,1,1692,-43,
+1609,69,1476,113,1558,226,
+1609,69,1692,183,1558,226,
+1742,113,1609,69,1692,183,
+1742,113,1825,226,1692,183,
+1692,-43,1609,69,1609,-69,
+1692,-43,1692,-183,1609,-69,
+1692,-43,1692,-183,1825,-139,
+1692,-43,1825,1,1825,-139,
+1558,366,1476,479,1609,523,
+1558,366,1692,409,1609,523,
+1692,323,1558,366,1692,409,
+1692,323,1825,366,1692,409,
+1825,226,1825,366,1692,323,
+1825,226,1692,183,1692,323,
+1742,479,1609,523,1692,409,
+1742,479,1825,366,1692,409,
+1476,-26,1476,113,1343,69,
+1476,-26,1343,-69,1343,69,
+1394,-139,1476,-26,1343,-69,
+1394,-139,1261,-183,1343,-69,
+1476,-339,1343,-296,1343,-436,
+1476,-339,1476,-479,1343,-436,
+1343,-296,1261,-183,1394,-139,
+1343,-296,1476,-253,1394,-139,
+1394,226,1394,366,1476,253,
+1394,226,1476,113,1476,253,
+1210,113,1343,69,1261,-43,
+1210,113,1128,1,1261,-43,
+1343,69,1476,113,1394,226,
+1343,69,1261,183,1394,226,
+1261,-43,1343,69,1343,-69,
+1261,-43,1261,-183,1343,-69,
+1261,-43,1261,-183,1128,-139,
+1261,-43,1128,1,1128,-139,
+1178,-296,1045,-339,1128,-453,
+1178,-296,1261,-409,1128,-453,
+1045,-339,912,-296,994,-183,
+1045,-339,1128,-226,994,-183,
+1178,-296,1045,-339,1128,-226,
+1178,-296,1261,-183,1128,-226,
+1343,-296,1261,-183,1178,-296,
+1343,-296,1261,-409,1178,-296,
+1343,-436,1343,-296,1261,-409,
+1343,-436,1261,-549,1261,-409,
+1128,-139,994,-183,1128,-226,
+1128,-139,1261,-183,1128,-226,
+2091,366,2173,479,2256,366,
+2091,366,2173,253,2256,366,
+2091,226,2091,366,2173,253,
+2091,226,2173,113,2173,253,
+1907,113,2040,69,1958,-43,
+1907,113,1825,1,1958,-43,
+2040,69,2173,113,2091,226,
+2040,69,1958,183,2091,226,
+2306,523,2389,636,2471,523,
+2306,523,2389,409,2471,523,
+2306,523,2389,636,2256,593,
+2306,523,2173,479,2256,593,
+2389,323,2256,366,2256,226,
+2389,323,2389,183,2256,226,
+2256,366,2173,479,2306,523,
+2256,366,2389,409,2306,523,
+2389,323,2256,366,2389,409,
+2389,323,2522,366,2389,409,
+2256,226,2256,366,2173,253,
+2256,226,2173,113,2173,253,
+2256,226,2173,113,2306,69,
+2256,226,2389,183,2306,69,
+1907,-339,2040,-296,1958,-183,
+1907,-339,1825,-226,1958,-183,
+1774,-296,1907,-339,1825,-226,
+1774,-296,1692,-183,1825,-226,
+1609,-296,1692,-183,1774,-296,
+1609,-296,1692,-409,1774,-296,
+2122,-183,2256,-139,2173,-26,
+2122,-183,2040,-69,2173,-26,
+2122,-183,2256,-139,2173,-253,
+2122,-183,2040,-296,2173,-253,
+1958,-43,1958,-183,1825,-139,
+1958,-43,1825,1,1825,-139,
+1958,-183,2040,-296,2122,-183,
+1958,-183,2040,-69,2122,-183,
+1958,-43,1958,-183,2040,-69,
+1958,-43,2040,69,2040,-69,
+1825,-139,1958,-183,1825,-226,
+1825,-139,1692,-183,1825,-226,
+2256,-139,2389,-183,2389,-43,
+2256,-139,2256,1,2389,-43,
+2173,-26,2256,-139,2256,1,
+2173,-26,2173,113,2256,1,
+2040,69,2173,113,2173,-26,
+2040,69,2040,-69,2173,-26,
+2306,69,2389,-43,2471,69,
+2306,69,2389,183,2471,69,
+2306,69,2389,-43,2256,1,
+2306,69,2173,113,2256,1,
+2686,366,2604,479,2522,366,
+2686,366,2604,253,2522,366,
+2686,226,2686,366,2604,253,
+2686,226,2604,113,2604,253,
+2737,69,2604,113,2686,226,
+2737,69,2820,183,2686,226,
+2389,636,2389,776,2522,733,
+2389,636,2522,593,2522,733,
+2471,523,2389,636,2522,593,
+2471,523,2604,479,2522,593,
+2389,323,2522,366,2522,226,
+2389,323,2389,183,2522,226,
+2522,366,2604,479,2471,523,
+2522,366,2389,409,2471,523,
+2522,226,2522,366,2604,253,
+2522,226,2604,113,2604,253,
+2522,226,2604,113,2471,69,
+2522,226,2389,183,2471,69,
+2522,-139,2389,-183,2389,-43,
+2522,-139,2522,1,2389,-43,
+2604,-26,2522,-139,2522,1,
+2604,-26,2604,113,2522,1,
+2737,69,2604,113,2604,-26,
+2737,69,2737,-69,2604,-26,
+2471,69,2389,-43,2522,1,
+2471,69,2604,113,2522,1,
+2655,776,2522,733,2604,619,
+2655,776,2737,663,2604,619,
+2655,776,2522,733,2604,846,
+2655,776,2737,889,2604,846,
+2820,776,2737,889,2655,776,
+2820,776,2737,663,2655,776,
+2686,366,2604,479,2737,523,
+2686,366,2820,409,2737,523,
+2604,619,2522,733,2522,593,
+2604,619,2604,479,2522,593,
+2737,523,2604,479,2604,619,
+2737,523,2737,663,2604,619,
+2471,-803,2604,-846,2522,-959,
+2471,-803,2389,-916,2522,-959,
+2338,-846,2471,-803,2389,-916,
+2338,-846,2256,-959,2389,-916,
+2173,-706,2173,-846,2040,-803,
+2173,-706,2040,-663,2040,-803,
+2173,-846,2256,-959,2338,-846,
+2173,-846,2256,-733,2338,-846,
+2686,-959,2820,-1002,2737,-1116,
+2686,-959,2604,-1072,2737,-1116,
+2686,-959,2820,-1002,2737,-889,
+2686,-959,2604,-846,2737,-889,
+2522,-1099,2522,-959,2389,-1002,
+2522,-1099,2389,-1142,2389,-1002,
+2522,-959,2604,-846,2686,-959,
+2522,-959,2604,-1072,2686,-959,
+2522,-1099,2522,-959,2604,-1072,
+2522,-1099,2604,-1212,2604,-1072,
+2389,-1002,2522,-959,2389,-916,
+2389,-1002,2256,-959,2389,-916,
+2389,-1002,2256,-959,2256,-1099,
+2389,-1002,2389,-1142,2256,-1099,
+1742,-706,1742,-846,1609,-803,
+1742,-706,1609,-663,1609,-803,
+1742,-846,1825,-959,1907,-846,
+1742,-846,1825,-733,1907,-846,
+1742,-706,1742,-846,1825,-733,
+1742,-706,1825,-593,1825,-733,
+1692,-549,1825,-593,1742,-706,
+1692,-549,1609,-663,1742,-706,
+1558,-593,1692,-549,1609,-663,
+1558,-593,1476,-706,1609,-663,
+1958,-1002,2040,-1116,2122,-1002,
+1958,-1002,2040,-889,2122,-1002,
+2040,-1116,2040,-1256,1907,-1212,
+2040,-1116,1907,-1072,1907,-1212,
+1958,-1002,2040,-1116,1907,-1072,
+1958,-1002,1825,-959,1907,-1072,
+2040,-803,1907,-846,1907,-706,
+2040,-803,2040,-663,1907,-706,
+1907,-846,1825,-959,1958,-1002,
+1907,-846,2040,-889,1958,-1002,
+2040,-803,1907,-846,2040,-889,
+2040,-803,2173,-846,2040,-889,
+1907,-706,1907,-846,1825,-733,
+1907,-706,1825,-593,1825,-733,
+2040,-1116,2040,-1256,2173,-1212,
+2040,-1116,2173,-1072,2173,-1212,
+2122,-1002,2040,-1116,2173,-1072,
+2122,-1002,2256,-959,2173,-1072,
+2173,-846,2256,-959,2122,-1002,
+2173,-846,2040,-889,2122,-1002,
+2256,-1099,2173,-1212,2306,-1256,
+2256,-1099,2389,-1142,2306,-1256,
+2256,-1099,2173,-1212,2173,-1072,
+2256,-1099,2256,-959,2173,-1072,
+2655,-1369,2737,-1256,2604,-1212,
+2655,-1369,2522,-1326,2604,-1212,
+2522,-1099,2604,-1212,2471,-1256,
+2522,-1099,2389,-1142,2471,-1256,
+2604,-1212,2737,-1256,2737,-1116,
+2604,-1212,2604,-1072,2737,-1116,
+2471,-1256,2604,-1212,2522,-1326,
+2471,-1256,2389,-1369,2522,-1326,
+2471,-1256,2389,-1369,2306,-1256,
+2471,-1256,2389,-1142,2306,-1256,
+2122,-1369,2040,-1256,2173,-1212,
+2122,-1369,2256,-1326,2173,-1212,
+2306,-1256,2173,-1212,2256,-1326,
+2306,-1256,2389,-1369,2256,-1326,
+614,-706,614,-846,481,-803,
+614,-706,481,-663,481,-803,
+614,-846,697,-959,779,-846,
+614,-846,697,-733,779,-846,
+614,-706,614,-846,697,-733,
+614,-706,697,-593,697,-733,
+430,-593,564,-549,481,-436,
+430,-593,348,-479,481,-436,
+564,-549,697,-593,614,-706,
+564,-549,481,-663,614,-706,
+430,-593,564,-549,481,-663,
+430,-593,348,-706,481,-663,
+830,-1002,912,-1116,994,-1002,
+830,-1002,912,-889,994,-1002,
+912,-1116,912,-1256,779,-1212,
+912,-1116,779,-1072,779,-1212,
+830,-1002,912,-1116,779,-1072,
+830,-1002,697,-959,779,-1072,
+912,-803,779,-846,779,-706,
+912,-803,912,-663,779,-706,
+779,-846,697,-959,830,-1002,
+779,-846,912,-889,830,-1002,
+912,-803,779,-846,912,-889,
+912,-803,1045,-846,912,-889,
+779,-706,779,-846,697,-733,
+779,-706,697,-593,697,-733,
+779,-706,697,-593,830,-549,
+779,-706,912,-663,830,-549,
+430,-139,564,-183,481,-296,
+430,-139,348,-253,481,-296,
+297,-183,430,-139,348,-253,
+297,-183,215,-296,348,-253,
+133,-43,133,-183,1,-139,
+133,-43,0,0,1,-139,
+133,-183,215,-296,297,-183,
+133,-183,215,-69,297,-183,
+646,-296,779,-339,697,-453,
+646,-296,564,-409,697,-453,
+646,-296,779,-339,697,-226,
+646,-296,564,-183,697,-226,
+481,-436,481,-296,348,-339,
+481,-436,348,-479,348,-339,
+481,-296,564,-183,646,-296,
+481,-296,564,-409,646,-296,
+481,-436,481,-296,564,-409,
+481,-436,564,-549,564,-409,
+348,-339,481,-296,348,-253,
+348,-339,215,-296,348,-253,
+348,-339,215,-296,215,-436,
+348,-339,348,-479,215,-436,
+779,-339,912,-296,912,-436,
+779,-339,779,-479,912,-436,
+697,-453,779,-339,779,-479,
+697,-453,697,-593,779,-479,
+564,-549,697,-593,697,-453,
+564,-549,564,-409,697,-453,
+830,-549,912,-436,994,-549,
+830,-549,912,-663,994,-549,
+830,-549,912,-436,779,-479,
+830,-549,697,-593,779,-479,
+1476,-1212,1343,-1256,1343,-1116,
+1476,-1212,1476,-1072,1343,-1116,
+1343,-1256,1261,-1369,1394,-1412,
+1343,-1256,1476,-1299,1394,-1412,
+1476,-1212,1343,-1256,1476,-1299,
+1476,-1212,1609,-1256,1476,-1299,
+1343,-803,1210,-846,1210,-706,
+1343,-803,1343,-663,1210,-706,
+1210,-846,1128,-959,1261,-1002,
+1210,-846,1343,-889,1261,-1002,
+1343,-803,1210,-846,1343,-889,
+1343,-803,1476,-846,1343,-889,
+1394,-593,1476,-706,1558,-593,
+1394,-593,1476,-479,1558,-593,
+1476,-706,1476,-846,1343,-803,
+1476,-706,1343,-663,1343,-803,
+1394,-593,1476,-706,1343,-663,
+1394,-593,1261,-549,1343,-663,
+1128,-1099,1045,-1212,1178,-1256,
+1128,-1099,1261,-1142,1178,-1256,
+1045,-1212,912,-1256,912,-1116,
+1045,-1212,1045,-1072,912,-1116,
+1128,-1099,1045,-1212,1045,-1072,
+1128,-1099,1128,-959,1045,-1072,
+1343,-1116,1261,-1002,1394,-959,
+1343,-1116,1476,-1072,1394,-959,
+1261,-1002,1128,-959,1128,-1099,
+1261,-1002,1261,-1142,1128,-1099,
+1343,-1116,1261,-1002,1261,-1142,
+1343,-1116,1343,-1256,1261,-1142,
+1394,-959,1261,-1002,1343,-889,
+1394,-959,1476,-846,1343,-889,
+1394,-959,1476,-846,1558,-959,
+1394,-959,1476,-1072,1558,-959,
+1045,-1212,912,-1256,994,-1369,
+1045,-1212,1128,-1326,994,-1369,
+1178,-1256,1045,-1212,1128,-1326,
+1178,-1256,1261,-1369,1128,-1326,
+1343,-1256,1261,-1369,1178,-1256,
+1343,-1256,1261,-1142,1178,-1256,
+1210,-846,1128,-959,1045,-846,
+1210,-846,1128,-733,1045,-846,
+1210,-706,1210,-846,1128,-733,
+1210,-706,1128,-593,1128,-733,
+1394,-593,1261,-549,1343,-436,
+1394,-593,1476,-479,1343,-436,
+1261,-549,1128,-593,1210,-706,
+1261,-549,1343,-663,1210,-706,
+994,-1002,912,-1116,1045,-1072,
+994,-1002,1128,-959,1045,-1072,
+912,-803,1045,-846,1045,-706,
+912,-803,912,-663,1045,-706,
+1045,-846,1128,-959,994,-1002,
+1045,-846,912,-889,994,-1002,
+1045,-706,1045,-846,1128,-733,
+1045,-706,1128,-593,1128,-733,
+1045,-706,1128,-593,994,-549,
+1045,-706,912,-663,994,-549,
+1045,-339,912,-296,912,-436,
+1045,-339,1045,-479,912,-436,
+1128,-453,1045,-339,1045,-479,
+1128,-453,1128,-593,1045,-479,
+1261,-549,1128,-593,1128,-453,
+1261,-549,1261,-409,1128,-453,
+994,-549,912,-436,1045,-479,
+994,-549,1128,-593,1045,-479,
+1476,-1212,1609,-1256,1609,-1116,
+1476,-1212,1476,-1072,1609,-1116,
+1609,-1256,1692,-1369,1558,-1412,
+1609,-1256,1476,-1299,1558,-1412,
+1742,-846,1825,-959,1692,-1002,
+1742,-846,1609,-889,1692,-1002,
+1609,-803,1742,-846,1609,-889,
+1609,-803,1476,-846,1609,-889,
+1476,-706,1476,-846,1609,-803,
+1476,-706,1609,-663,1609,-803,
+1825,-1099,1907,-1212,1774,-1256,
+1825,-1099,1692,-1142,1774,-1256,
+1825,-1099,1907,-1212,1907,-1072,
+1825,-1099,1825,-959,1907,-1072,
+1609,-1116,1692,-1002,1558,-959,
+1609,-1116,1476,-1072,1558,-959,
+1692,-1002,1825,-959,1825,-1099,
+1692,-1002,1692,-1142,1825,-1099,
+1609,-1116,1692,-1002,1692,-1142,
+1609,-1116,1609,-1256,1692,-1142,
+1558,-959,1692,-1002,1609,-889,
+1558,-959,1476,-846,1609,-889,
+1907,-1212,2040,-1256,1958,-1369,
+1907,-1212,1825,-1326,1958,-1369,
+1774,-1256,1907,-1212,1825,-1326,
+1774,-1256,1692,-1369,1825,-1326,
+1609,-1256,1692,-1369,1774,-1256,
+1609,-1256,1692,-1142,1774,-1256,
+614,-846,697,-959,564,-1002,
+614,-846,481,-889,564,-1002,
+481,-803,614,-846,481,-889,
+481,-803,348,-846,481,-889,
+430,-593,348,-706,266,-593,
+430,-593,348,-479,266,-593,
+348,-706,348,-846,481,-803,
+348,-706,481,-663,481,-803,
+697,-1099,779,-1212,646,-1256,
+697,-1099,564,-1142,646,-1256,
+697,-1099,779,-1212,779,-1072,
+697,-1099,697,-959,779,-1072,
+481,-1116,564,-1002,430,-959,
+481,-1116,348,-1072,430,-959,
+564,-1002,697,-959,697,-1099,
+564,-1002,564,-1142,697,-1099,
+481,-1116,564,-1002,564,-1142,
+481,-1116,481,-1256,564,-1142,
+430,-959,564,-1002,481,-889,
+430,-959,348,-846,481,-889,
+430,-959,348,-846,266,-959,
+430,-959,348,-1072,266,-959,
+82,-339,5,-453,-82,-339,
+82,-339,2,-226,-82,-339,
+5,-453,7,-593,133,-549,
+5,-453,133,-409,133,-549,
+82,-339,5,-453,133,-409,
+82,-339,215,-296,133,-409,
+133,-183,215,-296,82,-339,
+133,-183,2,-226,82,-339,
+1,-139,133,-183,2,-226,
+1,-139,-133,-183,2,-226,
+82,-706,82,-846,215,-803,
+82,-706,215,-663,215,-803,
+82,-846,1,-959,-82,-846,
+82,-846,8,-733,-82,-846,
+82,-706,82,-846,8,-733,
+82,-706,7,-593,8,-733,
+266,-593,133,-549,215,-436,
+266,-593,348,-479,215,-436,
+133,-549,7,-593,82,-706,
+133,-549,215,-663,82,-706,
+266,-593,133,-549,215,-663,
+266,-593,348,-706,215,-663,
+215,-436,133,-549,133,-409,
+215,-436,215,-296,133,-409,
+82,-846,1,-959,133,-1002,
+82,-846,215,-889,133,-1002,
+215,-803,82,-846,215,-889,
+215,-803,348,-846,215,-889,
+348,-706,348,-846,215,-803,
+348,-706,215,-663,215,-803,
+266,-959,133,-1002,215,-1116,
+266,-959,348,-1072,215,-1116,
+266,-959,133,-1002,215,-889,
+266,-959,348,-846,215,-889,
+830,-1369,912,-1256,779,-1212,
+830,-1369,697,-1326,779,-1212,
+430,-1412,564,-1369,481,-1256,
+430,-1412,348,-1299,481,-1256,
+297,-1369,430,-1412,348,-1299,
+297,-1369,215,-1256,348,-1299,
+646,-1256,779,-1212,697,-1326,
+646,-1256,564,-1369,697,-1326,
+481,-1116,481,-1256,348,-1212,
+481,-1116,348,-1072,348,-1212,
+481,-1256,564,-1369,646,-1256,
+481,-1256,564,-1142,646,-1256,
+348,-1212,481,-1256,348,-1299,
+348,-1212,215,-1256,348,-1299,
+348,-1212,215,-1256,215,-1116,
+348,-1212,348,-1072,215,-1116,
+82,-1212,1,-1099,-82,-1212,
+82,-1212,1,-1326,-82,-1212,
+1,-1099,1,-959,133,-1002,
+1,-1099,133,-1142,133,-1002,
+82,-1212,1,-1099,133,-1142,
+82,-1212,215,-1256,133,-1142,
+133,-1369,215,-1256,82,-1212,
+133,-1369,1,-1326,82,-1212,
+215,-1116,133,-1002,133,-1142,
+215,-1116,215,-1256,133,-1142,
+-481,-803,-614,-846,-614,-706,
+-481,-803,-481,-663,-614,-706,
+-614,-846,-697,-959,-564,-1002,
+-614,-846,-481,-889,-564,-1002,
+-481,-803,-614,-846,-481,-889,
+-481,-803,-348,-846,-481,-889,
+-430,-593,-348,-706,-266,-593,
+-430,-593,-348,-479,-266,-593,
+-348,-706,-348,-846,-481,-803,
+-348,-706,-481,-663,-481,-803,
+-430,-593,-348,-706,-481,-663,
+-430,-593,-564,-549,-481,-663,
+-697,-1099,-779,-1212,-646,-1256,
+-697,-1099,-564,-1142,-646,-1256,
+-779,-1212,-912,-1256,-912,-1116,
+-779,-1212,-779,-1072,-912,-1116,
+-697,-1099,-779,-1212,-779,-1072,
+-697,-1099,-697,-959,-779,-1072,
+-481,-1116,-564,-1002,-430,-959,
+-481,-1116,-348,-1072,-430,-959,
+-564,-1002,-697,-959,-697,-1099,
+-564,-1002,-564,-1142,-697,-1099,
+-481,-1116,-564,-1002,-564,-1142,
+-481,-1116,-481,-1256,-564,-1142,
+-430,-959,-564,-1002,-481,-889,
+-430,-959,-348,-846,-481,-889,
+-430,-959,-348,-846,-266,-959,
+-430,-959,-348,-1072,-266,-959,
+5,-453,7,-593,-133,-549,
+5,-453,-133,-409,-133,-549,
+-82,-339,5,-453,-133,-409,
+-82,-339,-215,-296,-133,-409,
+1,-139,-133,-183,-133,-43,
+1,-139,0,0,-133,-43,
+-133,-183,-215,-296,-82,-339,
+-133,-183,2,-226,-82,-339,
+-82,-706,-82,-846,-215,-803,
+-82,-706,-215,-663,-215,-803,
+-82,-706,-82,-846,8,-733,
+-82,-706,7,-593,8,-733,
+-266,-593,-133,-549,-215,-436,
+-266,-593,-348,-479,-215,-436,
+-133,-549,7,-593,-82,-706,
+-133,-549,-215,-663,-82,-706,
+-266,-593,-133,-549,-215,-663,
+-266,-593,-348,-706,-215,-663,
+-215,-436,-133,-549,-133,-409,
+-215,-436,-215,-296,-133,-409,
+-215,-436,-215,-296,-348,-339,
+-215,-436,-348,-479,-348,-339,
+-82,-846,1,-959,-133,-1002,
+-82,-846,-215,-889,-133,-1002,
+-215,-803,-82,-846,-215,-889,
+-215,-803,-348,-846,-215,-889,
+-348,-706,-348,-846,-215,-803,
+-348,-706,-215,-663,-215,-803,
+-266,-959,-133,-1002,-215,-1116,
+-266,-959,-348,-1072,-215,-1116,
+-266,-959,-133,-1002,-215,-889,
+-266,-959,-348,-846,-215,-889,
+-830,-1369,-912,-1256,-779,-1212,
+-830,-1369,-697,-1326,-779,-1212,
+-430,-1412,-564,-1369,-481,-1256,
+-430,-1412,-348,-1299,-481,-1256,
+-297,-1369,-430,-1412,-348,-1299,
+-297,-1369,-215,-1256,-348,-1299,
+-646,-1256,-779,-1212,-697,-1326,
+-646,-1256,-564,-1369,-697,-1326,
+-481,-1116,-481,-1256,-348,-1212,
+-481,-1116,-348,-1072,-348,-1212,
+-481,-1256,-564,-1369,-646,-1256,
+-481,-1256,-564,-1142,-646,-1256,
+-348,-1212,-481,-1256,-348,-1299,
+-348,-1212,-215,-1256,-348,-1299,
+-348,-1212,-215,-1256,-215,-1116,
+-348,-1212,-348,-1072,-215,-1116,
+1,-1099,1,-959,-133,-1002,
+1,-1099,-133,-1142,-133,-1002,
+-82,-1212,1,-1099,-133,-1142,
+-82,-1212,-215,-1256,-133,-1142,
+-133,-1369,-215,-1256,-82,-1212,
+-133,-1369,1,-1326,-82,-1212,
+-215,-1116,-133,-1002,-133,-1142,
+-215,-1116,-215,-1256,-133,-1142,
+-2338,-846,-2471,-803,-2389,-689,
+-2338,-846,-2256,-733,-2389,-689,
+-2471,-803,-2604,-846,-2522,-959,
+-2471,-803,-2389,-916,-2522,-959,
+-2338,-846,-2471,-803,-2389,-916,
+-2338,-846,-2256,-959,-2389,-916,
+-2173,-706,-2173,-846,-2040,-803,
+-2173,-706,-2040,-663,-2040,-803,
+-2173,-846,-2256,-959,-2338,-846,
+-2173,-846,-2256,-733,-2338,-846,
+-2173,-706,-2173,-846,-2256,-733,
+-2173,-706,-2256,-593,-2256,-733,
+-2686,-959,-2820,-1002,-2737,-1116,
+-2686,-959,-2604,-1072,-2737,-1116,
+-2686,-959,-2820,-1002,-2737,-889,
+-2686,-959,-2604,-846,-2737,-889,
+-2522,-1099,-2522,-959,-2389,-1002,
+-2522,-1099,-2389,-1142,-2389,-1002,
+-2522,-959,-2604,-846,-2686,-959,
+-2522,-959,-2604,-1072,-2686,-959,
+-2522,-1099,-2522,-959,-2604,-1072,
+-2522,-1099,-2604,-1212,-2604,-1072,
+-2389,-1002,-2522,-959,-2389,-916,
+-2389,-1002,-2256,-959,-2389,-916,
+-2389,-1002,-2256,-959,-2256,-1099,
+-2389,-1002,-2389,-1142,-2256,-1099,
+-1742,-706,-1742,-846,-1609,-803,
+-1742,-706,-1609,-663,-1609,-803,
+-1742,-846,-1825,-959,-1907,-846,
+-1742,-846,-1825,-733,-1907,-846,
+-1742,-706,-1742,-846,-1825,-733,
+-1742,-706,-1825,-593,-1825,-733,
+-1558,-593,-1692,-549,-1609,-436,
+-1558,-593,-1476,-479,-1609,-436,
+-1692,-549,-1825,-593,-1742,-706,
+-1692,-549,-1609,-663,-1742,-706,
+-1558,-593,-1692,-549,-1609,-663,
+-1558,-593,-1476,-706,-1609,-663,
+-1958,-1002,-2040,-1116,-2122,-1002,
+-1958,-1002,-2040,-889,-2122,-1002,
+-2040,-1116,-2040,-1256,-1907,-1212,
+-2040,-1116,-1907,-1072,-1907,-1212,
+-1958,-1002,-2040,-1116,-1907,-1072,
+-1958,-1002,-1825,-959,-1907,-1072,
+-2040,-803,-1907,-846,-1907,-706,
+-2040,-803,-2040,-663,-1907,-706,
+-1907,-846,-1825,-959,-1958,-1002,
+-1907,-846,-2040,-889,-1958,-1002,
+-2040,-803,-1907,-846,-2040,-889,
+-2040,-803,-2173,-846,-2040,-889,
+-1907,-706,-1907,-846,-1825,-733,
+-1907,-706,-1825,-593,-1825,-733,
+-1907,-706,-1825,-593,-1958,-549,
+-1907,-706,-2040,-663,-1958,-549,
+-2040,-1116,-2040,-1256,-2173,-1212,
+-2040,-1116,-2173,-1072,-2173,-1212,
+-2122,-1002,-2040,-1116,-2173,-1072,
+-2122,-1002,-2256,-959,-2173,-1072,
+-2173,-846,-2256,-959,-2122,-1002,
+-2173,-846,-2040,-889,-2122,-1002,
+-2256,-1099,-2173,-1212,-2306,-1256,
+-2256,-1099,-2389,-1142,-2306,-1256,
+-2256,-1099,-2173,-1212,-2173,-1072,
+-2256,-1099,-2256,-959,-2173,-1072,
+-2655,-1369,-2737,-1256,-2604,-1212,
+-2655,-1369,-2522,-1326,-2604,-1212,
+-2522,-1099,-2604,-1212,-2471,-1256,
+-2522,-1099,-2389,-1142,-2471,-1256,
+-2604,-1212,-2737,-1256,-2737,-1116,
+-2604,-1212,-2604,-1072,-2737,-1116,
+-2471,-1256,-2604,-1212,-2522,-1326,
+-2471,-1256,-2389,-1369,-2522,-1326,
+-2471,-1256,-2389,-1369,-2306,-1256,
+-2471,-1256,-2389,-1142,-2306,-1256,
+-2122,-1369,-2040,-1256,-2173,-1212,
+-2122,-1369,-2256,-1326,-2173,-1212,
+-2306,-1256,-2173,-1212,-2256,-1326,
+-2306,-1256,-2389,-1369,-2256,-1326,
+-614,-846,-697,-959,-779,-846,
+-614,-846,-697,-733,-779,-846,
+-614,-706,-614,-846,-697,-733,
+-614,-706,-697,-593,-697,-733,
+-430,-593,-564,-549,-481,-436,
+-430,-593,-348,-479,-481,-436,
+-564,-549,-697,-593,-614,-706,
+-564,-549,-481,-663,-614,-706,
+-830,-1002,-912,-1116,-994,-1002,
+-830,-1002,-912,-889,-994,-1002,
+-830,-1002,-912,-1116,-779,-1072,
+-830,-1002,-697,-959,-779,-1072,
+-912,-803,-779,-846,-779,-706,
+-912,-803,-912,-663,-779,-706,
+-779,-846,-697,-959,-830,-1002,
+-779,-846,-912,-889,-830,-1002,
+-912,-803,-779,-846,-912,-889,
+-912,-803,-1045,-846,-912,-889,
+-779,-706,-779,-846,-697,-733,
+-779,-706,-697,-593,-697,-733,
+-779,-706,-697,-593,-830,-549,
+-779,-706,-912,-663,-830,-549,
+-297,-183,-430,-139,-348,-26,
+-297,-183,-215,-69,-348,-26,
+-430,-139,-564,-183,-481,-296,
+-430,-139,-348,-253,-481,-296,
+-297,-183,-430,-139,-348,-253,
+-297,-183,-215,-296,-348,-253,
+-133,-183,-215,-296,-297,-183,
+-133,-183,-215,-69,-297,-183,
+-133,-43,-133,-183,-215,-69,
+-133,-43,-215,69,-215,-69,
+-646,-296,-779,-339,-697,-453,
+-646,-296,-564,-409,-697,-453,
+-779,-339,-912,-296,-830,-183,
+-779,-339,-697,-226,-830,-183,
+-646,-296,-779,-339,-697,-226,
+-646,-296,-564,-183,-697,-226,
+-481,-436,-481,-296,-348,-339,
+-481,-436,-348,-479,-348,-339,
+-481,-296,-564,-183,-646,-296,
+-481,-296,-564,-409,-646,-296,
+-481,-436,-481,-296,-564,-409,
+-481,-436,-564,-549,-564,-409,
+-348,-339,-481,-296,-348,-253,
+-348,-339,-215,-296,-348,-253,
+-779,-339,-912,-296,-912,-436,
+-779,-339,-779,-479,-912,-436,
+-697,-453,-779,-339,-779,-479,
+-697,-453,-697,-593,-779,-479,
+-564,-549,-697,-593,-697,-453,
+-564,-549,-564,-409,-697,-453,
+-830,-549,-912,-436,-994,-549,
+-830,-549,-912,-663,-994,-549,
+-830,-549,-912,-436,-779,-479,
+-830,-549,-697,-593,-779,-479,
+-1476,-1212,-1343,-1256,-1343,-1116,
+-1476,-1212,-1476,-1072,-1343,-1116,
+-1343,-1256,-1261,-1369,-1394,-1412,
+-1343,-1256,-1476,-1299,-1394,-1412,
+-1476,-1212,-1343,-1256,-1476,-1299,
+-1476,-1212,-1609,-1256,-1476,-1299,
+-1343,-803,-1210,-846,-1210,-706,
+-1343,-803,-1343,-663,-1210,-706,
+-1210,-846,-1128,-959,-1261,-1002,
+-1210,-846,-1343,-889,-1261,-1002,
+-1343,-803,-1210,-846,-1343,-889,
+-1343,-803,-1476,-846,-1343,-889,
+-1394,-593,-1476,-706,-1558,-593,
+-1394,-593,-1476,-479,-1558,-593,
+-1476,-706,-1476,-846,-1343,-803,
+-1476,-706,-1343,-663,-1343,-803,
+-1394,-593,-1476,-706,-1343,-663,
+-1394,-593,-1261,-549,-1343,-663,
+-1128,-1099,-1045,-1212,-1178,-1256,
+-1128,-1099,-1261,-1142,-1178,-1256,
+-1045,-1212,-912,-1256,-912,-1116,
+-1045,-1212,-1045,-1072,-912,-1116,
+-1128,-1099,-1045,-1212,-1045,-1072,
+-1128,-1099,-1128,-959,-1045,-1072,
+-1343,-1116,-1261,-1002,-1394,-959,
+-1343,-1116,-1476,-1072,-1394,-959,
+-1261,-1002,-1128,-959,-1128,-1099,
+-1261,-1002,-1261,-1142,-1128,-1099,
+-1343,-1116,-1261,-1002,-1261,-1142,
+-1343,-1116,-1343,-1256,-1261,-1142,
+-1394,-959,-1261,-1002,-1343,-889,
+-1394,-959,-1476,-846,-1343,-889,
+-1394,-959,-1476,-846,-1558,-959,
+-1394,-959,-1476,-1072,-1558,-959,
+-1045,-1212,-912,-1256,-994,-1369,
+-1045,-1212,-1128,-1326,-994,-1369,
+-1178,-1256,-1045,-1212,-1128,-1326,
+-1178,-1256,-1261,-1369,-1128,-1326,
+-1343,-1256,-1261,-1369,-1178,-1256,
+-1343,-1256,-1261,-1142,-1178,-1256,
+-1210,-846,-1128,-959,-1045,-846,
+-1210,-846,-1128,-733,-1045,-846,
+-1210,-706,-1210,-846,-1128,-733,
+-1210,-706,-1128,-593,-1128,-733,
+-1394,-593,-1261,-549,-1343,-436,
+-1394,-593,-1476,-479,-1343,-436,
+-1261,-549,-1128,-593,-1210,-706,
+-1261,-549,-1343,-663,-1210,-706,
+-994,-1002,-912,-1116,-1045,-1072,
+-994,-1002,-1128,-959,-1045,-1072,
+-912,-803,-1045,-846,-1045,-706,
+-912,-803,-912,-663,-1045,-706,
+-1045,-846,-1128,-959,-994,-1002,
+-1045,-846,-912,-889,-994,-1002,
+-1045,-706,-1045,-846,-1128,-733,
+-1045,-706,-1128,-593,-1128,-733,
+-1045,-706,-1128,-593,-994,-549,
+-1045,-706,-912,-663,-994,-549,
+-1128,-453,-1045,-339,-1178,-296,
+-1128,-453,-1261,-409,-1178,-296,
+-1045,-339,-912,-296,-912,-436,
+-1045,-339,-1045,-479,-912,-436,
+-1128,-453,-1045,-339,-1045,-479,
+-1128,-453,-1128,-593,-1045,-479,
+-1261,-549,-1128,-593,-1128,-453,
+-1261,-549,-1261,-409,-1128,-453,
+-1343,-436,-1261,-549,-1261,-409,
+-1343,-436,-1343,-296,-1261,-409,
+-994,-549,-912,-436,-1045,-479,
+-994,-549,-1128,-593,-1045,-479,
+-1476,-1212,-1609,-1256,-1609,-1116,
+-1476,-1212,-1476,-1072,-1609,-1116,
+-1609,-1256,-1692,-1369,-1558,-1412,
+-1609,-1256,-1476,-1299,-1558,-1412,
+-1742,-846,-1825,-959,-1692,-1002,
+-1742,-846,-1609,-889,-1692,-1002,
+-1609,-803,-1742,-846,-1609,-889,
+-1609,-803,-1476,-846,-1609,-889,
+-1476,-706,-1476,-846,-1609,-803,
+-1476,-706,-1609,-663,-1609,-803,
+-1825,-1099,-1907,-1212,-1774,-1256,
+-1825,-1099,-1692,-1142,-1774,-1256,
+-1825,-1099,-1907,-1212,-1907,-1072,
+-1825,-1099,-1825,-959,-1907,-1072,
+-1609,-1116,-1692,-1002,-1558,-959,
+-1609,-1116,-1476,-1072,-1558,-959,
+-1692,-1002,-1825,-959,-1825,-1099,
+-1692,-1002,-1692,-1142,-1825,-1099,
+-1609,-1116,-1692,-1002,-1692,-1142,
+-1609,-1116,-1609,-1256,-1692,-1142,
+-1558,-959,-1692,-1002,-1609,-889,
+-1558,-959,-1476,-846,-1609,-889,
+-1907,-1212,-2040,-1256,-1958,-1369,
+-1907,-1212,-1825,-1326,-1958,-1369,
+-1774,-1256,-1907,-1212,-1825,-1326,
+-1774,-1256,-1692,-1369,-1825,-1326,
+-1609,-1256,-1692,-1369,-1774,-1256,
+-1609,-1256,-1692,-1142,-1774,-1256,
+-2471,-803,-2604,-846,-2604,-706,
+-2471,-803,-2471,-663,-2604,-706,
+-2389,-689,-2471,-803,-2471,-663,
+-2389,-689,-2389,-549,-2471,-663,
+-2173,-706,-2256,-593,-2122,-549,
+-2173,-706,-2040,-663,-2122,-549,
+-2256,-593,-2389,-549,-2389,-689,
+-2256,-593,-2256,-733,-2389,-689,
+-2686,-593,-2604,-706,-2522,-593,
+-2686,-593,-2604,-479,-2522,-593,
+-2604,-706,-2604,-846,-2737,-803,
+-2604,-706,-2737,-663,-2737,-803,
+-2686,-593,-2604,-706,-2737,-663,
+-2686,-593,-2820,-549,-2737,-663,
+-2522,-593,-2604,-706,-2471,-663,
+-2522,-593,-2389,-549,-2471,-663,
+-2522,-593,-2389,-549,-2471,-436,
+-2522,-593,-2604,-479,-2471,-436,
+-1825,-453,-1907,-339,-1774,-296,
+-1825,-453,-1692,-409,-1774,-296,
+-1907,-339,-2040,-296,-2040,-436,
+-1907,-339,-1907,-479,-2040,-436,
+-1825,-453,-1907,-339,-1907,-479,
+-1825,-453,-1825,-593,-1907,-479,
+-1692,-549,-1825,-593,-1825,-453,
+-1692,-549,-1692,-409,-1825,-453,
+-1609,-436,-1692,-549,-1692,-409,
+-1609,-436,-1609,-296,-1692,-409,
+-2173,-339,-2306,-296,-2306,-436,
+-2173,-339,-2173,-479,-2306,-436,
+-2306,-296,-2389,-183,-2256,-140,
+-2306,-296,-2173,-253,-2256,-140,
+-2173,-339,-2306,-296,-2173,-253,
+-2173,-339,-2040,-296,-2173,-253,
+-2122,-549,-2040,-436,-1958,-549,
+-2122,-549,-2040,-663,-1958,-549,
+-2040,-436,-2040,-296,-2173,-339,
+-2040,-436,-2173,-479,-2173,-339,
+-2122,-549,-2040,-436,-2173,-479,
+-2122,-549,-2256,-593,-2173,-479,
+-1958,-549,-2040,-436,-1907,-479,
+-1958,-549,-1825,-593,-1907,-479,
+-2306,-296,-2389,-183,-2471,-296,
+-2306,-296,-2389,-409,-2471,-296,
+-2306,-436,-2306,-296,-2389,-409,
+-2306,-436,-2389,-549,-2389,-409,
+-2256,-593,-2389,-549,-2306,-436,
+-2256,-593,-2173,-479,-2306,-436,
+-2471,-436,-2471,-296,-2604,-339,
+-2471,-436,-2604,-479,-2604,-339,
+-2471,-436,-2471,-296,-2389,-409,
+-2471,-436,-2389,-549,-2389,-409,
+-2686,-593,-2820,-549,-2737,-436,
+-2686,-593,-2604,-479,-2737,-436,
+-2737,-436,-2737,-296,-2604,-339,
+-2737,-436,-2604,-479,-2604,-339,
+-2655,-183,-2522,-140,-2604,-26,
+-2655,-183,-2737,-70,-2604,-26,
+-2522,-140,-2389,-183,-2471,-296,
+-2522,-140,-2604,-253,-2471,-296,
+-2655,-183,-2522,-140,-2604,-253,
+-2655,-183,-2737,-296,-2604,-253,
+-2820,-183,-2737,-296,-2655,-183,
+-2820,-183,-2737,-70,-2655,-183,
+-2604,-339,-2471,-296,-2604,-253,
+-2604,-339,-2737,-296,-2604,-253,
+-912,209,-994,323,-861,366,
+-912,209,-779,253,-861,366,
+-994,323,-1128,366,-1128,226,
+-994,323,-994,183,-1128,226,
+-912,209,-994,323,-994,183,
+-912,209,-912,69,-994,183,
+-697,226,-779,113,-646,69,
+-697,226,-564,183,-646,69,
+-779,113,-912,69,-912,209,
+-779,113,-779,253,-912,209,
+-697,226,-779,113,-779,253,
+-697,226,-697,366,-779,253,
+-1261,323,-1394,366,-1394,226,
+-1261,323,-1261,183,-1394,226,
+-1394,366,-1476,479,-1343,523,
+-1394,366,-1261,409,-1343,523,
+-1261,323,-1394,366,-1261,409,
+-1261,323,-1128,366,-1261,409,
+-1210,113,-1128,226,-1045,113,
+-1210,113,-1128,-1,-1045,113,
+-1128,226,-1128,366,-1261,323,
+-1128,226,-1261,183,-1261,323,
+-1210,113,-1128,226,-1261,183,
+-1210,113,-1343,69,-1261,183,
+-1045,113,-1128,226,-994,183,
+-1045,113,-912,69,-994,183,
+-1045,113,-912,69,-994,-43,
+-1045,113,-1128,-1,-994,-43,
+-430,-139,-564,-183,-564,-43,
+-430,-139,-430,-3,-564,-43,
+-348,-26,-430,-139,-430,-3,
+-348,-26,-348,113,-430,-3,
+-133,-43,-215,69,-82,113,
+-133,-43,0,0,-82,113,
+-215,69,-348,113,-348,-26,
+-215,69,-215,-69,-348,-26,
+-697,-139,-830,-183,-830,-43,
+-697,-139,-697,-7,-830,-43,
+-697,-139,-830,-183,-697,-226,
+-697,-139,-564,-183,-697,-226,
+-646,69,-564,-43,-481,69,
+-646,69,-564,183,-481,69,
+-564,-43,-564,-183,-697,-139,
+-564,-43,-697,-7,-697,-139,
+-646,69,-564,-43,-697,-7,
+-646,69,-779,113,-697,-7,
+-481,69,-564,-43,-430,-3,
+-481,69,-348,113,-430,-3,
+-481,69,-348,113,-430,226,
+-481,69,-564,183,-430,226,
+-830,-183,-912,-296,-994,-183,
+-830,-183,-912,-69,-994,-183,
+-830,-43,-830,-183,-912,-69,
+-830,-43,-912,69,-912,-69,
+-779,113,-912,69,-830,-43,
+-779,113,-697,-7,-830,-43,
+-994,-43,-994,-183,-1128,-139,
+-994,-43,-1128,-1,-1128,-139,
+-994,-43,-994,-183,-912,-69,
+-994,-43,-912,69,-912,-69,
+-1958,323,-2091,366,-2091,226,
+-1958,323,-1958,183,-2091,226,
+-2091,366,-2173,479,-2040,523,
+-2091,366,-1958,409,-2040,523,
+-1958,323,-2091,366,-1958,409,
+-1958,323,-1825,366,-1958,409,
+-1907,113,-1825,226,-1742,113,
+-1907,113,-1825,-3,-1742,113,
+-1825,226,-1825,366,-1958,323,
+-1825,226,-1958,183,-1958,323,
+-1907,113,-1825,226,-1958,183,
+-1907,113,-2040,69,-1958,183,
+-2173,619,-2256,733,-2122,776,
+-2173,619,-2040,663,-2122,776,
+-2256,733,-2389,776,-2389,636,
+-2256,733,-2256,593,-2389,636,
+-2173,619,-2256,733,-2256,593,
+-2173,619,-2173,479,-2256,593,
+-1958,636,-2040,523,-1907,479,
+-1958,636,-1825,593,-1907,479,
+-2040,523,-2173,479,-2173,619,
+-2040,523,-2040,663,-2173,619,
+-1958,636,-2040,523,-2040,663,
+-1958,636,-1958,776,-2040,663,
+-1907,479,-2040,523,-1958,409,
+-1907,479,-1825,366,-1958,409,
+-1907,479,-1825,366,-1742,479,
+-1907,479,-1825,593,-1742,479,
+-1558,-139,-1476,-26,-1394,-139,
+-1558,-139,-1476,-253,-1394,-139,
+-1476,-26,-1476,113,-1609,69,
+-1476,-26,-1609,-70,-1609,69,
+-1558,-139,-1476,-26,-1609,-70,
+-1558,-139,-1692,-183,-1609,-70,
+-1476,-339,-1609,-296,-1609,-436,
+-1476,-339,-1476,-479,-1609,-436,
+-1609,-296,-1692,-183,-1558,-139,
+-1609,-296,-1476,-253,-1558,-139,
+-1476,-339,-1609,-296,-1476,-253,
+-1476,-339,-1343,-296,-1476,-253,
+-1558,226,-1558,366,-1692,323,
+-1558,226,-1692,183,-1692,323,
+-1558,366,-1476,479,-1394,366,
+-1558,366,-1476,253,-1394,366,
+-1558,226,-1558,366,-1476,253,
+-1558,226,-1476,113,-1476,253,
+-1742,113,-1609,69,-1692,-43,
+-1742,113,-1825,-3,-1692,-43,
+-1609,69,-1476,113,-1558,226,
+-1609,69,-1692,183,-1558,226,
+-1742,113,-1609,69,-1692,183,
+-1742,113,-1825,226,-1692,183,
+-1692,-43,-1609,69,-1609,-70,
+-1692,-43,-1692,-183,-1609,-70,
+-1692,-43,-1692,-183,-1825,-140,
+-1692,-43,-1825,-3,-1825,-140,
+-1558,366,-1476,479,-1609,523,
+-1558,366,-1692,409,-1609,523,
+-1692,323,-1558,366,-1692,409,
+-1692,323,-1825,366,-1692,409,
+-1825,226,-1825,366,-1692,323,
+-1825,226,-1692,183,-1692,323,
+-1742,479,-1609,523,-1692,636,
+-1742,479,-1825,593,-1692,636,
+-1742,479,-1609,523,-1692,409,
+-1742,479,-1825,366,-1692,409,
+-1476,-26,-1476,113,-1343,69,
+-1476,-26,-1343,-70,-1343,69,
+-1394,-139,-1476,-26,-1343,-70,
+-1394,-139,-1261,-183,-1343,-70,
+-1476,-339,-1343,-296,-1343,-436,
+-1476,-339,-1476,-479,-1343,-436,
+-1343,-296,-1261,-183,-1394,-139,
+-1343,-296,-1476,-253,-1394,-139,
+-1394,226,-1394,366,-1476,253,
+-1394,226,-1476,113,-1476,253,
+-1210,113,-1343,69,-1261,-43,
+-1210,113,-1128,-1,-1261,-43,
+-1343,69,-1476,113,-1394,226,
+-1343,69,-1261,183,-1394,226,
+-1261,-43,-1343,69,-1343,-70,
+-1261,-43,-1261,-183,-1343,-70,
+-1261,-43,-1261,-183,-1128,-139,
+-1261,-43,-1128,-1,-1128,-139,
+-1045,-339,-912,-296,-994,-183,
+-1045,-339,-1128,-226,-994,-183,
+-1178,-296,-1045,-339,-1128,-226,
+-1178,-296,-1261,-183,-1128,-226,
+-1343,-296,-1261,-183,-1178,-296,
+-1343,-296,-1261,-409,-1178,-296,
+-1128,-139,-994,-183,-1128,-226,
+-1128,-139,-1261,-183,-1128,-226,
+-2091,366,-2173,479,-2256,366,
+-2091,366,-2173,253,-2256,366,
+-2091,226,-2091,366,-2173,253,
+-2091,226,-2173,113,-2173,253,
+-1907,113,-2040,69,-1958,-43,
+-1907,113,-1825,-3,-1958,-43,
+-2040,69,-2173,113,-2091,226,
+-2040,69,-1958,183,-2091,226,
+-2306,523,-2389,636,-2471,523,
+-2306,523,-2389,409,-2471,523,
+-2306,523,-2389,636,-2256,593,
+-2306,523,-2173,479,-2256,593,
+-2389,323,-2256,366,-2256,226,
+-2389,323,-2389,183,-2256,226,
+-2256,366,-2173,479,-2306,523,
+-2256,366,-2389,409,-2306,523,
+-2389,323,-2256,366,-2389,409,
+-2389,323,-2522,366,-2389,409,
+-2256,226,-2256,366,-2173,253,
+-2256,226,-2173,113,-2173,253,
+-2256,226,-2173,113,-2306,69,
+-2256,226,-2389,183,-2306,69,
+-1907,-339,-2040,-296,-1958,-183,
+-1907,-339,-1825,-226,-1958,-183,
+-1774,-296,-1907,-339,-1825,-226,
+-1774,-296,-1692,-183,-1825,-226,
+-1609,-296,-1692,-183,-1774,-296,
+-1609,-296,-1692,-409,-1774,-296,
+-2122,-183,-2256,-140,-2173,-26,
+-2122,-183,-2040,-70,-2173,-26,
+-2122,-183,-2256,-140,-2173,-253,
+-2122,-183,-2040,-296,-2173,-253,
+-1958,-43,-1958,-183,-1825,-140,
+-1958,-43,-1825,-3,-1825,-140,
+-1958,-183,-2040,-296,-2122,-183,
+-1958,-183,-2040,-70,-2122,-183,
+-1958,-43,-1958,-183,-2040,-70,
+-1958,-43,-2040,69,-2040,-70,
+-1825,-140,-1958,-183,-1825,-226,
+-1825,-140,-1692,-183,-1825,-226,
+-2256,-140,-2389,-183,-2389,-43,
+-2256,-140,-2256,-4,-2389,-43,
+-2173,-26,-2256,-140,-2256,-4,
+-2173,-26,-2173,113,-2256,-4,
+-2040,69,-2173,113,-2173,-26,
+-2040,69,-2040,-70,-2173,-26,
+-2306,69,-2389,-43,-2471,69,
+-2306,69,-2389,183,-2471,69,
+-2306,69,-2389,-43,-2256,-4,
+-2306,69,-2173,113,-2256,-4,
+-2686,366,-2604,479,-2522,366,
+-2686,366,-2604,253,-2522,366,
+-2686,226,-2686,366,-2604,253,
+-2686,226,-2604,113,-2604,253,
+-2737,69,-2604,113,-2686,226,
+-2737,69,-2820,183,-2686,226,
+-2389,636,-2389,776,-2522,733,
+-2389,636,-2522,593,-2522,733,
+-2471,523,-2389,636,-2522,593,
+-2471,523,-2604,479,-2522,593,
+-2389,323,-2522,366,-2522,226,
+-2389,323,-2389,183,-2522,226,
+-2522,366,-2604,479,-2471,523,
+-2522,366,-2389,409,-2471,523,
+-2522,226,-2522,366,-2604,253,
+-2522,226,-2604,113,-2604,253,
+-2522,226,-2604,113,-2471,69,
+-2522,226,-2389,183,-2471,69,
+-2522,-140,-2389,-183,-2389,-43,
+-2522,-140,-2522,-5,-2389,-43,
+-2604,-26,-2522,-140,-2522,-5,
+-2604,-26,-2604,113,-2522,-5,
+-2737,69,-2604,113,-2604,-26,
+-2737,69,-2737,-70,-2604,-26,
+-2471,69,-2389,-43,-2522,-5,
+-2471,69,-2604,113,-2522,-5,
+-2655,776,-2522,733,-2604,619,
+-2655,776,-2737,663,-2604,619,
+-2522,733,-2389,776,-2471,889,
+-2522,733,-2604,846,-2471,889,
+-2655,776,-2522,733,-2604,846,
+-2655,776,-2737,889,-2604,846,
+-2820,776,-2737,889,-2655,776,
+-2820,776,-2737,663,-2655,776,
+-2686,366,-2604,479,-2737,523,
+-2686,366,-2820,409,-2737,523,
+-2604,619,-2522,733,-2522,593,
+-2604,619,-2604,479,-2522,593,
+-2737,523,-2604,479,-2604,619,
+-2737,523,-2737,663,-2604,619,
+-994,323,-1128,366,-1045,479,
+-994,323,-912,436,-1045,479,
+-861,366,-994,323,-912,436,
+-861,366,-779,479,-912,436,
+-697,226,-697,366,-564,323,
+-697,226,-564,183,-564,323,
+-697,366,-779,479,-861,366,
+-697,366,-779,253,-861,366,
+-1210,479,-1343,523,-1261,636,
+-1210,479,-1128,593,-1261,636,
+-1210,479,-1343,523,-1261,409,
+-1210,479,-1128,366,-1261,409,
+-1045,619,-1045,479,-912,523,
+-1045,619,-912,663,-912,523,
+-1045,479,-1128,366,-1210,479,
+-1045,479,-1128,593,-1210,479,
+-1045,619,-1045,479,-1128,593,
+-1045,619,-1128,733,-1128,593,
+-912,523,-1045,479,-912,436,
+-912,523,-779,479,-912,436,
+-912,523,-779,479,-779,619,
+-912,523,-912,663,-779,619,
+-266,226,-266,366,-133,323,
+-266,226,-133,183,-133,323,
+-266,366,-348,479,-430,366,
+-266,366,-348,253,-430,366,
+-266,226,-266,366,-348,253,
+-266,226,-348,113,-348,253,
+-215,69,-348,113,-266,226,
+-215,69,-133,183,-266,226,
+-82,113,-215,69,-133,183,
+-82,113,0,226,-133,183,
+-481,523,-564,636,-646,523,
+-481,523,-564,409,-646,523,
+-564,636,-564,776,-430,733,
+-564,636,-430,593,-430,733,
+-481,523,-564,636,-430,593,
+-481,523,-348,479,-430,593,
+-564,323,-430,366,-430,226,
+-564,323,-564,183,-430,226,
+-430,366,-348,479,-481,523,
+-430,366,-564,409,-481,523,
+-564,323,-430,366,-564,409,
+-564,323,-697,366,-564,409,
+-430,226,-430,366,-348,253,
+-430,226,-348,113,-348,253,
+-564,636,-564,776,-697,733,
+-564,636,-697,593,-697,733,
+-646,523,-564,636,-697,593,
+-646,523,-779,479,-697,593,
+-697,366,-779,479,-646,523,
+-697,366,-564,409,-646,523,
+-779,619,-697,733,-830,776,
+-779,619,-912,663,-830,776,
+-779,619,-697,733,-697,593,
+-779,619,-779,479,-697,593,
+-1774,889,-1907,932,-1825,1046,
+-1774,889,-1692,1002,-1825,1046,
+-1907,932,-2040,889,-1958,776,
+-1907,932,-1825,819,-1958,776,
+-1774,889,-1907,932,-1825,819,
+-1774,889,-1692,776,-1825,819,
+-1609,1029,-1609,889,-1476,932,
+-1609,1029,-1476,1072,-1476,932,
+-1609,889,-1692,776,-1774,889,
+-1609,889,-1692,1002,-1774,889,
+-1609,1029,-1609,889,-1692,1002,
+-1609,1029,-1692,1142,-1692,1002,
+-2256,733,-2389,776,-2306,889,
+-2256,733,-2173,846,-2306,889,
+-2122,776,-2256,733,-2173,846,
+-2122,776,-2040,889,-2173,846,
+-1958,636,-1958,776,-1825,733,
+-1958,636,-1825,593,-1825,733,
+-1958,776,-2040,889,-2122,776,
+-1958,776,-2040,663,-2122,776,
+-1825,733,-1958,776,-1825,819,
+-1825,733,-1692,776,-1825,819,
+-1825,733,-1692,776,-1692,636,
+-1825,733,-1825,593,-1692,636,
+-1178,1029,-1178,889,-1045,932,
+-1178,1029,-1045,1072,-1045,932,
+-1178,889,-1261,776,-1343,889,
+-1178,889,-1261,1002,-1343,889,
+-1178,1029,-1178,889,-1261,1002,
+-1178,1029,-1261,1142,-1261,1002,
+-994,1142,-1128,1186,-1045,1299,
+-994,1142,-912,1256,-1045,1299,
+-1128,1186,-1261,1142,-1178,1029,
+-1128,1186,-1045,1072,-1178,1029,
+-994,1142,-1128,1186,-1045,1072,
+-994,1142,-912,1029,-1045,1072,
+-1394,733,-1476,619,-1558,733,
+-1394,733,-1476,846,-1558,733,
+-1476,619,-1476,479,-1343,523,
+-1476,619,-1343,663,-1343,523,
+-1394,733,-1476,619,-1343,663,
+-1394,733,-1261,776,-1343,663,
+-1476,932,-1343,889,-1343,1029,
+-1476,932,-1476,1072,-1343,1029,
+-1343,889,-1261,776,-1394,733,
+-1343,889,-1476,846,-1394,733,
+-1476,932,-1343,889,-1476,846,
+-1476,932,-1609,889,-1476,846,
+-1343,1029,-1343,889,-1261,1002,
+-1343,1029,-1261,1142,-1261,1002,
+-1343,1029,-1261,1142,-1394,1186,
+-1343,1029,-1476,1072,-1394,1186,
+-1476,619,-1476,479,-1609,523,
+-1476,619,-1609,663,-1609,523,
+-1558,733,-1476,619,-1609,663,
+-1558,733,-1692,776,-1609,663,
+-1609,889,-1692,776,-1558,733,
+-1609,889,-1476,846,-1558,733,
+-1692,636,-1609,523,-1609,663,
+-1692,636,-1692,776,-1609,663,
+-1178,889,-1261,776,-1128,733,
+-1178,889,-1045,846,-1128,733,
+-1045,932,-1178,889,-1045,846,
+-1045,932,-912,889,-1045,846,
+-994,1142,-912,1029,-830,1142,
+-994,1142,-912,1256,-830,1142,
+-912,1029,-912,889,-1045,932,
+-912,1029,-1045,1072,-1045,932,
+-1261,636,-1343,523,-1343,663,
+-1261,636,-1261,776,-1343,663,
+-1045,619,-1128,733,-994,776,
+-1045,619,-912,663,-994,776,
+-1128,733,-1261,776,-1261,636,
+-1128,733,-1128,593,-1261,636,
+-994,776,-1128,733,-1045,846,
+-994,776,-912,889,-1045,846,
+-994,776,-912,889,-830,776,
+-994,776,-912,663,-830,776,
+-779,932,-646,889,-646,1029,
+-779,932,-779,1072,-646,1029,
+-646,889,-564,776,-697,733,
+-646,889,-779,846,-697,733,
+-779,932,-646,889,-779,846,
+-779,932,-912,889,-779,846,
+-912,1029,-912,889,-779,932,
+-912,1029,-779,1072,-779,932,
+-830,1142,-912,1029,-779,1072,
+-830,1142,-697,1186,-779,1072,
+-830,776,-697,733,-779,846,
+-830,776,-912,889,-779,846,
+-1907,932,-2040,889,-2040,1029,
+-1907,932,-1907,1072,-2040,1029,
+-1825,1046,-1907,932,-1907,1072,
+-1825,1046,-1825,1186,-1907,1072,
+-1609,1029,-1692,1142,-1558,1186,
+-1609,1029,-1476,1072,-1558,1186,
+-1692,1142,-1825,1186,-1825,1046,
+-1692,1142,-1692,1002,-1825,1046,
+-2173,932,-2306,889,-2306,1029,
+-2173,932,-2173,1072,-2306,1029,
+-2173,932,-2306,889,-2173,846,
+-2173,932,-2040,889,-2173,846,
+-2122,1142,-2040,1029,-1958,1142,
+-2122,1142,-2040,1256,-1958,1142,
+-2040,1029,-2040,889,-2173,932,
+-2040,1029,-2173,1072,-2173,932,
+-2122,1142,-2040,1029,-2173,1072,
+-2122,1142,-2256,1186,-2173,1072,
+-1958,1142,-2040,1029,-1907,1072,
+-1958,1142,-1825,1186,-1907,1072,
+-1958,1142,-1825,1186,-1907,1299,
+-1958,1142,-2040,1256,-1907,1299,
+-1128,1186,-1261,1142,-1261,1282,
+-1128,1186,-1128,1326,-1261,1282,
+-1558,1186,-1476,1299,-1394,1186,
+-1558,1186,-1476,1072,-1394,1186,
+-1558,1186,-1476,1299,-1609,1256,
+-1558,1186,-1692,1142,-1609,1256,
+-1394,1186,-1476,1299,-1343,1256,
+-1394,1186,-1261,1142,-1343,1256,
+-1692,1142,-1825,1186,-1742,1299,
+-1692,1142,-1609,1256,-1742,1299,
+-2389,1282,-2389,1142,-2256,1186,
+-2389,1282,-2256,1326,-2256,1186,
+-2306,889,-2389,776,-2471,889,
+-2306,889,-2389,1002,-2471,889,
+-2306,1029,-2306,889,-2389,1002,
+-2306,1029,-2389,1142,-2389,1002,
+-2122,1142,-2256,1186,-2173,1299,
+-2122,1142,-2040,1256,-2173,1299,
+-2256,1186,-2389,1142,-2306,1029,
+-2256,1186,-2173,1072,-2306,1029,
+-2604,932,-2471,889,-2471,1029,
+-2604,932,-2604,1072,-2471,1029,
+-2604,932,-2471,889,-2604,846,
+-2604,932,-2737,889,-2604,846,
+-2737,1029,-2737,889,-2604,932,
+-2737,1029,-2604,1072,-2604,932,
+-2655,1142,-2737,1029,-2604,1072,
+-2655,1142,-2522,1186,-2604,1072,
+-2389,1282,-2389,1142,-2522,1186,
+-2389,1282,-2522,1326,-2522,1186,
+-2471,1029,-2471,889,-2389,1002,
+-2471,1029,-2389,1142,-2389,1002,
+-2655,1142,-2522,1186,-2604,1299,
+-2655,1142,-2737,1256,-2604,1299,
+-2522,1186,-2389,1142,-2471,1029,
+-2522,1186,-2604,1072,-2471,1029,
+0,1046,0,1186,-133,1142,
+0,1046,-133,1002,-133,1142,
+-82,932,0,1046,-133,1002,
+-82,932,-215,889,-133,1002,
+0,733,-133,776,-133,636,
+0,733,0,593,-133,636,
+-133,776,-215,889,-82,932,
+-133,776,0,819,-82,932,
+-266,1186,-133,1142,-215,1029,
+-266,1186,-348,1072,-215,1029,
+-133,1142,0,1186,-82,1299,
+-133,1142,-215,1256,-82,1299,
+-266,1186,-133,1142,-215,1256,
+-266,1186,-348,1299,-215,1256,
+-215,1029,-133,1142,-133,1002,
+-215,1029,-215,889,-133,1002,
+-215,1029,-215,889,-348,932,
+-215,1029,-348,1072,-348,932,
+-266,366,-348,479,-215,523,
+-266,366,-133,409,-215,523,
+-133,323,-266,366,-133,409,
+-133,323,0,366,-133,409,
+0,226,0,366,-133,323,
+0,226,-133,183,-133,323,
+-348,619,-430,733,-297,776,
+-348,619,-215,663,-297,776,
+-348,619,-430,733,-430,593,
+-348,619,-348,479,-430,593,
+-133,636,-215,523,-82,479,
+-133,636,0,593,-82,479,
+-215,523,-348,479,-348,619,
+-215,523,-215,663,-348,619,
+-133,636,-215,523,-215,663,
+-133,636,-133,776,-215,663,
+-82,479,-215,523,-133,409,
+-82,479,0,366,-133,409,
+-430,733,-564,776,-481,889,
+-430,733,-348,846,-481,889,
+-297,776,-430,733,-348,846,
+-297,776,-215,889,-348,846,
+-133,776,-215,889,-297,776,
+-133,776,-215,663,-297,776,
+-348,932,-481,889,-481,1029,
+-348,932,-348,1072,-481,1029,
+-348,932,-481,889,-348,846,
+-348,932,-215,889,-348,846,
+-779,1299,-697,1186,-830,1142,
+-779,1299,-912,1256,-830,1142,
+-697,1186,-564,1142,-564,1282,
+-697,1186,-697,1326,-564,1282,
+-266,1186,-348,1299,-430,1186,
+-266,1186,-348,1072,-430,1186,
+-430,1186,-348,1299,-481,1256,
+-430,1186,-564,1142,-481,1256,
+-430,1186,-564,1142,-481,1029,
+-430,1186,-348,1072,-481,1029,
+-646,889,-564,776,-481,889,
+-646,889,-564,1002,-481,889,
+-646,1029,-646,889,-564,1002,
+-646,1029,-564,1142,-564,1002,
+-697,1186,-564,1142,-646,1029,
+-697,1186,-779,1072,-646,1029,
+-481,1029,-481,889,-564,1002,
+-481,1029,-564,1142,-564,1002,
+
+    };
+
+}
