+++ penrose_simulator/Assets/effects/Noise.cs
@@ -1,38 +1,23 @@
 ﻿using UnityEngine;
 using Random = UnityEngine.Random;
-// practive comment
+
 public class Noise : EffectBase {
 
-
   private Settings setting;
-
-  //private float scale;
-  //private float speed;
-  //private float amplifier;
-  //private float colorDelta;
-  private float timeScale;
-  private TileData[] tiles;
+  //private float timeScale;
 
   public override void Init() {
     base.Init();
     setting = new Settings();
-    tiles = controller.penrose.geometry.tileData;
-
-    //scale = Random.Range(0.05f, 0.2f);
-    //speed = Random.Range(0.1f, 1.5f);
-    //amplifier  = Random.Range(1, 5);
-    //colorDelta = Random.value;
-    //
-    //Debug.Log($"Scale: {scale}, Speed: {speed}, Amplifier: {amplifier}, ColorDelta: {colorDelta}");
   }
+
   public override void Draw() {
-    
     for(int i = 0; i < buffer.Length; i++) {
-      float scale = (1.0f+(controller.dance.decay*0.25f)) * setting.scale;
-      float x = tiles[i].center.x * scale;
-      float y = tiles[i].center.y * scale;
-      float z = controller.dance.fixedTime * setting.speed;
-      float n = Perlin.Noise(x,y ,z );
+      float scale = (1.0f + (controller.dance.decay * 0.25f)) * setting.scale;
+      float x     = controller.penrose.tiles[i].center.x * scale;
+      float y     = controller.penrose.tiles[i].center.y * scale;
+      float z     = controller.dance.fixedTime * setting.speed;
+      float n     = Perlin.Noise(x, y, z);
       n *= setting.amplifier;
 
       int v = (int)n;
@@ -56,6 +41,7 @@ public class Noise : EffectBase {
 
   [System.Serializable]
   public class Settings {
+
     public float scale;
     public float speed;
     public float amplifier;
@@ -64,9 +50,10 @@ public class Noise : EffectBase {
     public void Randomize() {
       scale      = Random.Range(0.05f, 0.2f);
       speed      = Random.Range(0.1f, 1.5f);
-      amplifier  = Random.Range(1, 5);
+      amplifier  = Random.Range(1f, 5f);
       colorDelta = Random.value;
     }
+
   }
 
 }
