--- penrose_simulator/Assets/palettes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e31e31deb9b1b4d419d290420fce5aa4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
